% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addImageLegend.R
\name{addImageLegend}
\alias{addImageLegend}
\title{Title}
\usage{
addImageLegend(channels = "DAPI", col = "blue", title, cex = 1)
}
\arguments{
\item{channels}{A vector of names for the channels (DAPI, GFP...).}

\item{col}{Channel colors}

\item{title}{A title for the image.}

\item{cex}{Expansion factor applied to test.}
}
\description{
Title
}
