% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMotifs.R
\name{addMotifs}
\alias{addMotifs}
\title{Add motifs to enrichment plots}
\usage{
addMotifs(plot.DT, pwms, cex.width = 1, cex.height = 1, lwd = 0.1)
}
\arguments{
\item{plot.DT}{Output of the plot.vl_enr or plot.vl_enr_cl methods containing plotting coordinates.}

\item{pwms}{A list of 'PWMatrix' whose names correspond to the 'motif' entries in plot.DT.}

\item{cex.width}{expansion factor for motif width.}

\item{cex.height}{expansion factor for motif height.}
}
\description{
Add motifs to enrichment plots
}
\examples{
For vl_enr object
pl <- plot(vl_enr)
addMotifs(pl, pwms = vl_motifs_DB_v2$pwms_perc)

For vl_enr_cl object
pl <- plot(vl_enr)
addMotifs(pl$plot.DT, pwms = vl_motifs_DB_v2$pwms_perc)

}
