% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRepelAxis.R
\name{addRepelAxis}
\alias{addRepelAxis}
\title{Add labels heatmap}
\usage{
addRepelAxis(side = 2, at, labels, cex = par("cex.axis"), width = 1)
}
\arguments{
\item{side}{The axis is placed as follows: 2 = left (default), 4 = right.}

\item{at}{Y position of the labels.}

\item{labels}{Labels}

\item{cex}{cex expansion factor for the axis. Default= par("cex.axis").}

\item{width}{width of the labels. Default= 1.}
}
\description{
Add axis with repelled labels to a plot.
}
