% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRepelLabels.R
\name{addRepelLabels}
\alias{addRepelLabels}
\title{addRepelLabels}
\usage{
addRepelLabels(
  x,
  y = NULL,
  labels,
  cex = 0.7,
  label.col = "black",
  seg.col = "black",
  rect.draw = TRUE,
  rect.col = adjustcolor("white", 0.3),
  no.overlap = FALSE,
  point_size = 0.5,
  points_padding_x = 0,
  points_padding_y = 0,
  box_padding_x = 0.5,
  box_padding_y = 0.5,
  force_push = 0.00001,
  force_pull = 0.00001
)
}
\arguments{
\item{x}{x coordinates.}

\item{y}{y coordinates.}

\item{labels}{labels to plot.}

\item{label.col}{A vector of colors for labels. Default= "black".}

\item{seg.col}{A vector of colors for segments. Default= "black".}

\item{rect.draw}{Draw rectangle boxes around labels? Default= TRUE.}

\item{rect.col}{A vector of colors for boxes. Default= adjustcolor("white", .3).}

\item{no.overlap}{If set to TRUE, overlapping names will be removed. Default= FALSE.}

\item{point_size}{Approximate size of points, defined as a fraction of char size.
Used in conjunction with 'point_padding' x and y (see below). Default= 0.5}

\item{box_padding_x}{Extra margin around text bounding boxes, defined as a fraction of char size. Default= 0.5.}

\item{box_padding_y}{Extra margin around text bounding boxes, defined as a fraction of char size. Default= 0.5.}

\item{force_push}{Magnitude of the push force. Default= 1e-5.}

\item{force_pull}{Magnitude of the pull force. Default= 1e-5.}

\item{label.cex}{Expansion factor for labels.}

\item{point_padding_x}{Extra margin around data points, defined as a fraction of char size. Default= 0.}

\item{point_padding_y}{Extra margin around data points, defined as a fraction of char size. Default= 0.}
}
\description{
Add repelled labels to an existing scatterplot.
}
\examples{
data("mtcars")
mtcars$car <- rownames(mtcars)

plot(x = mtcars$wt, y = mtcars$mpg)

addRepelLabels(x = mtcars$wt,
y = mtcars$mpg,
labels = mtcars$car,
cex = 0.7)

}
