% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSeqLogo.R
\name{addSeqLogo}
\alias{addSeqLogo}
\title{Plot logo from a numeric matrix}
\usage{
addSeqLogo(mat, x, y, cex.width = 1, cex.height = 1, pos = 4)
}
\arguments{
\item{mat}{A matrix (or a TFBSTools PFMatrix, PWMatrix or ICMatrix) with rows 'A', 'C', 'G', 'T'.}

\item{x}{Th x position where the logo should start or finish (depending on 'pos' argument).}

\item{y}{The y position of the center part of the logo. Default= 0.}

\item{cex.width}{A width expansion factor applied to letter heights.}

\item{pos}{Specifies on which side of the x position the logo should be plotted. Wither 2 (left) or 4 (right). Default= 4.}
}
\description{
Generic function that plots a logo out of an arbitrary numeric matrix.
}
\examples{
# Select br PPM
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
mot <- vlite::importJASPAR(pfm.file)

# Plot
vl_par()
plot(0, 0, type= "n")
addSeqLogo(mot$PFM[[1]], x= -1, y= .5)
addSeqLogo(mot$PPM[[1]], x= -0.5, y= 0)
addSeqLogo(mot$ICM[[1]], x= 0, y= -0.5)

}
