% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSeqLogoEnrPlot.R
\name{addSeqLogoEnrPlot}
\alias{addSeqLogoEnrPlot}
\title{Plot motif logos on the left of an enrichment plot}
\usage{
addSeqLogoEnrPlot(obj, ICM, cex.width = 1, cex.height = 1)
}
\arguments{
\item{obj}{An `vl_enr_cl` object as returned by the ?plot.vl_enr_cl funciton.}

\item{cex.width}{A width expansion factor applied to letter heights.}

\item{cex.height}{A height expansion factor applied to letter heights.}

\item{ICMatrixList}{An ICMatrixList object containing the motifs to add.}
}
\description{
Specialized funciton used to plot motif logs on the left of a vl_enr_cl plot.
}
\examples{
# Select br PPM
sel <- plot(
  enr,
  top.enrich = 5,
  order = "log2OR",
  padj.cutoff = 1e-3,
  col= c("pink", "red"),
  cex = .4,
  main = "ChIP motif enrich."
)
addSeqLogoEnrPlot(
  obj = sel,
  ICM = mot$ICM
)

}
