% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedEnrich.R
\name{bedEnrich}
\alias{bedEnrich}
\title{Find Overlapping Regions Between Two Sets of Genomic Intervals}
\usage{
bedEnrich(
  peaks,
  ctl.peaks,
  regions,
  pseudocount = 1,
  maxgap = -1L,
  minoverlap = 0L,
  ignore.strand = TRUE
)
}
\arguments{
\item{maxgap}{A single integer specifying the maximum gap allowed between 2 ranges for them to
be considered as overlapping. Default= -1.}

\item{minoverlap}{A single integer specifying the minimum overlap between 2 ranges for them to
be considered as overlapping. Default= 0.}

\item{ignore.strand}{If set to FALSE, only reports overlaps between regions that are on the same strand.
If set to TRUE (default), reports overlaps on both strands.}

\item{a}{Query regions in any format compatible with ?importBed.}

\item{b}{Target regions in any format compatible with ?importBed.}
}
\value{
A data.table with columns:
\itemize{
  \item idx.a: line index of the region in a.
  \item idx.b: line index of the overlapping region in b (NA if no overlap).
  \item overlap.start: overlap starting position (NA if no overlap).
  \item overlap.end: overlap ending position (NA if no overlap).
  \item overlap.width: overlap width (0 if no overlap).
}
Of note, regions in a with no overlap in b are not returned.
}
\description{
A wrapper around ?GenomicRanges::findOverlaps() that identifies, for each genomic range in a,
the overlapping regions in b.
}
