% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwGetSeqlengths.R
\name{bwGetSeqlengths}
\alias{bwGetSeqlengths}
\title{Retrieve the seqnames within a bigwig file}
\usage{
bwGetSeqlengths(bw)
}
\arguments{
\item{bw}{Path to a target bw file.}
}
\description{
Returns all the seqLengths of the chromosomes that are represented within a bigwig file.
}
\examples{
# Example track
bw <- "/groups/stark/vloubiere/projects/epigenetic_cancer/db/bw/ATAC/ATAC_PH18_merge.bw"
bwGetSeqlengths(bw= bw)

}
