% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwGetSeqnames.R
\name{bwGetSeqnames}
\alias{bwGetSeqnames}
\title{Retrieve the seqnames within a bigwig file}
\usage{
bwGetSeqnames(bw)
}
\arguments{
\item{bw}{Path to a target bw file.}
}
\description{
Returns all the seqnames that are represented within a bed file.
}
\examples{
# Example track
bw <- "/groups/stark/vloubiere/projects/epigenetic_cancer/db/bw/ATAC/ATAC_PH18_merge.bw"

Compute coverage
bed <- importBed(c("chr2R:10000000-10010000", "chr3R:10000000-10005000", "nonExistingChr:1000-2000"))
bwCoverage(bed= bed, bw= bw)

}
