% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwScreenshot.R
\name{bwScreenshot}
\alias{bwScreenshot}
\title{Create Genomic Screenshots with BigWig Tracks and Gene Annotations}
\usage{
bwScreenshot(
  bed,
  tracks = character(),
  track.names = function(x) gsub("(.*)[.].*$", "\\\\1", x),
  col = colorRampPalette(c("grey60", "grey10"))(length(tracks)),
  bw.max = NA,
  show.bw.range = TRUE,
  nbins = 1000,
  genome,
  gtf,
  sel.gene.symbols = NULL,
  border.col = NA,
  border.lwd = 1,
  bw.border.col = NA,
  bw.border.lwd = 1,
  bw.min = NA,
  ngenes = 1,
  cex.gene.symbol = 1,
  offset.gene.symbol = 0.25,
  col.gene.ps = "tomato",
  col.gene.ns = "cornflowerblue",
  exon.border = NA,
  gtf.transcript = NULL,
  gtf.exon = "exon",
  gtf.symbol = "gene_symbol",
  gtf.transcript.id = "transcript_id",
  region.width = 1,
  space.width = 1,
  bw.height = 1,
  bed.height = 1,
  space.height = 1,
  gene.height = 1,
  gene.space.height = 1,
  show.scale.bar = T,
  add = FALSE
)
}
\arguments{
\item{bed}{Input genomic ranges, in any format compatible with ?importBed().}

\item{tracks}{Character vector of paths to visualization tracks:
\itemize{
  \item BigWig (.bw) files for continuous signal tracks
  \item BED files for discrete regions (any format compatible with importBed)
}}

\item{track.names}{Function or character vector for track labels. Default extracts names
from file basenames.}

\item{col}{Colors for tracks. Default is a grey gradient from grey60 to grey10.}

\item{bw.max}{Maximum value at which bigwig coverage values will be clipped. Default= NA (no clipping).}

\item{show.bw.range}{Should the range of bw tracks be plotted? Default= TRUE.}

\item{nbins}{The number of bins used to quantify the signal (per region). Default= 1000}

\item{genome}{Genome assembly (e.g., "mm10", "dm6") for gene annotations.
If specified, nearby genes will be displayed.}

\item{gtf}{Character path to a custom GTF file for custom gene annotations (alternative to the 'genome' parameter).}

\item{sel.gene.symbols}{If specified, only selected gene symbols will be plotted. Default= NULL.}

\item{border.col}{The color used for track borders. Default= NA.}

\item{border.lwd}{The line width used for track borders. Default= 1.}

\item{bw.min}{Minimum signal value at which bigwig coverage values will be clipped. Default= NA (no clipping).}

\item{ngenes}{Integer. Number of nearest genes to display. Default = 1.}

\item{cex.gene.symbol}{Size of gene symbols. Default = 1.}

\item{offset.gene.symbol}{Vertical offset for gene symbols. Default= 1.}

\item{col.gene.ps}{Color for positive strand genes. Default = "tomato".}

\item{col.gene.ns}{Color for negative strand genes. Default = "cornflowerblue".}

\item{exon.border}{Color for exon rectangle borders. Default = NA, meaning no border.}

\item{gtf.transcript}{If specified, only features with the corresponding biotype will be plotted as transcripts. Default = NULL.}

\item{gtf.exon}{Name of exon features in the GTF file (default = "exon").}

\item{gtf.symbol}{Name of the attribute containing gene symbols in the GTF file. Default = "gene_symbol".}

\item{gtf.transcript.id}{Name of the attribute containing transcript IDs in the GTF file. Default = "transcript_id".}

\item{region.width}{Width of each genomic region. Default= 1.}

\item{space.width}{Horizontal space between regions. Default= 1.}

\item{bw.height}{Height of BigWig tracks. Default= 1.}

\item{bed.height}{Height of BED tracks. Default= 1.}

\item{space.height}{Vertical space between tracks. Default= 1.}

\item{gene.height}{Height of gene tracks. Default= 1.}

\item{gene.space.height}{Space between genes. Default= 1.}

\item{show.scale.bar}{Should the scale bar be plotted? Default= T.}

\item{add}{Should the plot beadded to the existing one? Default= FALSE.}
}
\value{
A screenshot of the genomic tracks/regions.

A genomic screenshot.
}
\description{
Creates publication-quality genomic screenshots displaying BigWig signal tracks,
BED regions, and optional gene annotations for specified genomic regions. Multiple
regions can be displayed side by side for comparison.
}
\examples{
# Plotting parameters for a nice layout
par(mai = c(0.9, 0.9, 0.9, 0.9),
    las = 1,
    tcl = -0.1,
    mgp = c(1.5, 0.35, 0),
    cex = 1,
    cex.lab = 9/12,
    cex.axis = 7/12,
    bty = "n",
    lend = 2)

# Simple example with two regions and 4 tracks
bwScreenshot(
  bed= c("chr3R:30760926-30794202", "chr3R:30765926-30789202"),
  tracks= c("/groups/stark/vloubiere/projects/epigenetic_cancer/db/bw/ATAC/ATAC_PH18_merge.bw",
            "/groups/stark/vloubiere/projects/epigenetic_cancer/db/peaks/ATAC/ATAC_PH18_conf_peaks.narrowPeak",
            "/groups/stark/vloubiere/projects/epigenetic_cancer/db/bw/ATAC/ATAC_PHD11_merge.bw",
            "/groups/stark/vloubiere/projects/epigenetic_cancer/db/peaks/ATAC/ATAC_PHD11_conf_peaks.narrowPeak"),
  genome= "dm6"
)

# Genes only
bwScreenshot(
  bed= c("chr3R:30760926-30794202", "chr3R:30765926-30789202"),
  genome= "dm6"
)

}
