% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callOrfTagHits.R
\name{callOrftagHits}
\alias{callOrftagHits}
\title{ORFtrap_call_hits}
\usage{
callOrftagHits(
  sorted.forward.counts,
  unsorted.forward.counts,
  sorted.reverse.counts = NULL,
  unsorted.reverse.counts = NULL,
  genome,
  output.prefix,
  padj.cutoff = 0.001,
  log2OR.cutoff = 1,
  log2OR.pseudocount = 1,
  min.ins.cov.sorted = 0,
  min.ins.cov.unsorted = 0,
  output.folder = "db/FC_tables/ORFtag/",
  bed.output.folder = "db/FC_tables/ORFtag/bed/"
)
}
\arguments{
\item{sorted.forward.counts}{Character vector of sorted (FACSed) count file paths generated by vl_ORFtag_pipeline().
Counts from all provided files will be merged.}

\item{unsorted.forward.counts}{Character vector of unsorted (input) count file paths generated by vl_ORFtag_pipeline().
Counts from all provided files will be merged.}

\item{sorted.reverse.counts}{If specified together with unsorted.reverse.counts, enrichments will also be computed using
reversed insertions and the result will be added in additional '.rev' columns (useful to detect insertion traps).}

\item{unsorted.reverse.counts}{If specified together with sorted.reverse.counts, enrichments will also be computed
using reversed insertions and the result will be added in additional '.rev' columns (useful to detect insertion traps).}

\item{genome}{Genome annotation to be used.
For now, only "mm10" and "hg38" are supported, and the corresponding non-first-exon .gtf will be used for sorting.}

\item{output.prefix}{Prefix used for output file names.}

\item{padj.cutoff}{The p.adjust cutoff to be used to call hits (>=). Default= 0.001}

\item{log2OR.cutoff}{The log2OR cutoff to be used to call hits (>=). Default= 1}

\item{log2OR.pseudocount}{Pseudocount used to avoid infinite values. Note that only the log2OR will be affected, not the fisher
p.values. Default= 1.}

\item{min.ins.cov.sorted}{Minimum number of duplication counts for an insertion to be considered TRUE in the sorted sample.
Default= 0 (no filtering).}

\item{min.ins.cov.unsorted}{Minimum number of duplication counts for an insertion to be considered TRUE in the unsorted sample.
Default= 0 (no filtering).}

\item{output.folder}{Output folder for FC files. Default= "db/FC_tables/ORFtag".}

\item{bed.output.folder}{Output folder where the bed files containing usable insertions on + and - strands will be saved. Default= "db/FC_tables/ORFtag/bed/".}
}
\value{
Returns FC tables containing DESeq2-like columns.
}
\description{
Function to call hits from a given screen. Receives sorted and unsorted counts as input, computes FC table
}
\examples{
# Call hits ----
# 124 hits should be identified with this dataset
callOrftagHits(sorted.forward.counts = c("db/gene_assignment/Activator2_sort_rep1_same_strand.txt",
                                         "db/gene_assignment/Activator2_sort_rep2_same_strand.txt"),
               unsorted.forward.counts = c("db/gene_assignment/Activator2_input_rep1_same_strand.txt",
                                           "db/gene_assignment/Activator2_input_rep2_same_strand.txt"),
               sorted.reverse.counts = c("db/gene_assignment/Activator2_sort_rep1_rev_strand.txt",
                                         "db/gene_assignment/Activator2_sort_rep2_rev_strand.txt"),
               unsorted.reverse.counts = c("db/gene_assignment/Activator2_input_rep1_rev_strand.txt",
                                           "db/gene_assignment/Activator2_input_rep2_rev_strand.txt"),
               genome = "mm10",
               output.prefix = "Activator_2_vs_input",
               output.folder = "/scratch-cbe/users/vincent.loubiere/")

# To call hits using strand bias (not used)
see ?callOrftagHits_strandBias()

}
