% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_alignDnaRsubread.R
\name{cmd_alignDnaRsubread}
\alias{cmd_alignDnaRsubread}
\title{Generate Commands for DNA Sequence Alignment Using Rsubread}
\usage{
cmd_alignDnaRsubread(
  fq1,
  fq2 = NULL,
  output.prefix,
  genome = NULL,
  genome.idx = NULL,
  bam.output.folder = "db/bam/DNAseq/",
  alignment.stats.output.folder = "db/alignment_stats/DNAseq/",
  Rpath = "Rscript",
  cores = 8
)
}
\arguments{
\item{fq1}{A character vector of .fq (or .fq.gz) file paths.}

\item{fq2}{For paired-end data, a character vector of .fq (or .fq.gz) file paths matching
fq1 files. Default= NULL.}

\item{output.prefix}{Prefix for the output files.}

\item{genome}{Reference genome identifier  (supported genomes: 'dm6', 'mm10'). Required if
genome.idx is not provided. Default= NULL.}

\item{genome.idx}{Path to Bowtie2 index files (without extensions).
Can be automatically defined using the `genome` argument. Default= NULL.}

\item{bam.output.folder}{Directory for the output BAM file. Default= "db/bam.output.folder/".}

\item{alignment.stats.output.folder}{Directory for alignment statistics.
Default= "db/alignment_stats/".}

\item{Rpath}{Path to the Rscript binary. Default= "Rscript".}

\item{cores}{Number of CPU cores to use. Default= 8.}
}
\value{
A `data.table` with:
- `file.type`: Output file labels ("bam", "align.stats").
- `path`: Paths to the output files.
- `cmd`: Shell command to run Rsubread.
- `cores`: The number of CPU cores to use.
- `job.name`: Default name for the job = "alnRsub".
}
\description{
Creates shell commands to align gDNA sequencing reads to a reference genome
using the Rsubread aligner. Outputs a BAM file and alignment statistics.
}
\examples{
# Align single-end reads to the mm10 genome
cmd <- cmd_alignDnaRsubread(
  fq1 = "sample_R1.fq.gz",
  output.prefix = "sample",
  genome = "mm10"
)
vl_submit(cmd, execute= FALSE)

# Align paired-end reads using a custom genome index
cmd <- cmd_alignDnaRsubread(
  fq1 = "sample_R1.fq.gz",
  fq2 = "sample_R2.fq.gz",
  output.prefix = "sample",
  genome.idx = "/path/to/custom/genome/index"
)
vl_submit(cmd, execute= FALSE)

}
