% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_alignRsubread.R
\name{cmd_alignRsubread}
\alias{cmd_alignRsubread}
\title{Title}
\usage{
cmd_alignRsubread(
  fq1,
  fq2 = NULL,
  output.prefix,
  genome,
  genome.idx = NULL,
  bam.output.folder = "db/bam.output.folder/",
  alignment.stats.output.folder = "db/alignment_stats/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
