% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_assignInsertions.R
\name{cmd_assignInsertions}
\alias{cmd_assignInsertions}
\title{Title}
\usage{
cmd_assignInsertions(
  bam,
  output.prefix = NULL,
  genome,
  gtf,
  bed.output.folder = "db/bed/",
  counts.output.folder = "db/counts/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{bed.output.folder}{}
}
\description{
Title
}
