% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_bedgraphToBigwig.R
\name{cmd_bedgraphToBigwig}
\alias{cmd_bedgraphToBigwig}
\title{Title}
\usage{
cmd_bedgraphToBigwig(
  bdg,
  output.prefix,
  bw.output.folder = "db/bw/",
  genome,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
