% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_bedToBigwig.R
\name{cmd_bedToBigwig}
\alias{cmd_bedToBigwig}
\title{Title}
\usage{
cmd_bedToBigwig(
  bed,
  genome,
  output.prefix = NULL,
  bw.output.folder = "db/bw/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
