% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_cellRangerAggr.R
\name{cmd_cellRangerAggr}
\alias{cmd_cellRangerAggr}
\title{Generate Cell Ranger Aggregate command}
\usage{
cmd_cellRangerAggr(
  molecule_info_paths,
  output.prefix,
  output.folder = "db/snRNASeq_10X/aggregated_samples/",
  cores = 4,
  mem = 32
)
}
\arguments{
\item{molecule_info_paths}{A character vector of paths to 'molecule_info.h5' files.}

\item{output.prefix}{Prefix for output files.}

\item{output.folder}{Directory where output files will be saved. Default= 'db/scRNASeq_10X/aggregated_samples/'.}

\item{cores}{Number of CPU cores to use. Default= 4.}

\item{mem}{Memory in GB to allocate. Default= 32.}
}
\description{
Generate Cell Ranger Aggregate command
}
