% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_checkVBCfile.R
\name{cmd_checkVBCfile}
\alias{cmd_checkVBCfile}
\title{Extracts the first 10,000 of a VBC tar file}
\usage{
cmd_checkVBCfile(
  vbcTarFile,
  fq.output.folder = paste0("/scratch-cbe/users/vincent.loubiere/", gsub(".tar.gz", "",
    basename(vbcTarFile))),
  nreads = 10000L,
  cores = 8,
  mem = 16,
  overwrite = FALSE
)
}
\arguments{
\item{vbcTarFile}{Path to a vbc tar.gz file.}

\item{fq.output.folder}{Output folder where fq files should be saved. Default= "/scratch-cbe/users/vincent.loubiere/<tar_file_basename>/".}

\item{nreads}{Number of reads that should be processed (for testing purposes). Default= 10,000L.}

\item{cores}{Number of CPU cores to use for samtools processing (when using BAM input). Default= 8L.}

\item{mem}{Memory to use for the job (Go). Default= 16.}

\item{overwrite}{If output files already exist, should they be overwritten? Default= FALSE.}
}
\value{
A data.table with three columns:
\itemize{
  \item `file.type`: Labels for output files (e.g.: "fq1", "fq2").
  \item `path`: Full paths to the output files.
  \item `cmd`: Shell command(s) to generate the output files.
  \item `cores`: The number of CPU cores to use.
  \item `job.name`: Default name for the job = "demultVBC".
}
}
\description{
Generates the call to a perl script to extract the head of VBC tar files for debugging purposes
}
