% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_countBCreads.R
\name{cmd_countBCreads}
\alias{cmd_countBCreads}
\title{Title}
\usage{
cmd_countBCreads(
  bam,
  output.prefix = NULL,
  counts.output.folder = "db/counts/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
