% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_cutnrunQC.R
\name{cmd_cutnrunQC}
\alias{cmd_cutnrunQC}
\title{Generate commands for peaks QC}
\usage{
cmd_cutnrunQC(
  peaks.file,
  bw.sample,
  bw.input = NULL,
  genome,
  output.prefix,
  pdf.output.folder = "db/QC/CUTNRUN/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{peaks.file}{Path to the sample's .narrowpeak or .broadPeak file.}

\item{bw.sample}{Path to sample bigwig.}

\item{bw.input}{Optional path to the input bigwig file that was used for peak calling.}

\item{genome}{A BSgenome name (e.g. "mm10", "dm6"...).}

\item{output.prefix}{Prefix for the output files. If not specified, will be generated from the bw.sample file name.}

\item{pdf.output.folder}{Output folder where the pdf file should be saved. Default= "db/QC/CUTNRUN/".}

\item{Rpath}{Path to the Rscript binary. Default= "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}
}
\value{
A data.table with:
- `file.type`: Output file labels (e.g: "peaks", "bedgraph").
- `path`: Paths to the output files.
- `cmd`: Shell command to run the peak calling pipeline.
- `job.name`: Default name for the job = "MACS2".
}
\description{
Creates commands to do QC of individual CUTNRUN/ChIP-seq/ATAC-Seq peaks.
}
