% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_DESeq2.R
\name{cmd_DESeq2}
\alias{cmd_DESeq2}
\title{Generate Commands for 'default' DESeq2 analysis}
\usage{
cmd_DESeq2(
  count.files,
  sample.names,
  conditions,
  ctl.conditions = unique(conditions),
  output.prefix,
  dds.output.folder = "db/dds/RNASeq/",
  FC.tables.output.folder = "db/FC_tables/RNASeq/",
  MAplot.output.folder = "pdf/MAplots/RNASeq/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\description{
Generate Commands for 'default' DESeq2 analysis
}
