% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_download.R
\name{cmd_download}
\alias{cmd_download}
\title{Download files}
\usage{
cmd_download(
  url,
  output.name = basename(url),
  output.folder,
  output.file.type = NA,
  gzip = FALSE
)
}
\arguments{
\item{url}{URL to download}

\item{output.folder}{Path where the output file will be saved.}

\item{output.file.type}{The type of output file (fq...). Default= NA.}

\item{gzip}{Should the output file be gzipped? Default= FALSE}

\item{file.typename}{The output file name (default= keep url basename).}
}
\description{
Download files
}
