% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_extractUnalignedReads.R
\name{cmd_exractUnalignedReads}
\alias{cmd_exractUnalignedReads}
\title{Generate Commands for Sequence Alignment Using Bowtie2}
\usage{
cmd_exractUnalignedReads(
  bam,
  fq.output.folder = "db/fq/",
  alignment.stats.output.folder = "db/alignment_stats/",
  cores = 8
)
}
\arguments{
\item{cores}{}
}
\description{
Generate Commands for Sequence Alignment Using Bowtie2
}
