% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_extractUnalignedReadsFromBam.R
\name{cmd_exractUnalignedReadsFromBam}
\alias{cmd_exractUnalignedReadsFromBam}
\title{Extract Unaligned Reads from a BAM File}
\usage{
cmd_exractUnalignedReadsFromBam(
  bam,
  fq.output.folder = "db/fq/",
  alignment.stats.output.folder = "db/alignment_stats/",
  cores = 8
)
}
\arguments{
\item{bam}{Path to the input BAM file. Only a single BAM file is allowed.}

\item{fq.output.folder}{Directory for the output FASTQ file. Default= "db/fq/".}

\item{alignment.stats.output.folder}{Directory for alignment statistics (not used in this function). Default= "db/alignment_stats/".}

\item{cores}{Number of CPU cores to use for samtools processing. Default= 8.}
}
\value{
A data.table with:
- `file.type`: Output file label ("fq1.unaligned").
- `path`: Path to the output FASTQ file.
- `cmd`: Shell command to extract unaligned reads.
- `cores`: The number of CPU cores to use.
- `job.name`: Default name for the job = "extractUnal".
}
\description{
Creates shell commands to extract unaligned reads from a BAM file and output them as a FASTQ file.
}
\examples{
# Extract unaligned reads from a BAM file
cmd <- cmd_exractUnalignedReadsFromBam(
  bam = "/data/bam/sample.bam"
)
vl_submit(cmd, execute= FALSE)

}
