% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_filterBamChrs.R
\name{cmd_filterBamChrs}
\alias{cmd_filterBamChrs}
\title{Filters certain chromosomes from bam file}
\usage{
cmd_filterBamChrs(
  bam,
  seqnames.regexpr = c(Dmel = "^(chr2L|chr2R|chr3L|chr3R|chr4|chrX|chrY)$"),
  output.prefix = NULL,
  bam.output.folder = "db/bam/",
  cores = 6
)
}
\arguments{
\item{bam}{Path to the input BAM file. Only a single BAM file is allowed.}

\item{seqnames.regexpr}{A named vector of length 1 containing the regular expression
used to filter input bam based on seqnames. The name will be appended to the output.prefix 
to create the output basename. Default= c("Dmel"= "^(chr2L|chr2R|chr3L|chr3R|chr4|chrX|chrY)$"); i.e. canonical
Dmel chromosomes.}

\item{output.prefix}{Prefix for the output filtered bam file. If not provided, it is derived from the input BAM filename.}

\item{bam.output.folder}{Output folder where filtered bam files will be saved.}

\item{cores}{Number of CPU cores to use. Default= 6.}
}
\value{
A data.table containing the command and output file names.
}
\description{
Creates shell commands to convert a BAM file to BigWig format. Supports single-end and paired-end sequencing data,
with optional fragment extension for paired-end reads.
}
