% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_filter_loom_file.R
\name{cmd_filterLoom}
\alias{cmd_filterLoom}
\title{Title}
\usage{
cmd_filterLoom(
  seurat_object,
  input.loom.file = NULL,
  cdition,
  output.folder = "db/velocity/"
)
}
\arguments{
\item{output.folder}{}
}
\description{
Title
}
