% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_infer_candidate_regulons.R
\name{cmd_infer_candidate_regulons}
\alias{cmd_infer_candidate_regulons}
\title{Compute candidate regulons per gene}
\usage{
cmd_infer_candidate_regulons(
  genes.file,
  TFs.file,
  cell.labels.file,
  N.iterations = 100L,
  cell.frac = 0.3,
  stability.cutoff = 0.6,
  method = "ELASTIC",
  starting.line = NULL,
  ending.line = NULL,
  output.prefix,
  output.folder = "db/candidate_regulons/",
  cores = 20,
  mem = 40,
  time = "8:00:00"
)
}
\arguments{
\item{genes.file}{An .rds file containing SCT-normalized counts for the genes to be tested (dependent variables).}

\item{TFs.file}{An .rds file containing SCT-normalized counts for the predictor TFs.}

\item{cell.labels.file}{An .rds file containing a data.table containing "cluster" and "cdition" columns.}

\item{N.iterations}{The number of iterations to be performed. Default= 100L.}

\item{cell.frac}{The fractions of cells to be used per iteration (random sampling). Default= .3.}

\item{stability.cutoff}{The fraction of time a TF should receive a coefficient to be considered as a candidate
activator/repressor. Default= 0.6.}

\item{method}{Should be one of 'LASSO' (quick) or 'ELASTIC' (less affected by colinearity between predictors but slower).}

\item{output.prefix}{The prefix of the output file. If missing, the basename of the genes_file will be used. If starting
and ending lines are specified, they will be appended to the output.prefix to avoid overwritting.}

\item{cores}{Number of cores to be used. Default= 20.}

\item{mem}{Memory dedicated to the job. Default= 40.}

\item{time}{Time limit for the job. Default= "8:00:00".}
}
\description{
Compute candidate regulons per gene using elastic net or lasso regression.
}
