% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_MAGECK_ORFeome.R
\name{cmd_MAGeCK_ORFeome}
\alias{cmd_MAGeCK_ORFeome}
\title{Run MAGeCK Analysis for ORFeome Screens}
\usage{
cmd_MAGeCK_ORFeome(
  sample.counts,
  sample.names = gsub("^(.*?_rep\\\\d+).*", "\\\\1", basename(sample.counts)),
  input.counts,
  input.names = gsub("^(.*?_rep\\\\d+).*", "\\\\1", basename(input.counts)),
  sample.cutoff.FUN = function(x) sum(x) >= 3,
  input.cutoff.FUN = function(x) sum(x) >= 0,
  row.cutoff.FUN = function(x) sum(x) >= 3,
  pseudocount = 0,
  sort = "pos",
  paired = FALSE,
  logFC.cutoff = log2(1.5),
  FDR.cutoff = 0.05,
  master.table.rds,
  output.prefix,
  FC.output.folder = "db/FC_tables/ORFeome/",
  MAGeCK.output.folder = "db/FC_tables/ORFeome/MAGeCK/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{sample.counts}{A character vector of paths to sample count files.}

\item{sample.names}{For each sample.counts file, the name of the corresponding sample. By default, it will be inferred from basename(sample.counts).}

\item{input.counts}{A character vector of paths to input count files.}

\item{input.names}{For each input.counts file, the name of the corresponding sample. By default, it will be inferred from basename(input.counts).}

\item{sample.cutoff.FUN}{A function to filter sample rows. Default: `function(x) sum(x) >= 3`.}

\item{input.cutoff.FUN}{A function to filter input rows. Default: `function(x) sum(x) >= 0`.}

\item{row.cutoff.FUN}{A function to filter rows using both sample and input columns. Default: `function(x) sum(x) >= 3`.}

\item{pseudocount}{A numeric value specifying the pseudocount to add to zero values in sample and input columns. Default= 0.}

\item{sort}{A character string specifying whether positive ('pos') or negative ('neg') enrichments are expected. Default= 'pos'.}

\item{paired}{Logical. Should samples be analyzed as paired? Default= FALSE.}

\item{logFC.cutoff}{A numeric value specifying the logFC cutoff for calling hits. Default= log2(1.5).}

\item{FDR.cutoff}{A numeric value specifying the FDR cutoff for calling hits. Default= 0.05.}

\item{master.table.rds}{Path to a .rds file containing the master table that will be merged to MAGeCK output.
Should at least contain the ORF ids in a column named 'id'.}

\item{output.prefix}{Output files prefix (e.g. 'screen1_sample1_vs_input').}

\item{FC.output.folder}{Directory where final FC tables should be saved. Default= 'db/FC_tables/ORFeome/'.}

\item{MAGeCK.output.folder}{Directory where MAGeCK intermediate files should be saved. Default= 'db/FC_tables/ORFeome/'.}

\item{Rpath}{Path to the Rscript executable. Default: "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}
}
\value{
A data.table containing columns:
- `file.type`: Labels for the output files (i.e., "raw.counts.table", "filtered.counts", "gene.summary", "volcano.plot", "gene.summary.master").
- `path`: Paths to the output files.
- `cmd`: Shell commands to run the MAGeCK analysis pipeline.
- `job.name`: Default name for the job = "MAGECK".
}
\description{
This function generates shell commands to perform MAGeCK analysis for ORFeome screens.
It creates raw and filtered count tables, runs MAGeCK for differential analysis, generates MA plots,
and merges results with a master table. The function supports filtering, pseudocount addition, and paired analysis.
}
