% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_orfeomeDictionary.R
\name{cmd_orfeomeDictionary}
\alias{cmd_orfeomeDictionary}
\title{Generates ORFeome dictionary}
\usage{
cmd_orfeomeDictionary(
  BC.fqs,
  bam.ORF,
  minNreads = 3,
  output.prefix,
  output.folder = "db/dictionary/",
  cores = 8,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{BC.fqs}{A vector of paths to .fq.gz files containing the trimmed BC sequences.}

\item{bam.ORF}{Path to a .bam file containing reads aligned to ORF sequences.}

\item{minNreads}{Minimum number of supporting reads for a BC/ORF combination to be considered valid.}

\item{output.prefix}{Prefix for output files.}

\item{output.folder}{Output directory for result files. Default: "db/dictionary/".}

\item{Rpath}{Path to the Rscript binary. Default: "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}
}
\value{
A data.table with:
- `file.type`: Output file label ("peaks", "stats").
- `path`: Path to the output file.
- `cmd`: Shell command to run the peak calling pipeline.
- `job.name`: Default name for the job = "Dictionary".
}
\description{
Generates a shell command to identify confident ORFs/BCs combinations.
}
\examples{
cmd <- cmd_orfeomeDictionary(
  BC.fqs = "/data/exp.fq.gz",
  bam.ORF = "/data/aligned.bam"
)
vl_submit(cmd, execute = FALSE)

}
