% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_submit.R
\name{cmd_submit}
\alias{cmd_submit}
\title{Title}
\usage{
cmd_submit(
  cmd,
  cores = 8,
  mem = 16,
  time = "08:00:00",
  job.name = "myJob",
  logs = "db/logs",
  overwrite = FALSE,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  execute = TRUE
)
}
\arguments{
\item{execute}{Should the command be submitted to the server? Default= TRUE}
}
\description{
Title
}
