% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_umiCountsFromBam.R
\name{cmd_umiCountsFromBam}
\alias{cmd_umiCountsFromBam}
\title{Title}
\usage{
cmd_umiCountsFromBam(
  bam,
  layout,
  output.prefix = NULL,
  umi.counts.output.folder = "db/umi_counts/",
  collapsed.bed.output.folder = "db/bed/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
