% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_umiToBigwigProseq.R
\name{cmd_umiToBigwigProseq}
\alias{cmd_umiToBigwigProseq}
\title{Title}
\usage{
cmd_umiToBigwigProseq(
  umi.counts,
  output.prefix = NULL,
  bw.output.folder = "db/bw/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
