% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_velocyto.R
\name{cmd_velocyto}
\alias{cmd_velocyto}
\title{Generate Cell Ranger Alignment Command}
\usage{
cmd_velocyto(cellranger.output.folder, gtf, cores = 8, mem = 64)
}
\arguments{
\item{mem}{}
}
\description{
Constructs a shell command to run 10x Genomics Cell Ranger for single-cell RNA-seq alignment and quantification, with resource specification and output management.
}
