% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeImage.R
\name{compositeImage}
\alias{compositeImage}
\title{Creates a composite image by merging different channels}
\usage{
compositeImage(im, red = NULL, green = NULL, blue = NULL, white = NULL)
}
\arguments{
\item{im}{The image object from which channels should be extracted.}

\item{red}{Index of the color chanel that should be shown in red. Default= NULL.}

\item{green}{Index of the color chanel that should be shown in green.  Default= NULL.}

\item{blue}{Index of the color chanel that should be shown in blue. Default= NULL.}

\item{white}{Index of the color chanel that should be shown in white. Default= NULL.
If all colors are set to FALSE, then an empty image is returned.}
}
\description{
Creates a composite image by merging different channels
}
