% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contribSeqlets.R
\name{contribSeqlets}
\alias{contribSeqlets}
\title{Return seqLets with high contribution scores}
\usage{
contribSeqlets(contrib, ext.size = 50, log2OR.cutoff = 1, FDR.cutoff = 0.05)
}
\arguments{
\item{contrib}{A contribution data.table imported with ?importContrib.}

\item{ext.size}{The size by which candidate peaks should be extended to retrieve
local background. Default= 50.}

\item{log2OR.cutoff}{Only the peaks with |log2OR| >= log2OR.cutoff are returned.}

\item{FDR.cutoff}{Only the peaks with FDR <= FDR.cutoff are returned.}
}
\value{
A gr data.table containing significant peaks, based on provided cutoffs.
}
\description{
Return seqLets with high contribution scores
}
\examples{
dir <- "/groups/stark/vloubiere/projects/epiDeepCancer/db/model_PHD11//contributions/"
files <- list.files(dir, ".h5", recursive= T, full.names = T)
contrib <- vlite::importContrib(files)
peaks <- contribSeqlets(contrib)

}
