% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaProcessing.R
\name{dnaProcessing}
\alias{dnaProcessing}
\title{Align gDNA using rsubread}
\usage{
dnaProcessing(
  fq1,
  fq2 = NULL,
  output.prefix,
  genome,
  genome.idx = NULL,
  allowMultiOverlap = FALSE,
  fq.output.folder = "db/fq/DNASeq/",
  bam.output.folder = "db/bam/DNASeq/",
  alignment.stats.output.folder = "db/alignment_stats/DNASeq/",
  counts.stats.output.folder = "db/stats/DNASeq/",
  counts.output.folder = "db/counts/DNASeq/",
  Rpath = "Rscript",
  cores = 8
)
}
\arguments{
\item{fq1}{A character vector of .fq (or .fq.gz) file paths.}

\item{fq2}{For paired-end data, a character vector of .fq (or .fq.gz) file paths matching fq1 files. Default= NULL.}

\item{output.prefix}{Prefix for output files.}

\item{genome}{Reference genome identifier (e.g., "mm10", "hg38").}

\item{genome.idx}{Path to the Rsubread genome index. If NULL, derived from genome. Default= NULL.}

\item{fq.output.folder}{Directory for trimmed FASTQ files. Default= "db/fq/DNASeq/".}

\item{bam.output.folder}{Directory for aligned BAM files. Default= "db/bam/DNASeq/".}

\item{alignment.stats.output.folder}{Directory for alignment statistics. Default= "db/alignment_stats/DNASeq/".}

\item{Rpath}{Path to the Rscript binary. Default= "Rscript".}

\item{cores}{Number of CPU cores to use. Default= 8.}
}
\value{
A data.table with:
- `file.types`: Types of output files.
- `path`: Paths to the output files.
- `cmd`: Shell commands for each step in the pipeline.
- `cores`: The number of CPU cores to use.
- `job.name`: Default name for the job = paste0("DNA_", output.prefix).
}
\description{
Generates a pipeline of commands to trim gDNA sequencing reads
and align them using the rsubread package.
}
\examples{
# Process paired-end RNA-Seq data
cmd <- dnaProcessing(
  fq1 = c("sample1_R1.fq.gz", "sample1_R1_rep2.fq.gz"),
  fq2 = c("sample1_R2.fq.gz", "sample1_R2_rep2.fq.gz"),
  output.prefix = "sample1",
  genome = "hg38",
  cores = 8
)
vl_submit(cmd, execute= FALSE)

}
