% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_iCistarget.R
\name{download_iCistarget}
\alias{download_iCistarget}
\title{Download iCisTarget report}
\usage{
download_iCistarget(url, dir, folder_name)
}
\arguments{
\item{url}{URL where the reports is to be found}

\item{dir}{Directory where the report should be saved.}

\item{folder_name}{Name of the unziped subfolder. By default, the downloaded subfolder will be named using the iCisTarget job name.}
}
\description{
Download an iCisTarget report, uncompresses it and stores it in a subfolder
}
