% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropboxDownload.R
\name{dropboxDownload}
\alias{dropboxDownload}
\title{Download a File or Folder from Dropbox}
\usage{
dropboxDownload(remote.path, local.path)
}
\arguments{
\item{remote.path}{A character string specifying the source path in Dropbox.}

\item{local.path}{A character string specifying the local destination path for the downloaded file or folder.}
}
\value{
No return value. Executes the Dropbox Uploader script to download the file or folder.
}
\description{
Uses the Dropbox Uploader script to download a file or folder from a specified Dropbox path to a local path.
}
\examples{
\dontrun{
dropboxDownload("/remote/folder/file.txt", "local/file.txt")  # Download a file
dropboxDownload("/remote/folder", "local/folder")  # Download a folder
}

}
