% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropboxUpload.R
\name{dropboxUpload}
\alias{dropboxUpload}
\title{Upload a File to Dropbox}
\usage{
dropboxUpload(local.path, remote.path = "./")
}
\arguments{
\item{local.path}{A character string specifying the local file or folder path to upload.}

\item{remote.path}{A character string specifying the destination path in Dropbox.}
}
\value{
No return value. Executes the Dropbox Uploader script to upload the file or folder.
}
\description{
Uses the Dropbox Uploader script to upload a file or folder to a specified Dropbox path.
}
\examples{
\dontrun{
dropboxUpload("local/file.txt", "/remote/folder/file.txt")  # Upload a file
dropboxUpload("local/folder", "/remote/folder")  # Upload a folder
}

}
