% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichBed.R
\name{enrichBed}
\alias{enrichBed}
\title{Find Overlapping Regions Between Two Sets of Genomic Intervals}
\usage{
enrichBed(
  peaks,
  ctl.peaks,
  regions,
  pseudocount = 1,
  maxgap = -1L,
  minoverlap = 0L,
  ignore.strand = TRUE
)
}
\arguments{
\item{peaks}{Peaks regions in any format compatible with ?importBed.}

\item{ctl.peaks}{Control peak regions with no overlaps with peaks, in any format compatible with ?importBed.}

\item{regions}{Regions for which overlaps with peaks have to be assessed, in any format compatible with ?importBed.}

\item{pseudocount}{The pseudocount to use to estimate the enrichment using the ?fisher.test function. Default= 1.
Note that the p-value will not be affected.}

\item{maxgap}{A single integer specifying the maximum gap allowed between 2 ranges for them to
be considered as overlapping. Default= -1.}

\item{minoverlap}{A single integer specifying the minimum overlap between 2 ranges for them to
be considered as overlapping. Default= 0.}

\item{ignore.strand}{If set to FALSE, only overlaps between regions that are on the same strand will be counted.
If set to TRUE (default), overlaps on both strands are counted.}
}
\value{
A list with elements:
\itemize{
  \item peaks.total: total number of peaks.
  \item peaks.overlaps: number of peaks overlapping regions.
  \item p.value: fisher test p.value.
  \item estimate: the fisher test odd ratio, based on the specified pseudocount.
}
}
\description{
A wrapper around ?GenomicRanges::findOverlaps() that identifies, for each genomic range in a,
the overlapping regions in b.
}
\examples{
peaks <- data.table(seqnames= "chr2R",
                    start= seq(1000, 10000, 1000))
peaks[, end:= start+100]
ctl.peaks <- resizeBed(peaks, "end", -100, 200)
enrichBed(peaks = peaks[1:5],
          ctl.peaks = ctl.peaks,
          regions = peaks)
}
