% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherTests.R
\name{fisherTests}
\alias{fisherTests}
\title{fisherTests}
\usage{
fisherTests(x.list, y.list, universe = NULL, alternative = "two.sided")
}
\arguments{
\item{x.list}{A list of elements to be compared to y.list in the search of overlaps.}

\item{y.list}{A list of elements to be compared to x.list in the search of overlaps.}

\item{universe}{If provided, will define the universe (background) of the test. Default
is to use the intersection between the (unique) values of x.list and y.list.}

\item{alternative}{The alternative to use for the fisher test. Default= "two.sided".}
}
\description{
Perform pairwise fisher tests between the elements of two lists.
}
