% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianBlur.R
\name{gaussianBlur}
\alias{gaussianBlur}
\title{applies a gasussian blur}
\usage{
gaussianBlur(x)
}
\arguments{
\item{x}{A numeric vector of the signal to be blurred}
}
\value{
Smoothed signal
}
\description{
radius= +/- 10 bins
}
