% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSgenomeSize.R
\name{getBSgenomeSize}
\alias{getBSgenomeSize}
\title{Get BS genome sizes}
\usage{
getBSgenomeSize(genome)
}
\arguments{
\item{genome}{BSgenome object to use. Example "dm3", "mm10"...}
}
\value{
data.table bed containing chromosomes start and (seqLength) coordinates
}
\description{
Very fast using data.table
}
\examples{
getBSgenomeSize("mm10")

}
