% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSsequence.R
\name{getBSsequence}
\alias{getBSsequence}
\title{Get genomic sequence}
\usage{
getBSsequence(bed, genome)
}
\arguments{
\item{bed}{Bed file for which regions have to be returned}

\item{genome}{BSgenome ID ("dm3", "dm6"...)}
}
\value{
Character vector of sequences
}
\description{
Returns the sequences of a bed.
}
\examples{
getBSsequence(vl_SUHW_top_peaks, "dm3")

}
