% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelEquation.R
\name{getModelEquation}
\alias{getModelEquation}
\title{model equation}
\usage{
getModelEquation(model, ...)
}
\arguments{
\item{model}{Model}

\item{...}{Arguments to 'format', such as 'digits' and 'trim'.}
}
\value{
Equation
}
\description{
Returns the equation from lm model. The code was found at https://stats.stackexchange.com/questions/63600/how-to-translate-the-results-from-lm-to-an-equation
}
