% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelExpVar.R
\name{getModelexpVar}
\alias{getModelexpVar}
\title{Compute explained variance}
\usage{
getModelexpVar(model)
}
\arguments{
\item{model}{An object containing the results returned by a model fitting function (e.g., lm or glm).}
}
\value{
% of explained variance
}
\description{
From Bernardo, computes the % explained variance for each predictor
}
