% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelRMSErsq.R
\name{getModelRMSErsq}
\alias{getModelRMSErsq}
\title{Compute RMSE and R2 from observed and predicted values}
\usage{
getModelRMSErsq(observed, predicted)
}
\arguments{
\item{observed}{vector of observed values}

\item{predicted}{vector of predicted values}
}
\value{
RMSE and R2
}
\description{
Comes from https://www.pluralsight.com/guides/linear-lasso-and-ridge-regression-with-r
}
