% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrepelScatterplot.R
\name{ggrepelScatterplot}
\alias{ggrepelScatterplot}
\title{ggrepelScatterplot}
\usage{
ggrepelScatterplot(
  x,
  y,
  col,
  label = rep("", length(x)),
  label.col = "black",
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  cex = 1,
  cex.pts = 1,
  cex.lab = 1,
  plot = T
)
}
\arguments{
\item{x}{x variable.}

\item{y}{y variable.}

\item{col}{Color of the dots.}

\item{label}{Labels to be added.}

\item{label.col}{Label color.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{main}{title.}

\item{cex}{Expansion factor for point size.}

\item{point_size}{point_size parameter ggrepel.}

\item{label_size}{label_size parameter ggrepel.}
}
\description{
A wrapper around ggrepel to plot a scatterplot with repel labels, but with parameters
matching the names of default R graphics.
}
