% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gHeatkey.R
\name{gHeatkey}
\alias{gHeatkey}
\title{Add Color Key (Legend) to Heatmap}
\usage{
gHeatkey(
  col,
  breaks,
  left = par("usr")[2] + diff(grconvertX(c(0, 1), "line", "user")),
  top = par("usr")[4],
  width = 0.75,
  height = 4,
  cex = 1,
  main = NA
)
}
\arguments{
\item{col}{Color vector of length(breaks)-1.}

\item{breaks}{Numeric vector of break points for color mapping, of length(col)+1.}

\item{left}{Numeric position for the left edge of the color key. Default is
calculated based on the current plot region plus one line width.}

\item{top}{Numeric position for the top edge of the color key. Default is
the top of the current plot region.}

\item{width}{Numeric width of the color key in lines (0.75 by default).}

\item{height}{Numeric height of the color key in lines (4 by default).}

\item{cex}{Numeric scaling factor for the heatkey. Default= 1.}

\item{main}{Character string for the title of the color key (NA by default).}

\item{border}{Color for the border of the color key. Use NA for no border
("black" by default).}

\item{show.breaks}{Logical; whether to show break values (TRUE by default).}
}
\value{
Adds a color key to the current plot. No return value.
}
\description{
Plots a color key (legend) for heatmaps.
This function is designed to work alongside heatmapvisualizations
and should be called after the main heatmap plot.
}
\details{
The function creates a vertical color key that maps colors to values in your
heatmap. It supports both continuous gradients and discrete color categories,
and can be customized in terms of position, size, and appearance.
}
\examples{
# Create example matrix
set.seed(1234)
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")

# Basic heatmap with default color key
gPar()
gImage(test)
gHeatkey()

}
\seealso{
\code{\link{gHeatmap}} for creating the main heatmap
}
