% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicScreenshot.R
\name{hicScreenshot}
\alias{hicScreenshot}
\title{Function to plot a HiC map with bw tracks on the bottom}
\usage{
hicScreenshot(
  mcool.file,
  misha.track,
  region,
  resolution,
  map.name = gsub(".mcool$|.track$", "", basename(mcool.file)),
  pdf.file = NULL,
  screenshot.cex.height = 1,
  tracks = character(),
  track.names = function(x) gsub("(.*)[.].*$", "\\\\1", x),
  col = colorRampPalette(c("grey60", "grey10"))(length(tracks)),
  bw.max = NA,
  nbins = 500,
  genome,
  gtf,
  sel.gene.symbols = NULL,
  border.col = NA,
  border.lwd = 1,
  bw.min = NA,
  ngenes = 1,
  cex.gene.symbol = 1,
  offset.gene.symbol = 0.25,
  col.gene.ps = "tomato",
  col.gene.ns = "cornflowerblue",
  exon.border = NA,
  gtf.transcript = NULL,
  gtf.exon = "exon",
  gtf.symbol = "gene_symbol",
  gtf.transcript.id = "transcript_id",
  region.width = 1,
  space.width = 1,
  bw.height = 1,
  bed.height = 1,
  space.height = 1,
  gene.height = 1,
  gene.space.height = 1,
  add = FALSE
)
}
\arguments{
\item{mcool.file}{Path to the misha file to be plotted.}

\item{region}{String specifying a unique region to plot, in the format "chr2L:16300000-16600000".}

\item{resolution}{The resolution to be used. Only the ones existing in the mcool file will work (see ?HiCExperiment::availableResolutions)}

\item{map.name}{The name of the hic map. By default, the basename of the mcool file will be used.}

\item{pdf.file}{Output pdf file. Should automatically adapts to the tracks being plotted.}

\item{screenshot.cex.height}{Expansion factor to apply to the height of the bottom screenshot. Default= 1.
For all other arguments, please refer to ?vlite::bwScreenshot.}
}
\value{
A plot of a HiC matrix with specific tracks
}
\description{
Wrapper around HiCExperiment::import and vlite::bwScreenshot to plot a HiC matrix with bw tracks.
}
