% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importContrib.R
\name{importContrib}
\alias{importContrib}
\title{Import contributions object}
\usage{
importContrib(h5, bed = NULL, fa = NULL, FUN = function(x) rowSums(x))
}
\arguments{
\item{h5}{A vector of paths to h5 file(s) containing the contribution scores.}

\item{bed}{Optional path(s) to Bed file(s) containing the coordinates for which contributions were computed.}

\item{fa}{Optional path(s) to fasta file(s) containing the sequences for which contributions were computed.}

\item{FUN}{The function to apply to each layer of the h5 object to retrieve the contribution scores.
By default, each layer is a matrix of dimensions nchar(sequence)*4. Thus, default= function(x) rowSums(x).}
}
\value{
A contribution data.table containing, for each region, the contribution scores (as list) and the corresponding sequence.
}
\description{
Import contributions object
}
\examples{
folder <- "/groups/stark/vloubiere/projects/epiDeepCancer/db/model_PH18/contributions/"
files <- list.files(folder, "h5$", recursive = TRUE, full.names = TRUE)
dat <- importContrib(h5= files)

}
