% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFq.R
\name{importFq}
\alias{importFq}
\title{importFq}
\usage{
importFq(fq.path, head)
}
\arguments{
\item{fq.path}{Path to the .fq of .fq.gz file to import.}

\item{head}{An optional integer specifying the number of reads that should be imported.}
}
\value{
A data.table with columns:
 - readID
 - Sequence
}
\description{
A wrapper to import fq files using data.table::fread().
}
