% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importJASPAR.R
\name{importJASPAR}
\alias{importJASPAR}
\title{importJASPAR}
\usage{
importJASPAR(
  combinedFile,
  pseudocount = 0.00001,
  bg = c(A = 0.25, C = 0.25, G = 0.25, T = 0.25)
)
}
\arguments{
\item{combinedFile}{Path to a combined .pfm file containing JASPAR-formatted raw PFMs.}

\item{pseudocount}{Pseudocount to be added to PPMs before logging. Default= 1e-5.}

\item{bg}{Background probabilities. Default= c("A"= .25, "C"= .25, "G"= .25, "T"= .25).}
}
\value{
A list containing PPM and PWM matrices stored in TFBSTools::PWMatrix sublists.
}
\description{
Imports a .pfm text file containing JASPAR PFMs and parse them into lists of PPM and PWM TFBSTools::PWMatrix objects.
}
\examples{
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
importJASPAR(pfm.file)

}
