% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMetadataSheet.R
\name{importMetadataSheet}
\alias{importMetadataSheet}
\title{See most recent error file in a directory}
\usage{
importMetadataSheet(metadata, first.col.name = "user")
}
\arguments{
\item{metadata}{Path to an excel metadata sheet}

\item{first.col.name}{If metadata in in .xlsx format, this character string will be used to detect the starting line. Default= "user".}
}
\description{
See most recent error file in a directory
}
