% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxProjectionImage.R
\name{maxProjectionImage}
\alias{maxProjectionImage}
\title{Maximum project of an image.}
\usage{
maxProjectionImage(im, start = 1, end = dim(im)[4])
}
\arguments{
\item{im}{Image to process.}

\item{start}{The first z-section to be included in the max projection.}

\item{end}{The last z-section to be included in the max projection.}
}
\description{
Maximum project of an image.
}
