% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifLassoRegression.R
\name{motifLassoRegression}
\alias{motifLassoRegression}
\title{LASSO regression using TF motifs}
\usage{
motifLassoRegression(response, counts)
}
\arguments{
\item{response}{Variable to predict}

\item{counts}{Matrix of motif counts}
}
\value{
Return a LASSO model
}
\description{
LASSO regression using TF motifs
}
\examples{
test <- motifLassoRegression(ATACSeq_FC, counts)

}
