% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifPosToBed.R
\name{motifPosToBed}
\alias{motifPosToBed}
\title{motifPosToMatrix}
\usage{
motifPosToBed(motifPos)
}
\arguments{
\item{motifPos}{A ?vl_motifPos() output data.table.}

\item{seqWidth.}{The width of the sequences for which motif positions have been computed.}

\item{cleanup.cache}{If set to TRUE, overwrites cached intermediate results. Default= FALSE.}
}
\value{
Returns a list of maximum scores' matrices (one per motif).
}
\description{
Transforms the output of vl_motifPos() to a list of scores' matrices.
}
