% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfeome_pipeline.R
\name{orfeomeProcessing}
\alias{orfeomeProcessing}
\title{Process OREfome Sequencing Data}
\usage{
orfeomeProcessing(
  vbcFile,
  layout,
  i7,
  i5 = "none",
  fq1,
  output.prefix,
  fq.output.folder = "db/fq/ORFeome/",
  bam.output.folder = "db/bam/ORFeome/",
  counts.output.folder = "db/counts/ORFeome/",
  alignment.stats.output.folder = "db/alignment_stats/ORFeome/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 8
)
}
\description{
Process OREfome Sequencing Data
}
