% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orftag_pipeline.R
\name{orftagProcessing}
\alias{orftagProcessing}
\title{Process ORtag Sequencing Data}
\usage{
orftagProcessing(
  vbcFile,
  layout,
  i7,
  i5 = "none",
  fq1,
  output.prefix,
  genome,
  genome.idx = NULL,
  gtf = NULL,
  fq.output.folder = "db/fq/ORFtag/",
  bam.output.folder = "db/bam/ORFtag/",
  alignment.stats.output.folder = "db/alignment_stats/ORFtag/",
  bed.output.folder = "db/bed/ORFtag/",
  counts.output.folder = "db/counts/ORFtag/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 8
)
}
\arguments{
\item{cores}{}
}
\description{
Process ORtag Sequencing Data
}
