% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orftagQC.R
\name{orftagQC}
\alias{orftagQC}
\title{Perform QC of ORFtag screen batches}
\usage{
orftagQC(
  sampleIDs,
  align.stats,
  counts.same.strand,
  dc.cutoff = 0,
  output.prefix,
  output.folder = "db/QC/ORFtag/",
  pdf.width = 9,
  pdf.height = 9
)
}
\arguments{
\item{sampleIDs}{A vector of sample IDs.}

\item{align.stats}{A vector of paths to alignment statistics files matching the sampleIDs.}

\item{counts.same.strand}{A vector of path of assigned insertions files matching the sampleIDs.}

\item{dc.cutoff}{The minimum number of duplicate counts for an insertion to be considered. Default= 0}

\item{output.prefix}{The output prefix for the output files.}

\item{output.folder}{Folder where output files will be saved. Default= "db/QC/ORFtag/".}

\item{pdf.width}{Width of the output pdf in inches. Default= 7.}

\item{pdf.height}{Height of the output pdf in inches. Default= 7.}
}
\description{
Perform QC of ORFtag screen batches
}
