% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_plot_auc_enrichment.R
\name{plot.vl_auc}
\alias{plot.vl_auc}
\title{Title}
\usage{
\method{plot}{vl_auc}(
  auc,
  cutoff = ifelse(value.var == "NES", 3, 0.6),
  min.diff.cutoff = 0,
  topN = Inf,
  order.var = "score",
  value.var = "NES",
  cluster.rows = T,
  cluster.cols = F,
  ranks.as.rows = FALSE,
  kmeans.k = NA,
  show.row.clusters = "right",
  show.numbers = NULL,
  legend.cex = 1,
  numbers.cex = 0.5,
  pdf.file = NULL
)
}
\arguments{
\item{auc}{A vl_auc object created by ?vl_motifAUC.}

\item{cutoff}{Cutoff on the maximum value across all conditions.}

\item{min.diff.cutoff}{Cutoff on the minimum difference between lowest and highest value.}

\item{topN}{After sorting the table using order.var (decreasing= T), subset only topN values for ploting.}

\item{order.var}{Value used to sort the data before subsetting topN values. Possible values are 'score' (use
the absolute of the value.var) or 'diff' (sort by difference between highest and lowest value.var).}

\item{value.var}{The value to be plotted. Possible value are "NES" or "auc". Default= "NES".}

\item{...}{Extra arguments to be passed to the heatmap funciton.}
}
\value{
Invisible returns the final matrix.
}
\description{
Title
}
