% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vl_enr.R
\name{plot.vl_enr}
\alias{plot.vl_enr}
\title{Plot Enrichment Results for vl_enr Objects}
\usage{
\method{plot}{vl_enr}(
  obj,
  log2OR.abs.cutoff = 0,
  padj.cutoff = 0.05,
  top.enrich = Inf,
  min.counts = 3L,
  order = "log2OR",
  xlab = "Odd Ratio (log2)",
  breaks = NULL,
  col = c("blue", "red")
)
}
\arguments{
\item{obj}{A data.table of class 'vl_enr' containing columns 'padj', 'log2OR' and 'set_hit'.
See ?vl_motifEnrich for an example.}

\item{log2OR.abs.cutoff}{Numeric cutoff used to filter absolute log2OR values before plotting.}

\item{padj.cutoff}{Numeric cutoff used to filter adjusted p-values before plotting.
Default= 0.05.}

\item{top.enrich}{Integer specifying the maximum number of top enrichments to plot, based on the
'order' parameter. Default= Inf (no selection).}

\item{min.counts}{Integer speifying the minimum number of counts required to include an enrichment
in the plot ('set_hit' column). Default= 3.}

\item{order}{The metric that should be used to order enrichments before selecting the top enrichments.
Possible values are 'padj' or 'log2OR'. Default= 'log2OR'.}

\item{xlab}{Label for the x-axis of the barplot. Default= 'Odd Ratio (log2)'.}

\item{breaks}{Numeric vector specifying the color breaks to use for the color scale. Defaults to the range of
filtered -log10(adjusted p-values).}

\item{col}{Color scale to use for the barplot. Default= c('blue', 'red').}
}
\value{
Invisibly returns the filtered and plotted data table.
}
\description{
This function generates a barplot to visualize enrichment results from an object of class 'vl_enr'.
The bars represent the log2 odds ratio (log2OR) of enrichment, and the color scale corresponds to
the adjusted p-values (-log10 transformed).
}
\examples{
# Example usage:
For an example using GO enrichment, see ?vl_GOenrich()
For an example using motif enrichment, see ?vl_motifEnrich()

# Assuming `enrichment_results` is an object of class `vl_enr`:
plot.vl_enr(enrichment_results, padj.cutoff = 0.01, top.enrich = 10, min.counts = 5)

}
