% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vl_enrRank.R
\name{plot.vl_enrRank}
\alias{plot.vl_enrRank}
\title{Function method to plot motifEnrichRanks recovery curves}
\usage{
\method{plot}{vl_enrRank}(
  obj,
  sel.motif,
  sel.group = NULL,
  col = "blue",
  mean.curve.col = "grey30",
  sd.curve.col = "limegreen",
  main = sel.motif
)
}
\arguments{
\item{obj}{An object as returned by ?motifEnrichRanks.}

\item{sel.motif}{The motif for which empirical cumulative recovery curve should be plotted.}

\item{col}{The color for the curve. Defaults= "blue.}

\item{mean.curve.col}{The color for the curve. Defaults= "grey30".}

\item{sd.curve.col}{The color for the curve. Defaults= "limegreen".}
}
\value{
A recovery plot
}
\description{
Function method to plot motifEnrichRanks recovery curves
}
