% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vl_iCisEnr.R
\name{plot.vl_iCisEnr}
\alias{plot.vl_iCisEnr}
\title{Function method to plot motifEnrichRanks recovery curves}
\usage{
\method{plot}{vl_iCisEnr}(
  enr,
  NES.cutoff = 3,
  simplify.clusters = FALSE,
  main = "Enriched motifs",
  legend.title = "NES",
  cluster.rows = TRUE,
  cluster.cols = FALSE,
  show.legend = TRUE,
  legend.cex = 0.6,
  show.row.clusters = FALSE,
  show.numbers = TRUE,
  numbers.cex = 0.4,
  add.logos =
    "/zssd/scratch/vincent.loubiere/motifs_db/JASPAR_CORE/20260812_JASPAR2026_CORE_NON_REDUNDANT.rds",
  logo.cex.width = 0.7,
  logo.cex.height = 0.7,
  ...
)
}
\arguments{
\item{enr}{An enrichment table as returned by ?vl_iCisEnrich.}

\item{NES.cutoff}{The NES cutoff to select enriched motifs. Default= 3.}

\item{simplify.clusters}{Should only the best motif per clsuter be returned? Default= FALSE.}

\item{main}{Title. Default= "Enriched motifs".}

\item{legend.title}{Default= "NES".}

\item{cluster.rows}{Default= TRUE.}

\item{cluster.cols}{Default= FALSE.}

\item{show.legend}{Default= TRUE.}

\item{legend.cex}{Default= 0.6.}

\item{show.row.clusters}{Default= FALSE.}

\item{show.numbers}{Default= TRUE/}

\item{numbers.cex}{Default= 0.4.}

\item{add.logos}{If a path to a .rds pwm db is provided, corresponding logos will be added
on the left of the heatmap. Default= "/zssd/scratch/vincent.loubiere/motifs_db/JASPAR_CORE/20260812_JASPAR2026_CORE_NON_REDUNDANT.rds".}

\item{logo.cex.width}{Width expansion factor for motif logos. Default= .7.}

\item{logo.cex.height}{Height expansion factor for motif logos. Default= .7.}

\item{...}{Additional arguments will be passed to ?vl_heatmap.}
}
\value{
A recovery plot
}
\description{
Function method to plot motifEnrichRanks recovery curves
}
