% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R
\name{plotImage}
\alias{plotImage}
\title{Title}
\usage{
plotImage(im, pxSizeX_um = NULL, scale.bar = 100, all = FALSE)
}
\arguments{
\item{im}{}

\item{jpeg.file}{}

\item{main}{}

\item{col}{}

\item{cex}{}
}
\description{
Title
}
