% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proseq_pipeline.R
\name{proseqProcessing}
\alias{proseqProcessing}
\title{Process PRO-Seq Sequencing Data}
\usage{
proseqProcessing(
  vbcFile,
  layout,
  eBC,
  i7,
  i5 = "none",
  fq1,
  fq2 = NULL,
  output.prefix,
  ref.genome,
  ref.genome.idx = NULL,
  spike.genome,
  spike.genome.idx = NULL,
  ref.gtf = NULL,
  fq.output.folder = "db/fq/PROSeq/",
  bam.output.folder = "db/bam/PROSeq/",
  alignment.stats.output.folder = "db/alignment_stats/PROSeq/",
  counts.output.folder = "db/counts/PROSeq/",
  counts.stats.output.folder = "db/stats/PROSeq/",
  bw.output.folder = "db/bw/PROSeq/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 8
)
}
\description{
Process PRO-Seq Sequencing Data
}
