% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwmPercToLog.R
\name{pwmPercToLog}
\alias{pwmPercToLog}
\title{PWM perc to log2}
\usage{
pwmPercToLog(
  perc_pwm,
  pseudocount = 1e-05,
  bg = c(A = 0.25, C = 0.25, G = 0.25, T = 0.25)
)
}
\arguments{
\item{perc_pwm}{A percentage PWM, where all columns sum to 1}

\item{pseudocount}{Pseudocount. Default= 1e-5}

\item{bg}{Background probabilities. Default= c("A"= .25, "C"= .25, "G"= .25, "T"= .25).}
}
\value{
Returns a log2 odd ratio pwm
}
\description{
PWM perc to log2
}
\examples{
vl_Dmel_motifs_DB_full$pwms_perc[[5]]@profileMatrix
pwmPercToLog(vl_Dmel_motifs_DB_full$pwms_perc[[5]]@profileMatrix)
}
