% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRegionsBed.R
\name{randomRegionsBed}
\alias{randomRegionsBed}
\title{Randomly Sample Control Genomic Regions}
\usage{
randomRegionsBed(
  bed,
  widths = rep(100, 1000),
  replace = T,
  no.overlaps = NULL,
  ignore.strand = TRUE
)
}
\arguments{
\item{bed}{Input genomic ranges (in any format compatible with ?importBed) from which random regions
are drawn.}

\item{widths}{Integer vector specifying the width of the regions to sample. The length of the vector
determines the number of sampled regions.}

\item{no.overlaps}{Genomic ranges for which overlaps should be avoided during sampling, in any format compatible
with ?importBed. Default= NULL.}

\item{ignore.strand}{If genomic ranges are provided in no.overlaps, only overlaps on the same strand will be
avoided. If set to TRUE (default), overlaps on both strands are avoided.}
}
\value{
A data.table of sampled control regions.
}
\description{
Randomly sample genomic regions from a bed with replacement, matching specified widths and potentially
avoiding overlaps with another set or regions.
}
\examples{
# Draw regions from dm6 canonical chromosomes
dm6.sizes <- getBSgenomeSize(genome= "dm6")
dm6.sizes <- dm6.sizes[seqnames \%in\% c("chr2L", "chr2R")]
randomRegionsBed(bed= dm6.sizes,  widths= rep(10, 10))

# Avoid overlaps with specific region
randomRegionsBed(bed= dm6.sizes, widths= rep(10, 10), no.overlaps= dm6.sizes[2])

}
