% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterScatterplot.R
\name{rasterScatterplot}
\alias{rasterScatterplot}
\title{Title}
\usage{
rasterScatterplot(
  x,
  y = NULL,
  pch = 16,
  col = adjustcolor("grey", 0.7),
  cex = 0.8,
  type = "p",
  frame = F,
  res = 600L,
  size = 2000L,
  xlab = NULL,
  ylab = NULL,
  xaxt = "s",
  yaxt = "s",
  xaxs = "r",
  yaxs = "r",
  axes = T,
  add = F,
  tmpdir = tempdir(),
  ...
)
}
\arguments{
\item{y}{default= NULL}

\item{type}{default= "p"}

\item{frame}{Default= F}

\item{res}{Nominal resolution of png image. default= 600}

\item{size}{Size of the png image. Default= 2000}

\item{...}{Extra plotting parameters}
}
\value{
Raster scatterplot
}
\description{
Title
}
\examples{
rasterScatterplot(1:3)
}
