% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizeBed.R
\name{resizeBed}
\alias{resizeBed}
\title{Resize Genomic Regions in a BED File}
\usage{
resizeBed(
  bed,
  center = "center",
  upstream = 500L,
  downstream = 500L,
  genome = NULL,
  ignore.strand = FALSE
)
}
\arguments{
\item{bed}{Input genomic ranges in a format compatible with ?importBed().}

\item{center}{Specifies the origin for resizing. Options are:
\itemize{
  \item `center`: Resize symmetrically around the midpoint of the region (default).
  \item `start`: Resize from the start of the region.
  \item `end`: Resize from the end of the region.
  \item `region`: Extend or shorten the entire region's boundaries.
}}

\item{upstream}{Integer. Number of bases to extend upstream from the specified origin. Default= 500.}

\item{downstream}{Integer. Number of bases to extend downstream from the specified origin. Default= 500L.}

\item{genome}{A BS genome name (e.g. "dm6", "mm10").
If specified, resized regions exceeding chromosome lengths will be clipped. Default= NULL.}

\item{ignore.strand}{Logical. If set to TRUE, upstream and downstream resizing is performed irrespective
of the strand. Default= FALSE.}
}
\description{
Adjusts the size of genomic coordinates.
}
\details{
Resized regions are returned as a gr data.table containing any additional columns from the input.
}
\examples{
# Import example BED file
bed <- importBed(c("chr2L:1000-2000:+", "chr2L:1000-2000:-"))

# Resize using different centering options
resizeBed(bed, center = "start", upstream = 500, downstream = 1000)[]
resizeBed(bed, center = "center", upstream = 500, downstream = 1000)[]
resizeBed(bed, center = "end", upstream = 500, downstream = 1000)[]
resizeBed(bed, center = "region", upstream = 500, downstream = 1000)[]

# Ignore strand during resizing
resizeBed(bed, center = "center", upstream = 500, downstream = 1000, ignore.strand = TRUE)[]

}
