% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revCompDNA.R
\name{revCompDNA}
\alias{revCompDNA}
\title{reverseComplement}
\usage{
revCompDNA(DNA_char, complement = T, reverse = T)
}
\arguments{
\item{DNA_char}{DNA character string}
}
\value{
Reverse Complemented DNA sequence
}
\description{
Reverse complement DNA sequence character string
}
\examples{
revCompDNA("ATCG")

}
