% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcessing.R
\name{rnaseqProcessing}
\alias{rnaseqProcessing}
\title{Process RNA-Seq Data}
\usage{
rnaseqProcessing(
  fq1,
  fq2 = NULL,
  output.prefix,
  genome,
  genome.idx = NULL,
  annotation.file = NULL,
  GTF.attrType.extra = NULL,
  allowMultiOverlap = FALSE,
  fq.output.folder = "db/fq/RNASeq/",
  bam.output.folder = "db/bam/RNASeq/",
  alignment.stats.output.folder = "db/alignment_stats/RNASeq/",
  counts.stats.output.folder = "db/stats/RNASeq/",
  counts.output.folder = "db/counts/RNASeq/",
  bw.output.folder = "db/bw/RNASeq/",
  Rpath = "Rscript",
  cores = 8
)
}
\arguments{
\item{fq1}{A character vector of .fq (or .fq.gz) file paths.}

\item{fq2}{For paired-end data, a character vector of .fq (or .fq.gz) file paths matching fq1 files. Default= NULL.}

\item{output.prefix}{Prefix for output files.}

\item{genome}{Reference genome identifier (e.g., "mm10", "hg38").}

\item{genome.idx}{Path to the Rsubread genome index. If NULL, derived from genome. Default= NULL.}

\item{annotation.file}{Path to a '.gtf', '.gtf.gz' or '.saf' annotation file. If genome is specified,
the corresponding gtf file will be used. SAF format should contain: GeneID, Chr, Start, End, Strand. Default= NULL.}

\item{GTF.attrType.extra}{Additional GTF attribute to include in the output (e.g., gene symbol). Default= NULL.}

\item{allowMultiOverlap}{If a read overlaps more than one feature, should be assigned to all overlapping features? Default= FALSE.}

\item{fq.output.folder}{Directory for trimmed FASTQ files. Default= "db/fq/RNASeq/".}

\item{bam.output.folder}{Directory for aligned BAM files. Default= "db/bam/RNASeq/".}

\item{alignment.stats.output.folder}{Directory for alignment statistics. Default= "db/alignment_stats/RNASeq/".}

\item{counts.stats.output.folder}{Directory for gene count statistics. Default= "db/stats/RNASeq/".}

\item{counts.output.folder}{Directory for gene counts. Default= "db/counts/RNASeq/".}

\item{bw.output.folder}{Directory for BigWig files. Default= "db/bw/RNASeq/".}

\item{Rpath}{Path to the Rscript binary. Default= "Rscript".}

\item{cores}{Number of CPU cores to use. Default= 8.}
}
\value{
A data.table with:
- `file.types`: Types of output files.
- `path`: Paths to the output files.
- `cmd`: Shell commands for each step in the pipeline.
- `cores`: The number of CPU cores to use.
- `job.name`: Default name for the job = paste0("RNA_", output.prefix).
}
\description{
Implements a pipeline for processing RNA-Seq data, including:
1. Adapter trimming
2. Genome alignment
3. Gene counting
4. BigWig track generation
}
\examples{
# Process paired-end RNA-Seq data
cmd <- rnaseqProcessing(
  fq1 = c("sample1_R1.fq.gz", "sample1_R1_rep2.fq.gz"),
  fq2 = c("sample1_R2.fq.gz", "sample1_R2_rep2.fq.gz"),
  output.prefix = "sample1",
  genome = "hg38",
  gtf = "/data/annotations/hg38.gtf",
  GTF.attrType.extra = "gene_name",
  cores = 8
)
vl_submit(cmd, execute= FALSE)

# Then, differential analysis should be performed using ?cmd_DESeq2()

}
