% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveImage.R
\name{saveImage}
\alias{saveImage}
\title{Title}
\usage{
saveImage(
  im,
  jpeg.file,
  scale.metadata = NULL,
  scale.bar = 100,
  horizontal.lines = NULL,
  box = FALSE,
  legend = NULL,
  legend.col = "white",
  cex = 1,
  lwd = 1
)
}
\arguments{
\item{im}{}

\item{jpeg.file}{}

\item{cex}{}

\item{main}{}

\item{col}{}
}
\description{
Title
}
