% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_colors.R
\name{sc_colors}
\alias{sc_colors}
\title{Plot and cluster top marker genes from single-cell transcriptome data}
\usage{
sc_colors(clusters, unique.lvl = FALSE)
}
\arguments{
\item{clusters}{A factor vector containin clusters to color.}

\item{unique.lvl}{If set to TRUE, one color is returned per unique level inside clusters.}
}
\value{
The default colors of ggplot
}
\description{
Plot and cluster top marker genes from single-cell transcriptome data
}
