% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_computePairwiseMarkerGenes.R
\name{sc_computePariwiseMarkerGenes}
\alias{sc_computePariwiseMarkerGenes}
\title{Compute Pariwise Marker Genes}
\usage{
sc_computePariwiseMarkerGenes(
  dat,
  select.clusters,
  ctl.clusters = NULL,
  value.var = Idents(dat),
  add.fisher = FALSE
)
}
\arguments{
\item{dat}{A Seurat object (must contain an SCT assay).}

\item{select.clusters}{Character vector of cluster names to analyze. If missing, marker genes are computed for all clusters, by
comparing observations within all the selected clusters vs. outside.}

\item{value.var}{A factor vector of clusters of length ncol(dat). Default= Idents(dat).}

\item{add.fisher}{Logical; if TRUE, adds Fisher's exact test statistics for each gene/cluster.}
}
\value{
A data.table with marker gene statistics for each cluster, including average expression, log fold change, adjusted p-values, AUC, and
optionally Fisher's test results.
}
\description{
Identifies marker genes for a merged group of clusters.
}
\details{
- Uses the SCT assay's data slot for expression values.
- Only clusters in `select.clusters` are analyzed if provided.
- Fisher's test is performed on active/inactive cell counts if `add.fisher = TRUE`.
}
