% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_diffMarkers.R
\name{sc_diffMarkers}
\alias{sc_diffMarkers}
\title{A function to analyse differential marker genes between two sets of clusters}
\usage{
sc_diffMarkers(
  dat,
  select.clusters,
  exclude.clusters = NULL,
  output.prefix,
  pairwise = T,
  logFC.cutoff = 0.25,
  wilcox.padj.cutoff = 0.05,
  log2OR.cutoff = 2,
  fisher.padj.cutoff = 0.0000000001,
  pdf.output.folder = "pdf/",
  GO.output.folder = "db/GO/",
  open.pdf = TRUE,
  cleanup.cache = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dat}{A Seurat object.}

\item{select.clusters}{A character vector of the cluster names of interest.}

\item{exclude.clusters}{Specified clusters will be excluded before any analysis.}

\item{output.prefix}{Basename of output files, which will also be used as title for the plots.}

\item{pairwise}{A character vector of cluster names to be used as controls. Default= TRUE,}

\item{logFC.cutoff}{logFC cutoff (wilcoxon test). Default= 0.25.}

\item{wilcox.padj.cutoff}{p.adjust cutoff for the wilcoxon test p.value.}

\item{log2OR.cutoff}{log2OR cutoff (fisher.test). Default= 2.}

\item{fisher.padj.cutoff}{p.adjust cutoff for the fisher test p.value.}

\item{pdf.output.folder}{Folder where pdf files will be saved. Default= "pdf".}

\item{GO.output.folder}{Folder where GO.rds enrichment files will be saved. Default= "db/GO".}

\item{open.pdf}{Should the pdf file be opened At the end of the function? Default= TRUE.}

\item{cleanup.cache}{Should all enrichments be computed again, even if temporary files already exist? Default= FALSE.}

\item{verbose}{For debugging purposes. Default= FALSE.}
}
\description{
Compute pairwise comparison between clusters.
}
