% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_topMarkers.R
\name{sc_topMarkers}
\alias{sc_topMarkers}
\title{Plot and cluster top marker genes from single-cell transcriptome data}
\usage{
sc_topMarkers(
  dat,
  topN = Inf,
  perc.cutoff = 10,
  perc.diff.cutoff = 20,
  padj.wilcox.cutoff = 0.05,
  logFC.cutoff = 0.25,
  padj.fisher.cutoff = NULL,
  log2OR.cutoff = NULL,
  auc.cutoff = NULL,
  order.var = "perc.diff",
  alternative = "greater"
)
}
\arguments{
\item{dat}{A data.table with marker gene statistics. See ?sc_computeMarkerGenes.}

\item{topN}{Number of top genes per cluster. Default= Inf (all markers).}

\item{perc.cutoff}{Minimum percentage of expressing within cluster.}

\item{perc.diff.cutoff}{Minimum difference in percentage of expressing cells within cluster vs. outside cluster.}

\item{padj.wilcox.cutoff}{Adjusted Wilcoxon p-value cutoff. Default= 0.05.}

\item{logFC.cutoff}{Log fold change cutoff. Default= 0.25.}

\item{padj.fisher.cutoff}{Adjusted Fisher p-value cutoff. Default= NULL.}

\item{log2OR.cutoff}{Log2 odds ratio cutoff. Default= NULL.}

\item{auc.cutoff}{AUC cutoff. Default= NULL.}

\item{order.var}{Variable to order genes by ("perc.diff", "auc", "logFC", "log2OR").}

\item{alternative}{The direction in which filtering should be done. SHould be one of 'greater' or 'smaller'.}
}
\value{
The subset of the rds object corresponding to top marker genes.
}
\description{
Plot and cluster top marker genes from single-cell transcriptome data
}
