% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_UMAP.R
\name{sc_UMAP}
\alias{sc_UMAP}
\title{Plot UMAP from seurat object}
\usage{
sc_UMAP(
  dat,
  value.var = Idents(dat),
  show.labels = TRUE,
  main = NULL,
  col = NULL,
  NA.col = "lightgrey",
  zlim = NULL,
  legend.main = NULL,
  cex.pts = 0.4,
  transparency = 0.7,
  cex.lab = 0.7,
  cex.legend = 0.7,
  randomize = FALSE,
  layer = "SCT"
)
}
\arguments{
\item{dat}{A seurat object.}

\item{value.var}{The value to plot. Should either be a factor vector matching the number of cells in the object,
or a unique gene symbol.}

\item{show.labels}{Should the factor levels be shown on the UMAP itself? Otherwise, they will be shown in a legend.}

\item{cex.lab}{Expansion factor for labels.}

\item{cex.legend}{Expansion factor for the legend.}

\item{layer}{The layer to use. Default= "SCT".}
}
\value{
Invisibly returns a list containing the ordered heatmap matrix.
}
\description{
Function to plot rasterized umaps from a seurat object.
}
