% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCENIClite_infer_regulons.R
\name{SCENIClite_infer_regulons}
\alias{SCENIClite_infer_regulons}
\title{Title}
\usage{
SCENIClite_infer_regulons(
  seurat.object,
  N.variable.genes = NULL,
  min.act.cells = 0.025,
  TF.gene.names,
  N.genes = cores,
  N.iterations = 100L,
  cell.frac = 0.3,
  stability.cutoff = 0.6,
  method = "ELASTIC",
  output.prefix,
  output.folder = "db/candidate_regulons/",
  cores = 20,
  mem = cores * 2,
  tmp.folder = "/scratch-cbe/users/vincent.loubiere/candidate_regulons/",
  overwrite = F
)
}
\arguments{
\item{method}{}
}
\description{
Title
}
