% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starrseqProcessing.R
\name{starrseqProcessing}
\alias{starrseqProcessing}
\title{Process STARR-Seq sequencing Data}
\usage{
starrseqProcessing(
  fq1,
  fq2 = NULL,
  output.prefix,
  genome,
  genome.idx = NULL,
  max.ins = 2000,
  bed.subset = NULL,
  fq.output.folder = "db/fq/STARRSeq/",
  bam.output.folder = "db/bam/STARRSeq/",
  alignment.stats.output.folder = "db/alignment_stats/STARRSeq/",
  umi.counts.output.folder = "db/umi_counts/STARRSeq/",
  bed.output.folder = "db/bed/STARRSeq/",
  bw.output.folder = "db/bw/STARRSeq/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 8
)
}
\arguments{
\item{fq1}{A character vector of .fq (or .fq.gz) file paths.}

\item{fq2}{For paired-end data, a character vector of .fq (or .fq.gz) file paths matching fq1 files. Default= NULL.}

\item{output.prefix}{Prefix for output files.}

\item{genome}{Reference genome identifier (e.g., "mm10", "hg38").}

\item{genome.idx}{Path to the Rsubread genome index. If NULL, derived from genome. Default= NULL.}

\item{max.ins}{Bowtie2 maximum insert size for paired-end alignment. Default= 2000.}

\item{bed.subset}{Optional path to a bed file. If provided, only reads overlapping these regions on the same strand are used to generate
bigwig tracks. Default= NULL (all reads are used).}

\item{fq.output.folder}{Directory where trimmed fastq files will be saved. Default= "db/fq/STARRSeq/".}

\item{bam.output.folder}{Directory where aligned bam files will be saved. Default= "db/bam/STARRSeq/".}

\item{umi.counts.output.folder}{Directory where UMI counts will be saved. Default= "db/umi_counts/STARRSeq/".}

\item{bed.output.folder}{Directory where UMI-collapsed bed files will be saved. Default= "db/bed/STARRSeq/".}

\item{bw.output.folder}{Directory where bigwig tracks will be saved. Default= "db/bw/STARRSeq/".}

\item{Rpath}{Path to the Rscript executable. Default= "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}

\item{cores}{Number of CPU cores to use. Default= 8.}
}
\value{
A data.table with:
- `file.type`: Type of output file.
- `path`: Path to the output file.
- `cmd`: Shell command for each step in the pipeline.
- `cores`: The number of CPU cores to use.
- `job.name`: Default name for the job = paste0("STARR_", output.prefix).
}
\description{
Implements a pipeline for processing ORFeome sequencing data, including:
1. Adapter trimming
2. Genome alignment
3. UMI collapsing
4. Bigwig tracks
}
\examples{
# Process ORFeome sequencing data
cmd <- starrseqProcessing(
  fq1 = c("sample1_R1.fq.gz", "sample1_R1_rep2.fq.gz"),
  output.prefix = "sample1",
  cores = 8
)
vl_submit(cmd, execute= FALSE)

# Then, peaks should be called using ?cmd_MAGeCK_ORFeome()

}
