% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringGetDB.R
\name{stringGetDB}
\alias{stringGetDB}
\title{Build STRING db}
\usage{
stringGetDB(
  species,
  network_type = "full",
  version = "12.0",
  input_directory = tempdir()
)
}
\arguments{
\item{species}{either "Dm" or "Mm"}

\item{network_type}{The type of interactions to be used. Can be one of "full" (full functional, default) or "physical" if only physical interactions are to be considered}

\item{version}{databse version. default= "12.0"}

\item{input.directory}{The input directory where db will be saved. Default= tempdir().}
}
\description{
Build STRING db
}
