% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitCav.R
\name{submitCav}
\alias{submitCav}
\title{Title}
\usage{
submitCav(
  cmd,
  conda = "vl_conda",
  logs = "db/logs/",
  name = "vl",
  create.output.dirs = TRUE,
  overwrite = FALSE,
  execute = TRUE,
  wdir = getwd()
)
}
\arguments{
\item{cmd}{}

\item{logs}{}

\item{name}{}

\item{wdir}{}
}
\description{
Title
}
