% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractBed.R
\name{subtractBed}
\alias{subtractBed}
\title{Subtract Genomic Regions from Reference Intervals}
\usage{
subtractBed(a, b, min.width = 1L, ignore.strand = TRUE)
}
\arguments{
\item{a}{Query regions in any format compatible with ?importBed().}

\item{b}{Target regions in any format compatible with ?importBed().}

\item{min.width}{Integer specifying the minimum width required for resized regions.
Regions smaller than this are discarded. Default= 1L.}

\item{ignore.strand}{If set to FALSE, only subtracts overlapping features that are on the same strand.
If set to TRUE (default), subtracts overlapping feature(s) regardless of their strand(s).}
}
\value{
A gr data.table containing the remaining portions of a, after subtracting the regions
defined in b.
}
\description{
Resize regions in a by subtracting regions in b in order to remove overlaps between
the two. If regions in a fully overlap region(s) in b, they will therefore be removed.
}
\examples{
# Create example regions
a <- importBed(c("chr1:100-500:-"))
b <- importBed(c("chr1:200-300:+", "chr1:400-450:-", "chr1:425-475:-"))

# Basic example
subtractBed(a, b)

# Minimum width of 50bp
subtractBed(a, b, min.width = 50L)

# Only subtract regions with similar strand
subtractBed(a, b, ignore.strand= FALSE)

}
