% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_cache_file.R
\name{vl_cache_file}
\alias{vl_cache_file}
\title{Title}
\usage{
vl_cache_file(input.list)
}
\arguments{
\item{input.list}{list of elements used by the function which will be used to generate a reproducible path for the cache file.}
}
\description{
Title
}
