% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_download.R
\name{vl_dropbox_download}
\alias{vl_dropbox_download}
\title{Download a File or Folder from Dropbox}
\usage{
vl_dropbox_download(local_path, remote_path)
}
\arguments{
\item{local_path}{A character string specifying the local destination path for the downloaded file or folder.}

\item{remote_path}{A character string specifying the source path in Dropbox.}
}
\value{
No return value. Executes the Dropbox Uploader script to download the file or folder.
}
\description{
Uses the Dropbox Uploader script to download a file or folder from a specified Dropbox path to a local path.
}
\examples{
\dontrun{
vl_dropbox_download("local/file.txt", "/remote/folder/file.txt")  # Download a file
vl_dropbox_download("local/folder", "/remote/folder")  # Download a folder
}

}
