% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_upload.R
\name{vl_dropbox_upload}
\alias{vl_dropbox_upload}
\title{Upload a File to Dropbox}
\usage{
vl_dropbox_upload(local_path, remote_path)
}
\arguments{
\item{local_path}{A character string specifying the local file or folder path to upload.}

\item{remote_path}{A character string specifying the destination path in Dropbox.}
}
\value{
No return value. Executes the Dropbox Uploader script to upload the file or folder.
}
\description{
Uses the Dropbox Uploader script to upload a file or folder to a specified Dropbox path.
}
\examples{
\dontrun{
vl_dropbox_upload("local/file.txt", "/remote/folder/file.txt")  # Upload a file
vl_dropbox_upload("local/folder", "/remote/folder")  # Upload a folder
}

}
