% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_heatmap.R
\name{vl_heatmap}
\alias{vl_heatmap}
\title{Create Enhanced Heatmap Visualization}
\usage{
vl_heatmap(
  x,
  cluster.rows = TRUE,
  cluster.cols = FALSE,
  kmeans.k = NA,
  som.grid = NULL,
  order.cl = FALSE,
  breaks = NULL,
  col = NULL,
  row.annotations = NULL,
  col.annotations = NULL,
  show.legend = "right",
  legend.title = "Value",
  row.annotations.title = "Rows",
  col.annotations.title = "Columns",
  show.numbers = FALSE,
  show.rownames = TRUE,
  show.colnames = TRUE,
  main = NA,
  useRaster = FALSE,
  clustering.distance.rows = "euclidean",
  clustering.distance.cols = "euclidean",
  clustering.method = "complete",
  cutree.rows = NULL,
  cutree.cols = NULL,
  show.row.clusters = "right",
  show.col.clusters = "top",
  show.row.dendro = TRUE,
  show.col.dendro = TRUE,
  row.gap.axis = NA,
  row.gap.width = 1/10,
  col.gap.width = 1/10,
  cluster.seed = 3453,
  row.clusters.col = c("grey90", "grey40"),
  col.clusters.col = c("grey90", "grey40"),
  row.annotations.col = rainbow(9)[1:7],
  col.annotations.col = rainbow(9)[1:7],
  na.col = "lightgrey",
  legend.cex = 1,
  legend.length.cex = 1,
  numbers.cex = 0.7,
  show.grid = FALSE,
  grid.lwd = 0.25,
  plot = T,
  pdf.file = NULL,
  pdf.cell.size = 1,
  pdf.par = TRUE,
  pdf.close = TRUE
)
}
\arguments{
\item{x}{A matrix, data.table or data.frame of numeric values or factors (that will be coerced to numeric) to be displayed as a heatmap.}

\item{cluster.rows}{Logical or vector specifying row clustering:
* TRUE: perform clustering (default).
* FALSE: no clustering.
* vector: pre-defined clustering. Will be coerced to factors.
* data.table/matrix: clustering will be performed on this subset instead of x.}

\item{cluster.cols}{Similar to cluster.rows but for columns (of note, 'matrix' option is not available here).
Default= FALSE.}

\item{kmeans.k}{Integer specifying the number of k-means clusters for rows. Defaults= NA (uses hierarchical clustering).}

\item{som.grid}{Vector of two integers specifying the size of the grid to use for SOM-based clustering, using the default
parameters specific in ?somClustering. Default= NULL.}

\item{order.cl}{Should clusters be ordered from top left to bottom right? Default= FALSE.}

\item{breaks}{A numeric vector specifying the breakpoints for color mapping.
Can either be the length of col (centered breaks) or one element longer (in which case breaks will correspond to edges).
If set to NULL, defaults to the different levels for factors (centered breaks), a diverging scale
of 21 values (centered on 0) or to 20 values spanning the range of x (breaks as edges).}

\item{col}{A vector of colors corresponding to the values in breaks.
Can either be the length of breaks (centered breaks) or one element shorter (breaks used as edges).
If set to NULL, rainbow colors are used for factors, a diverging blue-white-red palette for data
spanning positive and negative values, or a sequential blue-yellow palette for data with a unique sign.}

\item{row.annotations}{A vector of length nrow(x) containing row annotations.}

\item{col.annotations}{A vector of length ncol(x) containing row annotations.}

\item{show.legend}{Should the legend be plotted?
Possible values are TRUE, FALSE, "right" (similar to TRUE) or "top". Default= "right".}

\item{legend.title}{Character string for the legend title. Default= "Value".}

\item{row.annotations.title}{Title for the row annotations legend. Default= "Rows".}

\item{col.annotations.title}{Title for the column annotations legend. Default= "Columns".}

\item{show.numbers}{Logical or matrix; if set to TRUE or a matrix is provided, displays values in cells.}

\item{show.rownames}{Logical; whether to display row names. Default= TRUE.}

\item{show.colnames}{Logical; whether to display column names. Default= TRUE.}

\item{main}{Add title to the heatmap. Default= NA.}

\item{useRaster}{logical; if TRUE a bitmap raster is used to plot the image instead of polygons. Default= FALSE.}

\item{clustering.distance.rows}{Character string specifying the distance metric for rows. Default is "euclidean".}

\item{clustering.distance.cols}{Similar to clustering.distance.rows but for columns.}

\item{clustering.method}{Character string specifying the hierarchical clustering method. Default is "complete".}

\item{cutree.rows}{Integer specifying the number of row clusters to cut the tree into. Default= NULL.}

\item{cutree.cols}{Similar to cutree.rows but for columns. Default= NULL.}

\item{show.row.clusters}{Character specifying the position of row cluster visualization: "right", "left", or FALSE.}

\item{show.col.clusters}{Character specifying the position of column cluster visualization: "top", "bottom", or FALSE.}

\item{show.row.dendro}{If rows are clustered using hclust, should the dendrogram be shown? Default= TRUE.}

\item{show.col.dendro}{If cols are clustered using hclust, should the dendrogram be shown? Default= TRUE.}

\item{row.gap.axis}{Default= NA.}

\item{row.gap.width}{The height of the gap between row clusters, expressed as a fraction of rows' height. Default= 1/10.}

\item{col.gap.width}{The width of the gap between column clusters, expressed as a fraction of columns' height. Default= 1/10.}

\item{cluster.seed}{Integer seed for reproducible clustering. Default= 3453.}

\item{row.clusters.col}{A vector of two colors for the row cluster gradient. Default is c("grey90", "grey40").}

\item{col.clusters.col}{A vector of two colors for the column cluster gradient. Default is c("grey90", "grey40").}

\item{row.annotations.col}{Row annotations colors. Default= rainbow(9)[1:7].}

\item{col.annotations.col}{Col annotations colors. Default= rainbow(9)[1:7].}

\item{na.col}{Color for NA values. Default= "ligthgrey".}

\item{legend.cex}{Numeric scaling factor for the legend. Default= 1.}

\item{legend.length.cex}{Legend length expansion factor. Default= 1.}

\item{numbers.cex}{Numeric scaling factor for the size of displayed numbers. Default= 0.7.}

\item{show.grid}{Should the matric grid be shown? Default= FALSE.}

\item{grid.lwd}{Line width used for the grid.}

\item{pdf.file}{If specified, the heatmap will be saved into the provided pdf file, with an optimized and nice layout.
Default= NULL}

\item{pdf.cell.size}{The cell size on printed pdf in inches. Default= 1.
Default= NULL}

\item{pdf.par}{If TRUE, the plotting parameters of the PDF will be automatically set to nice default values. Otherwise,
the current par() settings will be used. Default= TRUE.}

\item{pdf.close}{If TRUE, the pdf file will be closed by the end of the script. Otherwise, it stays open, allowing to
add extra elements to the plot.}
}
\value{
Invisibly returns a list with two `data.table` objects:
* row: Contains row names, ordering, and clustering (if applicable).
* col: Contains column names, ordering, and clustering (if applicable).
}
\description{
Generates a customizable heatmap with support for hierarchical clustering, k-means clustering,
dendrograms, and cluster visualization. This function provides advanced options for visualizing
matrix data with clustering and color-coded annotations.
}
\examples{
# Create example matrix
set.seed(1234)
mat <- matrix(rnorm(200), 20, 10)
colnames(mat) <- paste("Sample", 1:10, sep = "")
rownames(mat) <- paste("Gene", 1:20, sep = "")
mat <- round(mat, 2)

# Basic heatmap
vl_par()
vl_heatmap(mat)

# Hierarchical clustering
vl_heatmap(mat,
cluster.rows = T,
cluster.cols = T,
cutree.rows = 3,
cutree.cols = 3,
show.numbers = T)

}
