% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_hilbertCurves.R
\name{vl_hilbertCurves}
\alias{vl_hilbertCurves}
\title{Function to plot hilbert curves}
\usage{
vl_hilbertCurves(
  track,
  title = "pixel mode",
  genome = "dm6",
  width = 2500,
  canonical.chr = c("chrX", "chr2L", "chr2R", "chr3L", "chr3R", "chr4"),
  clip = quantile(cov, 0.9995),
  highlight = NULL
)
}
\arguments{
\item{track}{Path to the bw file to be plotted.}

\item{title}{The tile of the heatmap. Default= "pixel mode".}

\item{genome}{The genome to bin. Default= "dm6.}

\item{width}{The width of the bins used to tile the genome. Default= 2500.}

\item{canonical.chr}{The canonical chromosomes that should be plotted. Default= c("chrX", "chr2L", "chr2R", "chr3L", "chr3R", "chr4").}

\item{clip}{The values used to clip the data. Default= quantile(x, .9995).}

\item{highlight}{A set of regions that should be higlighted on the heatmap. The content of the 'name' column will be added as text.}
}
\description{
Wrapper around ?HilbertCurve package (mode= "pixel").
}
