% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_iCisEnrich.R
\name{vl_iCisEnrich}
\alias{vl_iCisEnrich}
\title{Title}
\usage{
vl_iCisEnrich(
  regions,
  max.rank = 5000,
  N.permutations = NA,
  decreasing = F,
  seed = 1,
  cleanup.cache = FALSE,
  what = "enr",
  pwm.db.file =
    "/zssd/scratch/vincent.loubiere/motifs_db/JASPAR_CORE/20260812_JASPAR2026_CORE_NON_REDUNDANT.rds",
  ranks.file =
    "/zssd/scratch/vincent.loubiere/projects/iCis_lite/Rdata/ranks/v2_core/ranks2.rds",
  tmp.dir = "/zssd/scratch/vincent.loubiere/projects/iCis_lite/db/tmp/"
)
}
\arguments{
\item{regions}{A list of regions to be overlapped with db peaks to compute enrichment.}

\item{max.rank}{Default= 5000}

\item{N.permutations}{Default= NA (no pvalues).}

\item{decreasing}{Default= FALSE.}

\item{seed}{Default= 1.}

\item{cleanup.cache}{Default= FALSE.}

\item{ranks.file}{The ranks to be used. Default to max cluster buster scores and number of cluster motifs,
saved at "/zssd/scratch/vincent.loubiere/projects/iCis_lite/Rdata/ranks/v2_core/ranks2.rds".}
}
\description{
Title
}
