% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_motifCounts.R
\name{vl_motifCounts}
\alias{vl_motifCounts}
\title{Count Motif Occurrences in Genomic Regions}
\usage{
vl_motifCounts(
  sequences,
  bed,
  pwm_log_odds,
  genome = NULL,
  bg = "genome",
  p.cutoff = 5e-05,
  cleanup.cache = FALSE,
  what = "motifCounts"
)
}
\arguments{
\item{sequences}{A named character vector of sequences to analyze. This argument takes precedence over the bed argument.}

\item{bed}{Genomic ranges in a format compatible with ?importBed, from which genomic sequences will be retrieved when sequences is set to NULL.}

\item{pwm_log_odds}{A PWMatrixList (in log2 odds ratio format) containing motifs to count.}

\item{genome}{Genome to use as background when bg = 'genome' and/or to retrieve sequences (when bed is specified).
Default= NULL.}

\item{bg}{Background model for motif detection. Options are 'genome', 'subject' (inferred from input sequences) or 'even' (0.25, 0.25, 0.25, 0.25). Default= 'genome'.}

\item{p.cutoff}{p-value cutoff for motif detection. Default= 5e-5.}

\item{cleanup.cache}{Logical. If set to TRUE, clears cached intermediate results. Default= FALSE.}

\item{what}{The values that should be returned. Should be one of 'motifCounts', 'motifMatches', 'motifScores'. Default= 'motifCounts'.}
}
\value{
A matrix of motif counts (on both strands).
}
\description{
Wrapper around ?motifmatchr::matchMotifs that counts motif occurrences in a set of genomic coordinates (or sequences) using a PWMatrixList.
Note that here the strand is not considered (motifs are called on both strands).
If you want to have the score associated to the strand, use ?vl_motifPos.
}
\examples{
# Download Dev enhancer from pe-STARR-Seq paper
tmp <- tempfile(pattern = '.xlsx')
download.file(url = 'https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41467-024-52921-2/MediaObjects/41467_2024_52921_MOESM4_ESM.xlsx',
              destfile = tmp)
dev <- readxl::read_xlsx(tmp, sheet = 2, skip = 1)
dev <- as.data.table(dev)
dev <- dev[group=='dev' & detail \%in\% c('medium', 'strong')]

# Load motifs
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
pwm <- importJASPAR(pfm.file)$pwms_log_odds

# Compute counts from coordinates
mot1 <- vl_motifCounts(bed= dev, genome= 'dm3', pwm_log_odds= pwm)

# Compute counts from sequences
mot2 <- vl_motifCounts(sequences = dev$enhancer_sequence, genome= 'dm3', pwm_log_odds= pwm)

# Both approaches should be identical
identical(mot1, mot2)

}
