% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_motifEnrichRanks.R
\name{vl_motifEnrichRanks}
\alias{vl_motifEnrichRanks}
\title{Title}
\usage{
vl_motifEnrichRanks(
  labels,
  ranks.matrix,
  max.rank = 5000,
  N.permutations = 1000,
  decreasing = F,
  seed = 1,
  cleanup.cache = FALSE,
  tmp.dir = "/zssd/scratch/vincent.loubiere/projects/iCis_lite/db/tmp/"
)
}
\arguments{
\item{labels}{A logical vector of labels matching the number of rows in ranks.matrix.}

\item{ranks.matrix}{A numeric matrix of motif ranks (one motif per column).}

\item{max.rank}{The maximum rank up to which cumulative recovery should be computed. Default= 5000.}

\item{N.permutations}{The number of permutations used for empirical p-values. 
If set to NA, p-values are not computed. Default= 1,000.}

\item{decreasing}{Should the provided ranks be used in decreasing order? Default = FALSE.}

\item{cleanup.cache}{Should the cache be cleaned? Default= FALSE.}
}
\description{
Title
}
