% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_pie.R
\name{vl_pie}
\alias{vl_pie}
\title{Clean pie chart}
\usage{
vl_pie(x, col = NULL, labels = "r", remove.zeros = TRUE, ...)
}
\arguments{
\item{x}{Variable to be summarized}

\item{col}{colors to be used}

\item{labels}{Type of labels to be added. "r"= number, "p"= percentage}

\item{...}{extra arguments passed to pie(x, ...)}
}
\value{
Plots a pie chart
}
\description{
Clean pie chart
}
\examples{
vl_pie(rep(1:3, 1:3))
}
