% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_plot.R
\name{vl_plot}
\alias{vl_plot}
\title{Scatterplot function}
\usage{
vl_plot(
  x,
  y = NULL,
  pch = 19,
  cex = 0.8,
  col = adjustcolor("grey", 0.7),
  xaxt = "s",
  frame = F,
  axes = T,
  ...
)
}
\arguments{
\item{y}{Default= NULL}

\item{pch}{Default= 19}

\item{cex}{Default= .8}

\item{col}{Default= adjustcolor("grey", .7)}

\item{xaxt}{Default= "s"}

\item{frame}{Default= F}

\item{axes}{Default= T}
}
\description{
A wrapper around ?plot with nicer default settings.
}
\examples{
vl_par()
vl_plot(1:3)
}
