% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_PPV.R
\name{vl_PPV}
\alias{vl_PPV}
\title{Positive Predicted Value}
\usage{
vl_PPV(
  predicted,
  label,
  Nleft = 100,
  plot = FALSE,
  xlim = NULL,
  ylim = NULL,
  xlab = "Prediction score",
  ylab = "Positive pred. value (\%)",
  lty.1 = 1,
  lty.2 = 3,
  col = "black",
  show.max = TRUE,
  show.max.value = show.max,
  cex.max = 0.7,
  pos.max = 2,
  offset.max = 0.5,
  show.pred.cutoff = T,
  add = FALSE,
  return.value.at = NULL,
  ...
)
}
\arguments{
\item{predicted}{Predicted values from the model (ranging from 0 to 1).}

\item{label}{A vector of logical labels (or that can be coerced to logical).}

\item{Nleft}{Minimum number of remaining positive sequences (prediction>=cutoff) before computing max PPV.
Default= 100.}

\item{plot}{Should the PPV be plotted?}

\item{xlim}{x limits for plotting. Default= NULL.}

\item{ylim}{y limits for plotting. Default= NULL.}

\item{xlab}{x label. Default= "Prediction score".}

\item{ylab}{y label. Default= "Positive pred. value (percentage)".}

\item{lty.1}{Line type before Nleft. Default= 1.}

\item{lty.2}{Line type after Nleft. Default= 3.}

\item{col}{Color.}

\item{cex.max}{cex for the max PPV text label.}

\item{pos.max}{pos for the PPV text label Default= 3.}

\item{offset.max}{Offset for plotting the PPV text label. Default= .5.}

\item{add}{When plot is set to TRUE, should only the lines/label be added to an existing plot?}

\item{return.value.at}{If specified, return the PPV value at the last predicted value <= return.value.at.}

\item{...}{Extra arguments to be passed to lines}
}
\value{
PPV at a given threshold and/or PPV curve with max PPV with a minimum number of observations.
}
\description{
Given a set of predicted values and a vector of logical labels, computes the Positive Predictive Values for
increasing predicted thresholds.
}
\examples{
# Create synthetic data
predicted <- c(1,1,0.8,0.4,.1,0)
label <- c(T,T,T,F,F,F)

# Plot PPV
vl_PPV(predicted, label, Nleft= 3, plot= T) # 100\% TPR with at least 3 obs
vl_PPV(predicted, label, Nleft= 3, return.value.at= .4) # 75\% TPR with pred >= .4 (and at least 3 obs)  
vl_PPV(predicted, label, Nleft= 4, plot= T) # 100\% TPR with at least 4 obs

}
