% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_prAUC.R
\name{vl_prAUC}
\alias{vl_prAUC}
\title{PR AUC}
\usage{
vl_prAUC(
  rank,
  label,
  compute.NES = TRUE,
  N.iter = 100L,
  decreasing = TRUE,
  plot = FALSE,
  main = "PR",
  xlab = "Recall (TPR)",
  ylab = "Precision (PPV)",
  type = "l",
  add = FALSE,
  ...
)
}
\arguments{
\item{rank}{A numeric variable used for ranking (higher = more positive when decreasing=TRUE).}

\item{label}{A vector of logical labels (or that can be coerced to logical).}

\item{compute.NES}{Should a Normalized Enrichment Score be computed by label permutation? Default= TRUE.}

\item{N.iter}{Number of iterations used to compute the NES. Default= 100L.}

\item{decreasing}{Should the sort order for rank be decreasing? If FALSE, ranks are inverted for PRROC. Default= TRUE.}

\item{plot}{Should the PR curve be plotted? Default= FALSE.}

\item{main}{Main title. Default= "PR".}

\item{xlab}{xlab. Default= "Recall (TPR)".}

\item{ylab}{ylab. Default= "Precision (PPV)".}

\item{type}{The type of plot. Default= "l".}

\item{add}{Should the PR curve be added to an existing plot?}

\item{...}{Extra arguments passed to plot (when plot=TRUE and add=FALSE) or lines (when add=TRUE).}

\item{xlim}{Plot xlim. Default= c(0, 1).}

\item{ylim}{Plot ylim. Default= c(0, 1).}
}
\value{
PR AUC (and NES if compute.NES=TRUE)
}
\description{
Computes Precision-Recall AUC (PR-AUC) and optionally plots the PR curve.
}
