% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_motifAUC.R, R/vl_rankAUC.R
\name{vl_rankAUC}
\alias{vl_rankAUC}
\title{Title}
\usage{
vl_rankAUC(
  rankings,
  labels,
  rank.name = "rank",
  label.name = "motif",
  cleanup.cache = FALSE
)

vl_rankAUC(
  rankings,
  labels,
  rank.name = "rank",
  label.name = "motif",
  cleanup.cache = FALSE
)
}
\arguments{
\item{labels}{}
}
\description{
Title

Title
}
