% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_rocAUC_grid.R
\name{vl_rocAUC_grid}
\alias{vl_rocAUC_grid}
\title{Paralle ROC AUC}
\usage{
vl_rocAUC_grid(
  ranks,
  labels,
  compute.NES = TRUE,
  N.iter = 100L,
  decreasing = TRUE,
  ranks.name = "rank",
  labels.name = "label",
  cleanup.cache = FALSE
)
}
\arguments{
\item{ranks}{A matrix (or.data.table) of variables (one per column) used for ranking (decreasing= TRUE by default. See decreasing argument).}

\item{labels}{A matrix (or data.table) of logical labels (one per column) for which ROC AUC will be computed .}

\item{compute.NES}{Should the Normalized Enrichment Score be computed (Niter= 1000)? Default= TRUE.}

\item{N.iter}{Number of iterations used to compute the NES. Default= 100L}

\item{decreasing}{Should the sort order for ranks be increasing or decreasing? Default= TRUE.}

\item{ranks.name}{The name to give to the ranks variables. Default= "rank".}

\item{labels.name}{The name to give to the labels variables. Default= "label".}

\item{cleanup.cache}{Should the temporary files be overwritten?}
}
\value{
ROC AUC values and NES for all ranks/labels combinations.
}
\description{
A wrapper around vl_rocAUC that computes ROC AUC and NES for many rank/label combinations.
}
\examples{
# Create synthetic data
rank <- 10:1
ranks <- matrix(c(rank, rev(rank)), ncol= 2)
label <- rep(c(T, F), each= 5)
labels <- matrix(c(label, rev(label)), ncol= 2)

# Compute ROC AUC
vl_rocAUC_grid(ranks= ranks, labels= labels)

}
