% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_scancel.R
\name{vl_scancel}
\alias{vl_scancel}
\title{Cancel SLURM Jobs Except Specified Ones}
\usage{
vl_scancel(user = "vincent.loubiere", ignore.name = c("[RStudio", "jupyter_"))
}
\arguments{
\item{user}{Username for which jobs should be canceled. Defaults to `vincent.loubiere`.}

\item{ignore.name}{Names of the jobs that should not be cancelled. Defaults to c('[RStudio', 'jupyter_').}
}
\value{
No return value. Executes the `scancel` command to cancel jobs.
}
\description{
Cancels all SLURM jobs for a user, except those with names matching the specified patterns.
}
\examples{
\dontrun{
vl_scancel()  # Cancel jobs for the default user
vl_scancel(user = "another.user", ignore.name = c("my_job"))
}

}
