% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_seqLogoMat.R
\name{vl_seqLogoMat}
\alias{vl_seqLogoMat}
\title{Plot logo from a numeric matrix}
\usage{
vl_seqLogoMat(
  mat,
  name = NULL,
  xlab = "Index",
  ylab = "Frequency",
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{mat}{A matrix (or a TFBSTools PFMatrix, PWMatrix or ICMatrix) with rows 'A', 'C', 'G', 'T'.}

\item{name}{The name of the motif which will be added as a title.}

\item{xlab}{xlab. Default= "Index".}

\item{ylab}{ylab. Default= "Frequency".}

\item{xlim}{xlim for the plot. Default= NULL.}

\item{ylim}{ylim for the plot. Default= NULL.}

\item{...}{Extra parameters for the plot function.}
}
\description{
Generic function that plots a logo out of an arbitrary numeric matrix.
}
\examples{
# Select br PPM
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
PPM <- importJASPAR(pfm.file, pseudocount = 0)$PPM[["br"]]

# Plot
vl_par()
vl_seqLogoMat(PPM, name= "br")

}
