% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_squeue.R
\name{vl_squeue}
\alias{vl_squeue}
\title{Squeue to the System}
\usage{
vl_squeue(user = "vincent.loubiere")
}
\arguments{
\item{user}{A character string specifying the username. Defaults to `"vincent.loubiere"`.}
}
\value{
A data table containing the output of the `squeue` command.
}
\description{
This function retrieves the SLURM job queue for a specified user by running the `squeue` command
on the SLURM server and reads the output into a data table.
}
\examples{
\dontrun{
# Retrieve the SLURM job queue for the default user
vl_squeue()

# Retrieve the SLURM job queue for a different user
vl_squeue(user = "another.user")
}

}
