% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_submit.R
\name{vl_submit}
\alias{vl_submit}
\title{Submit Commands to a Server with Resource Management}
\usage{
vl_submit(
  cmd,
  cores = 8,
  mem = 32,
  time = "08:00:00",
  job.name = "myJob",
  logs = "db/logs/",
  modules = c("build-env/f2022", "mamba/24.3.0-0"),
  conda = "/groups/stark/conda_envs/.conda/envs/vlite_pipelines/",
  overwrite = FALSE,
  execute = TRUE,
  create.output.dirs = TRUE
)
}
\arguments{
\item{cmd}{A data.table containing the commands to execute, which will all be submitted as a single job. Intended columns are:
- `file.type`: Type of output file.
- `path`: Path to the output file.
- `cmd`: Shell command to execute.
- `cores`: (Optional) The number of CPU cores to allocate to the job (only the value from the first row will be used).
- `mem`: (Optional) Memory allocation in GB (only the value from the first row will be used).
- `time`: (Optional) Maximum runtime for the job (HH:MM:SS, only the value from the first row will be used).
- `job.name`: (Optional) The name of the job, only the value from the first row will be used.
- `logs`: (Optional) Directory for log files, only the value from the first row will be used
- `modules`: (Optional) A vector or list of modules to be loaded before executing the command.}

\item{cores}{Number of CPU cores to allocate. If the cmd object contains a 'cores' column, it will override this argument.
Default= 8.}

\item{mem}{Memory allocation in GB. If the cmd object contains a 'mem' column, it will override this argument. Default= 32.}

\item{time}{Maximum runtime for the job (HH:MM:SS). If the cmd object contains a 'time' column, it will override this argument.
Default= '08:00:00'.}

\item{job.name}{Name of the job. If the cmd object contains a 'job.name' column, it will override this argument.
Default= "myJob".}

\item{logs}{Directory for log files. If the cmd object contains a 'logs' column, it will override this argument.
Default = 'db/logs'.}

\item{modules}{A character vector of modules to be loaded before executing the command. If the cmd object contains a 'modules'
column, it will override this argument. Default= c("build-env/f2022", "mamba/24.3.0-0").
By default, the conda module used by the pipelines is being loaded.}

\item{conda}{Path to a conda environment. If specified, it will be activated after loading the modules.
If the cmd object contains a 'conda' column, it will override this argument.
Default= "/groups/stark/conda_envs/.conda/envs/vlite_pipelines/".
By default, the conda environment required to run vlite_pipelines is being activated.}

\item{overwrite}{If set to TRUE, overwrites existing output files. Default= FALSE.}

\item{execute}{If set to FALSE, the command is returned and is not submitted to the cluster. Default= TRUE.}

\item{create.output.dirs}{Should missing output directories be created before executing the command? Default= TRUE.}
}
\description{
Used in vlite pipeline to submit shell commands to the server Handles job submission, resource allocation, and optional directory
creation for output files.
}
\examples{
# Submit a command to the server
cmd <- data.table(
  file.type = "bam",
  path = "output.bam",
  cmd = "echo 'Hello world'",
  job.name= "test"
)
vl_submit(cmd)

# Return the constructed command without submitting
vl_submit(cmd, execute = FALSE)

}
