% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano Plot}
\usage{
volcanoPlot(
  log2FC,
  padj,
  labels = rep("", length(log2FC)),
  label.col = "black",
  main = NULL,
  log2FC.cutoff = 1,
  padj.cutoff = 0.05
)
}
\arguments{
\item{log2FC}{}

\item{padj}{}

\item{labels}{}

\item{label.col}{}

\item{main}{}

\item{log2FC.cutoff}{}

\item{padj.cutoff}{}
}
\value{
A volcano plot where specific points can be labelled
}
\description{
A wrapper around ggrepelScatterplot to plot a nice volcano Plot where specific
points can be labelled using ggrepel.
}
