% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addImageLegend.R
\name{addImageLegend}
\alias{addImageLegend}
\title{Title}
\usage{
addImageLegend(channels = "DAPI", col = "blue", title = "Image")
}
\arguments{
\item{channels}{}

\item{col}{}

\item{title}{}
}
\description{
Title
}
