% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRsq.R
\name{addRsq}
\alias{addRsq}
\title{Plots Rsq or PCC coeff}
\usage{
addRsq(rsq, pos = "topleft", adjusted = F, bty = "n", digits = 2, ...)
}
\arguments{
\item{rsq}{Rsquare value. If numeric, will be rounded, while characters will be printed as is}

\item{pos}{Position. Default= "topleft" (see x argument in ?legend for more detail).}

\item{adjusted}{Is the Rsq adjusted? Default= FALSE.}

\item{bty}{Box around the legend? Default= FALSE.}

\item{digits}{Rounding digits. Default= 2.}

\item{...}{}
}
\value{
Add Rsquare as a legend on aplot
}
\description{
Plots Rsq or PCC coeff
}
