% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSeqLogo.R
\name{addSeqLogo}
\alias{addSeqLogo}
\title{Plot logo from a numeric matrix}
\usage{
addSeqLogo(
  mat,
  x = 0,
  y = 0,
  width = 1,
  pos = 4,
  cex.width = 1,
  cex.height = 1,
  ...
)
}
\arguments{
\item{mat}{A matrix (or a TFBSTools PFMatrix, PWMatrix or ICMatrix) with rows 'A', 'C', 'G', 'T'.}

\item{...}{Extra parameters for the plot function.}

\item{name}{The name of the motif which will be added as a title.}

\item{xlab}{xlab. Default= "Index".}

\item{ylab}{ylab. Default= "Frequency".}

\item{xlim}{xlim for the plot. Default= NULL.}

\item{ylim}{ylim for the plot. Default= NULL.}
}
\description{
Generic function that plots a logo out of an arbitrary numeric matrix.
}
\examples{
# Select br PPM
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
PPM <- importJASPAR(pfm.file, pseudocount = 0)$PPM[["br"]]

# Plot
vl_par()
addSeqLogo(PPM, name= "br")

}
