% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSeqLogo.R
\name{addSeqLogo}
\alias{addSeqLogo}
\title{Add ICM logos to an existing plot}
\usage{
addSeqLogo(
  pwm,
  x,
  y,
  pos = 2,
  cex.width = 1,
  cex.height = 1,
  min.content = 0.05
)
}
\arguments{
\item{pwm}{List of percentage PWM matrices which will be converted to ICM before plotting.}

\item{x}{x positions (see pos argument).}

\item{y}{y positions (centered).}

\item{pos}{Either 2 (left) of 4 (right)}

\item{cex.width}{width expansion factor.}

\item{cex.height}{height expansion factor.}

\item{min.content}{Flanks with a smaller summed content are not plotted.}
}
\description{
plot seqlogo from pwm matrix
}
\examples{
#' # Retrieve br motif
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
pwm <- importJASPAR(pfm.file)$pwms_log_odds[["br"]]

# Plot
plot(0, 1, type= "n")
addSeqLogo(pwm, x = -1, y= 1, cex.height = 3, cex.width = 2, pos= 2)
# Higher min.content cutoff
addSeqLogo(pwm, x = -1, y= 0.8, cex.height = 3, cex.width = 2, pos= 2, min.content= 1)

}
