% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSeqLogo.R
\name{addSeqLogo}
\alias{addSeqLogo}
\title{Add ICM logos to an existing plot}
\usage{
addSeqLogo(
  ICM,
  sel = NULL,
  x = NULL,
  y = NULL,
  pos = 2,
  cex.width = 1,
  cex.height = 1
)
}
\arguments{
\item{ICM}{A list of ICMs to be plotted.}

\item{sel}{Indices or names of the ICM that should be plotted. Should match the length of x and y.
Default= NULL.}

\item{x}{x positions (see pos argument).}

\item{y}{y positions (centered).}

\item{pos}{Either 2 (left) of 4 (right)}

\item{cex.width}{width expansion factor.}

\item{cex.height}{height expansion factor.}

\item{min.content}{Flanks with a smaller summed content are not plotted.}
}
\description{
plot seqlogo from pwm matrix
}
\examples{
# Retrieve br motif
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
ICM <- importJASPAR(pfm.file)$ICM

# Plot
plot(0, 1, type= "n")
addSeqLogo(
ICM, 
x = -1,
y= seq(0.8, 1.2, .2),
sel= c("br", "dl", "Eip74EF"),
cex.height = 1.5,
cex.width = 1, 
pos= 2
)

}
