% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSeqLogoEnrPlot.R
\name{addSeqLogoEnrPlot}
\alias{addSeqLogoEnrPlot}
\title{Plot motif logos on the left of an enrichment plot}
\usage{
addSeqLogoEnrPlot(obj, ICM, cex.width = 1, cex.height = 1)
}
\arguments{
\item{obj}{Either a data.table containing 'name' and 'y' columns, similar to the $rows data.table
returned by vl_heatmap or the `vl_enr_cl` object returned by the ?plot.vl_enr_cl function.
The 'name' values should match the motif names present in the provided ICMatrixList (see next arugment),
and y corresponds to the y plotting positions of each row.}

\item{cex.width}{A width expansion factor applied to letter heights.}

\item{cex.height}{A height expansion factor applied to letter heights.}

\item{ICMatrixList}{An ICMatrixList object containing the motifs to add.}
}
\description{
Specialized funciton used to plot motif logs on the left of a heatmap/balloon plot.
}
\examples{
# Select br PPM
sel <- plot(
  enr,
  top.enrich = 5,
  order = "log2OR",
  padj.cutoff = 1e-3,
  col= c("pink", "red"),
  cex = .4,
  main = "ChIP motif enrich."
)
addSeqLogoEnrPlot(
  obj = sel,
  ICM = mot$ICM
)

}
