% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloons_plot.R
\name{balloons_plot}
\alias{balloons_plot}
\title{Ballons plot}
\usage{
balloons_plot(
  var1,
  var2,
  color.breaks = NULL,
  col = c("#440154FF", "#472D7BFF", "#3B528BFF", "#2C728EFF", "#21908CFF", "#27AD81FF",
    "#5DC863FF", "#AADC32FF", "#FDE725FF"),
  cex = 1,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  gap.axis = 0,
  var1.legend.title = NULL,
  var2.legend.title = NULL,
  legend.left.pos = NULL,
  legend.size.breaks = NULL
)
}
\arguments{
\item{var1}{A matrix of numeric values controling balloons'sizes.}

\item{var2}{A matrix of numeric values controling balloons'colors.}

\item{color.breaks}{Breaks used for coloring (see var 2).}

\item{col}{Vector of colors used for coloring}

\item{cex}{Expansion factor for balloons. default= 1.}

\item{main}{Title. Default= NA}

\item{gap.axis}{See ?par(). Default= 0}

\item{xaxt}{See ?par(). Default= "s".}

\item{yaxt}{See ?par(). Default= "s".}
}
\value{
Balloon plot
}
\description{
Starting from two matrices (dot size and colors, respectively), generates a balloon plot
}
\examples{
var1 <- matrix(-3:5, ncol= 3)
var2 <- matrix(-3:5, ncol= 3)
balloons_plot(var1= var1,
                 var2= var2)

}
