% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloons_key.R
\name{balloonskey}
\alias{balloonskey}
\title{Plots balloons legend}
\usage{
balloonskey(
  breaks,
  labels,
  main = NA,
  cex,
  left = par("usr")[2],
  top = par("usr")[4],
  ticks.cex = 0.6,
  main.cex = 1
)
}
\arguments{
\item{breaks}{Numeric values specifying size breaks.}

\item{labels}{Labels corresponding to each size.}

\item{main}{Title}

\item{cex}{cex parameter affecting balloons sizes.}

\item{left}{left position.}

\item{top}{top position.}

\item{ticks.cex}{cex expansion factor for ticks labels}

\item{main.cex}{cex expansion factor for legend title}

\item{height}{height}
}
\value{
Plots heatkey
}
\description{
Plots balloons legend
}
