% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwCoverage.R
\name{bwCoverage}
\alias{bwCoverage}
\title{bw coverage}
\usage{
bwCoverage(bed, bw)
}
\arguments{
\item{bed}{BED file input compatible with '?importBed', representing the genomic ranges for which mean bw signal will be returned.}

\item{bw}{Path to a target bw file.}
}
\description{
Compute mean signal from a bigwig file at a set of coordinates. Regions with no overlap will return NA.
}
\examples{
bwCoverage(bed= bed, bw= bw)

}
