% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwScreenshot.R
\name{bwScreenshot}
\alias{bwScreenshot}
\title{Create Genomic Screenshots with BigWig Tracks and Gene Annotations}
\usage{
bwScreenshot(
  bed,
  tracks = character(),
  track.names = function(x) gsub("(.*)[.].*$", "\\\\1", x),
  col = colorRampPalette(c("grey60", "grey10"))(length(tracks)),
  border.col = NA,
  border.lwd = 0.5,
  bw.max = NA,
  bw.min = NA,
  bw.n.breaks = 100,
  genome,
  ngenes = 1,
  sel.gene.symbols = NULL,
  cex.symbol = 0.7,
  offset.symbol = 0.25,
  col.ns = "cornflowerblue",
  col.ps = "tomato",
  exon.border = NA,
  gtf,
  gtf.transcript = "mRNA",
  gtf.exon = "exon",
  gtf.symbol = "gene_symbol",
  gtf.transcript.id = "transcript_id",
  region.width = 100,
  space.width = 30,
  bw.height = 100,
  bed.height = 20,
  space.height = 2,
  gene.height = 4,
  gene.space.height = 7,
  add = FALSE
)
}
\arguments{
\item{bed}{Input genomic ranges, in any format compatible with ?importBed().}

\item{tracks}{Character vector of paths to visualization tracks:
\itemize{
  \item BigWig (.bw) files for continuous signal tracks
  \item BED files for discrete regions (any format compatible with importBed)
}}

\item{track.names}{Function or character vector for track labels. Default extracts names
from file basenames.}

\item{col}{Colors for tracks. Default is a grey gradient from grey60 to grey10.}

\item{bw.max}{Numeric. Maximum signal value for BigWig tracks. Values above this
will be clipped. Default (NA) uses track maximum.}

\item{bw.n.breaks}{The number of breaks to which the signal will be simplified, to avoid polygon vector with too many points.
If set to NA, the signal is not simplified and is plotted as is.}

\item{genome}{Character. Genome assembly (e.g., "mm10", "dm6") for gene annotations.
If specified, nearby genes will be displayed.}

\item{ngenes}{Integer. Number of nearest genes to display. Default = 1.}

\item{sel.gene.symbols}{If specified, only selected gene symbols will be plotted. Default= NULL.}

\item{gtf}{Character path to a custom GTF file for custom gene annotations (alternative to the 'genome' parameter).}

\item{gtf.transcript}{Character. Name of transcript features in the GTF file (default = "mRNA").}

\item{gtf.exon}{Character. Name of exon features in the GTF file (default = "exon").}

\item{gtf.symbol}{Character. Name of the attribute containing gene symbols in the GTF file
(default = "gene_symbol").}

\item{gtf.transcript.id}{Character. Name of the attribute containing transcript IDs in the GTF file
(default = "transcript_id").}
}
\value{
A screenshot of the genomic tracks/regions.

A genomic screenshot.
}
\description{
Creates publication-quality genomic screenshots displaying BigWig signal tracks,
BED regions, and optional gene annotations for specified genomic regions. Multiple
regions can be displayed side by side for comparison.
}
\section{Visual Parameters}{

**Track Appearance:**
\itemize{
  \item `bw.height`: Height of BigWig tracks (default = 100)
  \item `bed.height`: Height of BED tracks (default = 20)
  \item `space.height`: Vertical space between tracks (default = 2)
}

**Region Layout:**
\itemize{
  \item `region.width`: Width of each genomic region (default = 100)
  \item `space.width`: Horizontal space between regions (default = 30)
}

**Gene Annotations:**
\itemize{
  \item `gene.height`: Height of gene tracks (default = 4)
  \item `gene.space.height`: Space between genes (default = 7)
  \item `cex.symbol`: Size of gene symbols (default = 0.7)
  \item `offset.symbol`: Vertical offset for gene symbols (default = 0.25)
  \item `col.ns`: Color for negative strand genes (default = "cornflowerblue")
  \item `col.ps`: Color for positive strand genes (default = "tomato")
  \item `exon.border`: Color for exon rectangle borders (default = NA, meaning no border)
}
}

\examples{
# Plotting parameters for a nice layout
par(mai = c(0.9, 0.9, 0.9, 0.9),
    las = 1,
    tcl = -0.1,
    mgp = c(1.5, 0.35, 0),
    cex = 1,
    cex.lab = 9/12,
    cex.axis = 7/12,
    bty = "n",
    lend = 2)

# Simple example with two regions and 4 tracks
bwScreenshot(
  bed= c("chr3R:30760926-30794202", "chr3R:30765926-30789202"),
  tracks= c("/groups/stark/vloubiere/projects/epigenetic_cancer/db/bw/ATAC/ATAC_PH18_merge.bw",
            "/groups/stark/vloubiere/projects/epigenetic_cancer/db/peaks/ATAC/ATAC_PH18_conf_peaks.narrowPeak",
            "/groups/stark/vloubiere/projects/epigenetic_cancer/db/bw/ATAC/ATAC_PHD11_merge.bw",
            "/groups/stark/vloubiere/projects/epigenetic_cancer/db/peaks/ATAC/ATAC_PHD11_conf_peaks.narrowPeak"),
  genome= "dm6"
)

# Genes only
bwScreenshot(
  bed= c("chr3R:30760926-30794202", "chr3R:30765926-30789202"),
  genome= "dm6"
)

}
