% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipBed.R
\name{clipBed}
\alias{clipBed}
\title{Clip Genomic Regions to Defined Limits}
\usage{
clipBed(a, b, ignore.strand = TRUE)
}
\arguments{
\item{a}{Query regions in any format compatible with ?importBed.}

\item{b}{Target regions in any format compatible with ?importBed.}

\item{ignore.strand}{If set to FALSE and strand column is provided, only clips boundaries
that are on the same strand. If set to TRUE (default), clips boundaries regardless of their strand.}
}
\value{
A data.table containing the remaining portions of a, after clipping them using the boundaries
defined in b.
}
\description{
A wrapper around ?GenomicRanges::findOverlaps that clips the regions in a that extend beyond
the regions in b.
}
\examples{
# Create example regions
a <- importBed(c("chr1:100-500:+"))
b <- importBed(c("chr1:200-300:+", "chr1:400-450:-", "chr1:425-475:-"))

# Basic example
clipBed(a, b)[]

# Strand-specific
clipBed(a, b, ignore.strand = FALSE)[]

}
