% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipBed.R
\name{clipBed}
\alias{clipBed}
\title{Clip Genomic Regions to Defined Limits}
\usage{
clipBed(a, b, min.width = 1L, ignore.strand = TRUE)
}
\arguments{
\item{a}{Query regions in any format compatible with ?importBed().}

\item{b}{Target regions in any format compatible with ?importBed().}

\item{min.width}{Integer specifying the minimum width required for clipped regions.
Regions smaller than this are discarded. Default= 1L.}

\item{ignore.strand}{If set to FALSE, only clips at boundaries that are on the same strand.
If set to TRUE (default), clips boundaries regardless of their strands.}
}
\value{
A gr data.table containing the remaining portions of a, after clipping them using the boundaries
defined in b.
}
\description{
Clips genomic regions in a to the boundaries defined by regions in b.
This function is used for restricting regions in a to valid genomic
boundaries, such as chromosome limits or other predefined regions.
}
\examples{
# Create example regions
a <- importBed(c("chr1:100-500:+"))
b <- importBed(c("chr1:200-300:+", "chr1:400-450:-", "chr1:425-475:-"))

# Basic example
clipBed(a, b)

# Require minimum width of 100bp
clipBed(a, b, min.width= 100)

# Strand-specific
clipBed(a, b, ignore.strand = FALSE)

}
