% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestBed.R
\name{closestBed}
\alias{closestBed}
\title{Find Closest Genomic Features Between Two Sets of Regions}
\usage{
closestBed(a, b, k = NA, ignore.strand = TRUE)
}
\arguments{
\item{a}{Query regions in any format compatible with ?importBed.}

\item{b}{Target regions in any format compatible with ?importBed.}

\item{k}{If specified, the k nearest neighbors are computed. If set to NA (default), only the closest
regions are returned, including ties.}

\item{ignore.strand}{If set to FALSE, only closest features that are on the same strand are reported.
If set to TRUE (default), closest features are reported on both strands. See details.}
}
\value{
A data.table with columns:
\itemize{
  \item idx.a: line index of the regions in a.
  \item idx.b: line index of the closest region(s) in b.
  \item dist: distance between closest regions (see details).
}
}
\description{
A wrapper around `GenomicRanges::nearest()` and `GenomicRanges::nearestKNeighbors()` that computes,
for each genomic range in a, the nearest genomic ranges in b and the stranded distance between them.
}
\details{
**Distance Calculation**:
- Features that have no nearest neighbor (e.g., chromosome missing in b) are not returned
- For overlapping features, a distance of 0 is returned.
- For non-overlapping features, the genomic distance between their closest boundaries are returned:
  * Negative distances indicate upstream features.
  * Positive distances indicate downstream features.
  * Unstranded (*) features are treated as `+`.
}
\examples{
# Create example regions
a <- importBed(c("chr2R:100-200:+", "chr2R:100-200:-"))
b <- importBed(c("chr2R:300-400:+", "chr2R:300-400:-", "chr2R:3000-3100:+"))

# Find single closest features
closestBed(a, b)[]

Only consider features that are on the same strand
closestBed(a, b, ignore.strand = F)[]

# Return all features at the second closest distance (including ties):
closestBed(a, b, k= 3)[]
closestBed(a, b, k= 3, ignore.strand= FALSE)[]

}
