% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_alignBowtie.R
\name{cmd_alignBowtie}
\alias{cmd_alignBowtie}
\title{Generate Commands for Sequence Alignment Using Bowtie2}
\usage{
cmd_alignBowtie(
  fq1,
  output.prefix,
  genome,
  genome.idx = NULL,
  max.mismatch = 2,
  bam.output.folder = "db/bam/",
  alignment.stats.output.folder = "db/alignment_stats/",
  cores = 8
)
}
\arguments{
\item{cores}{}
}
\description{
Generate Commands for Sequence Alignment Using Bowtie2
}
