% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_alignCellRangerArc.R
\name{cmd_alignCellRangerArc}
\alias{cmd_alignCellRangerArc}
\title{Generate Cell Ranger ARC Alignment Command for Multiome Data}
\usage{
cmd_alignCellRangerArc(
  gex.dir,
  gex.sample.name,
  atac.dir,
  atac.sample.name,
  index,
  min.gex.count = NULL,
  min.atac.count = NULL,
  output.prefix,
  output.folder = "db/multiome_10X/",
  cores = 10,
  mem = 64
)
}
\arguments{
\item{gex.dir}{Path to the directory containing GEX FASTQ files.}

\item{gex.sample.name}{The sample name prefix for GEX FASTQs.
cellranger-arc will only search for filenames matching sampleName_S*_L.*.}

\item{atac.dir}{Path to the directory containing ATAC FASTQ files.}

\item{atac.sample.name}{The sample name prefix for ATAC FASTQs.
cellranger-arc will only search for filenames matching sampleName_S*_L.*.}

\item{index}{Path to the Cell Ranger ARC reference.}

\item{min.gex.count}{Cell caller override: define the minimum number of GEX UMI counts for a cell barcode. Note: this option must be specified in conjunction with min.atac.count.
With min.atac.count=X and min.gex.count=Y a barcode is defined as a cell if it contains at least X ATAC counts AND at least Y.}

\item{min.atac.count}{See min.gex.count argument above.}

\item{output.prefix}{Prefix for output files and the run ID.}

\item{output.folder}{Directory where output files will be saved. Default= 'db/multiome_10X/'.}

\item{cores}{Number of CPU cores to use. Default= 10.}

\item{mem}{Memory in GB to allocate. Default= 64.}
}
\description{
Constructs a shell command to run 10x Genomics Cell Ranger ARC for paired
multiome (GEX and ATAC) quantification. This function is intended for
real data and does not create dummy files.
}
