% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_bamToBigwig.R
\name{cmd_bamToBigwig}
\alias{cmd_bamToBigwig}
\title{Title}
\usage{
cmd_bamToBigwig(
  bam,
  layout,
  output.prefix = NULL,
  extend.PE.fragments = FALSE,
  extsize = 0,
  bw.output.folder = "db/bw/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
