% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_bedgraphToBigwig.R
\name{cmd_bedgraphToBigwig}
\alias{cmd_bedgraphToBigwig}
\title{Convert BedGraph to BigWig Format}
\usage{
cmd_bedgraphToBigwig(
  bdg,
  output.prefix,
  genome,
  scaling.factor = 1L,
  bw.output.folder = "db/bw/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{bdg}{Path to the input BedGraph file. Only a single file is allowed.}

\item{output.prefix}{Prefix for the output BigWig file. If missing, will be constructed from the
basename of the bdg file.}

\item{genome}{A BSgenome ("mm10", "dm3"...).}

\item{bw.output.folder}{Directory for the BigWig file. Default= "db/bw/".}

\item{Rpath}{Path to the Rscript binary. Default= "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}
}
\value{
A `data.table` with:
- `file.type`: Output file label ("bw").
- `path`: Path to the BigWig file.
- `cmd`: Shell command to run the BedGraph to BigWig conversion.
- `job.name`: Default name for the job = "bdgToBw".
}
\description{
Creates shell commands to convert a BedGraph file to BigWig format using a genome sizes file.
}
\examples{
# Convert a BedGraph file to BigWig format
cmd <- cmd_bedgraphToBigwig(
  bdg = "/data/peaks/sample.bdg",
  output.prefix = "sample",
  genome = "mm10"
)
vl_submit(cmd, execute= FALSE)

}
