% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_bedToBigwig.R
\name{cmd_bedToBigwig}
\alias{cmd_bedToBigwig}
\title{Convert BED to BigWig Format}
\usage{
cmd_bedToBigwig(
  bed,
  genome,
  bed.subset = NULL,
  output.prefix = NULL,
  bw.output.folder = "db/bw/",
  Rpath = "Rscript"
)
}
\arguments{
\item{bed}{Path to the input BED file. Only a single BED file is allowed.}

\item{genome}{A BS genome name use to retrieve seqnames.}

\item{bed.subset}{Optional path to a bed file. If provided, only reads overlapping these regions on the same
strand are used. Default= NULL (no filtering).}

\item{output.prefix}{Prefix for the output BigWig file. If not provided, it is derived from the input bed filename.}

\item{bw.output.folder}{Directory for the BigWig file. Default= "db/bw/".}

\item{Rpath}{Path to the Rscript binary. Default= "Rscript".}
}
\value{
A data.table with:
- `file.type`: Output file label ("bw").
- `path`: Path to the BigWig file.
- `cmd`: Shell command to run the BED to BigWig conversion.
- `job.name`: Default name for the job = "bedToBw".
}
\description{
Creates shell commands to convert a BED file to a CPM-normalized bigwig file.
}
\examples{
# Convert a BED file to BigWig format
cmd <- cmd_bedToBigwig(
  bed = "/data/bed/sample.bed",
  genome = "mm11"
)
vl_submit(cmd, execute= FALSE)

}
