% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_confidentPeaks.R
\name{cmd_confidentPeaks}
\alias{cmd_confidentPeaks}
\title{Title}
\usage{
cmd_confidentPeaks(
  replicates_peaks_files,
  merge_peaks_file,
  output_prefix,
  conf_peaks_output_folder = "db/peaks/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
