% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_confidentPeaks.R
\name{cmd_confidentPeaks}
\alias{cmd_confidentPeaks}
\title{Generate Commands for Identifying Confident Peaks}
\usage{
cmd_confidentPeaks(
  replicates.peaks.files,
  merge.peaks.file,
  output.prefix,
  conf.peaks.output.folder = "db/peaks/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{replicates.peaks.files}{Vector of paths to replicate peak files in .narrowPeak or .broadPeak format.}

\item{merge.peaks.file}{Path to the merged peak file in .narrowPeak or .broadPeak format.}

\item{output.prefix}{Prefix for the output file.}

\item{conf.peaks.output.folder}{Directory for the confident peaks file. Default= "db/peaks/".}

\item{Rpath}{Path to the Rscript binary. Default= "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}
}
\value{
A data.table with:
- `file.type`: Output file label ("confident.peaks").
- `path`: Path to the confident peaks file.
- `cmd`: Shell command to run the confident peaks pipeline.
- `job.name`: Default name for the job = "confPeaks".
}
\description{
Creates shell commands to identify confident peaks across replicates using merged peak files.
Outputs a file containing confident peaks.
}
\examples{
# Identify confident peaks across replicates
cmd <- cmd_confidentPeaks(
  replicates.peaks.files = c("/data/peaks/rep1.narrowPeak", "/data/peaks/rep2.narrowPeak"),
  merge.peaks.file = "/data/peaks/merged.narrowPeak",
  output.prefix = "sample1"
)
vl_submit(cmd, execute= FALSE)

}
