% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_countBCreads.R
\name{cmd_countBCreads}
\alias{cmd_countBCreads}
\title{Generate Commands for Counting Barcode Reads}
\usage{
cmd_countBCreads(
  bam,
  output.prefix = NULL,
  counts.output.folder = "db/counts/",
  Rpath = "Rscript"
)
}
\arguments{
\item{bam}{Path to the input BAM file. Only a single BAM file is allowed.}

\item{output.prefix}{Prefix for the output file. If not provided, it is derived from the input BAM filename.}

\item{counts.output.folder}{Directory for the counts file. Default= "db/counts/".}

\item{Rpath}{Path to the Rscript binary. Default= "Rscript".}
}
\value{
A data.table with:
- `file.type`: Output file label ("counts.BC").
- `path`: Path to the output counts file.
- `cmd`: Shell command to run the barcode counting pipeline.
- `job.name`: Default name for the job = "countBCreads".
}
\description{
Creates shell commands to count barcode reads from a BAM file using a predefined barcode dictionary.
Outputs a counts file.
}
\examples{
# Count barcodes in a BAM file
cmd <- cmd_countBCreads(
  bam = "/data/bam/sample.bam"
)
vl_submit(cmd, execute= FALSE)

}
