% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_countRsubread.R
\name{cmd_countRsubread}
\alias{cmd_countRsubread}
\title{Title}
\usage{
cmd_countRsubread(
  bam,
  layout,
  output.prefix = NULL,
  genome,
  gtf = NULL,
  GTF.attrType.extra = NULL,
  counts.stats.output.folder = "db/count_stats/",
  counts.output.folder = "db/counts/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{Rpath}{}
}
\description{
Title
}
