% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_download.R
\name{cmd_download}
\alias{cmd_download}
\title{Download files}
\usage{
cmd_download(
  url,
  output.name = basename(url),
  submit = TRUE,
  output.folder,
  output.file.type = NA,
  gzip = FALSE
)
}
\arguments{
\item{url}{URL to download}

\item{output.name}{The output file name (default= keep url basename).}

\item{submit}{If set to TRUE, the command will directly be submitted. If set to FALSE,
a 'cmd' object compatible with ?vl_submit will be returned. Default= TRUE.}

\item{output.folder}{Path where the output file will be saved.}

\item{output.file.type}{The type of output file (fq...). Default= NA.}

\item{gzip}{Should the output file be gzipped? Default= FALSE}
}
\value{
A `data.table` with:
- `file.type`: Output file labels.
- `path`: Paths to trimmed FASTQ files.
- `cmd`: Shell command to run Trim Galore.
- `job.name`: Default name for the job = "download".
}
\description{
Download files
}
\examples{
vl_submit("url/to/data", execute= FALSE)

}
