% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_downloadSRA.R
\name{cmd_downloadSRA}
\alias{cmd_downloadSRA}
\title{Download and Process SRA Files}
\usage{
cmd_downloadSRA(
  SRR,
  layout,
  output.prefix = SRR,
  fq.output.folder = "db/fq/",
  gzip = TRUE,
  tmp.folder = paste0(fq.output.folder, "/tmp/"),
  threads = 4,
  mem = 2048
)
}
\arguments{
\item{SRR}{SRA accession number}

\item{layout}{Either "SINGLE" or "PAIRED" for read layout}

\item{output.prefix}{Output file prefix (default= SRR number).}

\item{fq.output.folder}{Directory for output FASTQ files.}

\item{gzip}{Logical, whether to gzip the output files.}

\item{tmp.folder}{Directory for temporary files.}

\item{threads}{Number of threads (default: 4)}

\item{mem}{Memory in MB (default: 2048)

#' @return A `data.table` with:
- `file.type`: Output file labels. "fq1" or "fq2".
- `path`: Paths to the fastq files.
- `cmd`: Shell command to download and process SRA file.
- `job.name`: Default name for the job = "dlSRA".}
}
\description{
Download and Process SRA Files
}
\examples{
cmd <- cmd_downloadSRA("SRR123456",
                       layout="PAIRED",
                       fq.output.folder="./fastq",
                       tmp.folder="./tmp")
# Submit command
vl_submit(cmd, execute= FALSE)

}
