% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_mergeBigwig.R
\name{cmd_mergeBigwig}
\alias{cmd_mergeBigwig}
\title{Convert Multiple BigWig Files to a Merged BigWig Format}
\usage{
cmd_mergeBigwig(
  bw,
  output.prefix,
  bw.output.folder = "db/bw/",
  tmp.bdg.folder = "db/bw/tmp/",
  genome,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  bigWigMergePath =
    "/software/2020/software/kent_tools/20190507-linux.x86_64/bin/bigWigMerge"
)
}
\arguments{
\item{bw}{Character vector of paths to input BigWig files.}

\item{output.prefix}{Prefix for the output merged BigWig file. If missing, will be constructed from the
basename of the first provided bw path.}

\item{bw.output.folder}{Directory for the output files. Default is "db/bw/".}

\item{tmp.bdg.folder}{Folder to save temporary .bedgraph files. Default is "db/bw/tmp/".}

\item{genome}{A BSgenome ("mm10", "dm3"...).}

\item{Rpath}{Path to the Rscript binary. Default is "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}

\item{bigWigMergePath}{Path to the tool executable.
Default= "/software/2020/software/kent_tools/20190507-linux.x86_64/bin/bigWigMerge"}
}
\value{
A `data.table` with:
- `file.type`: Output file label ("bw").
- `path`: Path to the merged BigWig file.
- `cmd`: Shell command to run the BigWig merge and BedGraph to BigWig conversion.
- `job.name`: Default name for the job = "bwMerge".
}
\description{
Creates shell commands to merge multiple BigWig files and convert the merged BedGraph to BigWig format using a
genome sizes file.
}
\examples{
# Merge BigWig files and convert to a single BigWig
cmd <- cmd_bedgraphToBigwig(
  bw = c("/data/bw/sample_1.bw", "/data/bw/sample_2.bw"),
  output.prefix = "sample",
  genome = "mm10"
)
vl_submit(cmd, execute = FALSE)

}
