% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_peakCalling.R
\name{cmd_peakCalling}
\alias{cmd_peakCalling}
\title{Generate Commands for Peak Calling Using MACS2}
\usage{
cmd_peakCalling(
  bam,
  bam.input = NULL,
  layout,
  output.prefix,
  keep.dup = 1,
  compute.model = FALSE,
  extsize = 200,
  shift = 0,
  broad = FALSE,
  genome.macs2,
  genome,
  peaks.output.folder = "db/peaks/",
  bw.output.folder = "db/bw/",
  Rpath = "Rscript"
)
}
\arguments{
\item{bam}{A character vector specifying path(s) to treatment BAM file(s).}

\item{bam.input}{A character vector specifying path(s) to input BAM file(s). Default= NULL.}

\item{layout}{Sequencing layout, either "SINGLE" or "PAIRED".}

\item{output.prefix}{Prefix for the output files.}

\item{keep.dup}{Number of duplicates to keep or "all". Default= 1L.}

\item{compute.model}{Should MACS2 peak model be computed? Of note, for paired-end reads,
modelling will be skipped. Default= FALSE.}

\item{extsize}{If compute.model is set to TRUE and single-end reads are provided, integer value
specifying by how many bases the reads should be extended.
Typically 200 for ChIP-Seq, DNAse, CutNrun; 75 for ATAC-Seq. Default= 200L.}

\item{shift}{If compute.model is set to TRUE and single-end reads are provided, integer value
specifying by how many bases the reads should be extended.
Typically 0 for ChIP-Seq, DNAse & CutNrun; -35 for ATAC-Seq, -100 for DNAse. Default= 0L.}

\item{broad}{Logical. Whether to call broad peaks. Default= FALSE.}

\item{genome.macs2}{Genome size parameter for MACS2 (e.g., "dm, "mm", "hs").}

\item{genome}{A BSgenome name (e.g. "mm10", "dm6"...) that will be used to generate the bigwig track.}

\item{peaks.output.folder}{Directory where MACS2 output files should be saved. Default= "db/peaks/".}

\item{bw.output.folder}{Directory where bigwig files should be saved. Default= "db/bw/".}

\item{Rpath}{Path to the Rscript binary. Default= "Rscript".}
}
\value{
A data.table with:
- `file.type`: Output file labels (e.g: "peaks", "bedgraph").
- `path`: Paths to the output files.
- `cmd`: Shell command to run the peak calling pipeline.
- `job.name`: Default name for the job = "MACS2".
}
\description{
Creates shell commands to perform peak calling on BAM files using MACS2.
Supports single-sample and merged peak calling, with optional input controls,
and outputs peak files and bedGraph files.
}
\examples{
# Basic peak calling without input controls
cmd <- cmd_peakCalling(
  bam = "/data/bam/rep1.bam",
  layout = "PAIRED",
  genome.macs2 = "mm",
  output.prefix = "sample1"
)
vl_submit(cmd, execute= FALSE)

# Peak calling with input controls
cmd <- cmd_peakCalling(
  bam = "/data/bam/rep1.bam",
  bam.input = "/data/bam/input1.bam",
  layout = "PAIRED",
  genome.macs2 = "mm",
  output.prefix = "sample1"
)
vl_submit(cmd, execute= FALSE)

}
