% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_peakCalling.R
\name{cmd_peakCalling}
\alias{cmd_peakCalling}
\title{Generate Commands for Peak Calling Using MACS2}
\usage{
cmd_peakCalling(
  bam,
  bam.input = NULL,
  layout,
  output.prefix,
  keep.dup = 1,
  extsize = 200,
  shift = 0,
  genome.macs2,
  peaks.output.folder = "db/peaks/",
  broad = FALSE
)
}
\arguments{
\item{bam}{A character vector specifying path(s) to treatment BAM file(s).}

\item{bam.input}{A character vector specifying path(s) to input BAM file(s). Default= NULL.}

\item{layout}{Sequencing layout, either "SINGLE" or "PAIRED".}

\item{output.prefix}{Prefix for the output files.}

\item{keep.dup}{Number of duplicates to keep or "all". Default= 1L.}

\item{extsize}{Integer value by which reads should be extended.
Typically 200 for ChIP-Seq, DNAse, CutNrun; 75 for ATAC-Seq. Default= 200L.}

\item{shift}{Integer value by which reads will be shifted.
Typically 0 for ChIP-Seq, DNAse & CutNrun; -35 for ATAC-Seq, -100 for DNAse. Default= 0L.}

\item{genome.macs2}{Genome size parameter for MACS2 (e.g., "mm", "hs").}

\item{peaks.output.folder}{Directory for peak files. Default= "db/peaks/".}

\item{broad}{Logical. Whether to call broad peaks. Default= FALSE.}
}
\value{
A data.table with:
- `file.type`: Output file labels (e.g: "peaks", "bedgraph").
- `path`: Paths to the output files.
- `cmd`: Shell command to run the peak calling pipeline.
- `job.name`: Default name for the job = "MACS2".
}
\description{
Creates shell commands to perform peak calling on BAM files using MACS2.
Supports single-sample and merged peak calling, with optional input controls,
and outputs peak files and bedGraph files.
}
\examples{
# Basic peak calling without input controls
cmd <- cmd_peakCalling(
  bam = "/data/bam/rep1.bam",
  layout = "PAIRED",
  genome.macs2 = "mm",
  output.prefix = "sample1"
)
vl_submit(cmd, execute= FALSE)

# Peak calling with input controls
cmd <- cmd_peakCalling(
  bam = "/data/bam/rep1.bam",
  bam.input = "/data/bam/input1.bam",
  layout = "PAIRED",
  genome.macs2 = "mm",
  output.prefix = "sample1"
)
vl_submit(cmd, execute= FALSE)

}
