% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_peakCallingFromBw.R
\name{cmd_peakCallingFromBw}
\alias{cmd_peakCallingFromBw}
\title{Call Peaks from bigWig Files}
\usage{
cmd_peakCallingFromBw(
  experiment.bw.file,
  input.bw.file = NULL,
  bed,
  genome,
  output.prefix,
  output.folder = "db/peaks/",
  zscore.cutoff = 1.64,
  local.enr.cutoff = 1.5,
  local.fdr.cutoff = 0.05,
  input.enr.cutoff = 1.5,
  input.fdr.cutoff = 0.05,
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"
)
}
\arguments{
\item{experiment.bw.file}{Path to the experiment bigWig (.bw) file. Must be unique.}

\item{input.bw.file}{Optional path to input/control bigWig (.bw) file. Must be unique or NULL.}

\item{bed}{Genomic ranges for which peaks will be called, in any format compatible with ?importBed.
If genome is specified (see below), this argument will be ignored.}

\item{genome}{A BSgenome name specifying the genome for which peaks should be called. For custom regions,
see the bed argument.}

\item{output.prefix}{Prefix for output files.}

\item{output.folder}{Output directory for result files. Default: "db/peaks/".}

\item{zscore.cutoff}{Z-score cutoff for putative peaks. Default: 1.64.}

\item{local.enr.cutoff}{Local background enrichment cutoff. Default: 1.5.}

\item{local.fdr.cutoff}{Local background FDR cutoff. Default: 0.05.}

\item{input.enr.cutoff}{Input enrichment cutoff. Default: 1.5.}

\item{input.fdr.cutoff}{Input FDR cutoff. Default: 0.05.}

\item{Rpath}{Path to the Rscript binary. Default: "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript".}
}
\value{
A data.table with:
- `file.type`: Output file label ("peaks", "stats").
- `path`: Path to the output file.
- `cmd`: Shell command to run the peak calling pipeline.
}
\description{
Generates a shell command to call peaks from a ChIP-seq experiment bigWig file, with optional input control, over specified genomic regions.
Outputs a .narrowPeak file and a statistics .txt file.
}
\examples{
cmd <- cmd_peakCallingFromBw(
  experiment.bw.file = "/data/exp.bw",
  input.bw.file = "/data/input.bw",
  genome.regions = "/data/regions.bed",
  output.prefix = "sample1"
)
vl_submit(cmd, execute = FALSE)

}
