% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_trimIlluminaAdaptors.R
\name{cmd_trimIlluminaAdaptors}
\alias{cmd_trimIlluminaAdaptors}
\title{Generate Trim Galore Commands for FASTQ Files}
\usage{
cmd_trimIlluminaAdaptors(fq1, fq2 = NULL, fq.output.folder = "db/fq/")
}
\arguments{
\item{fq1}{A character vector of .fq (or .fq.gz) file paths.}

\item{fq2}{For paired-end data, a character vector of .fq (or .fq.gz) file paths matching fq1 files. Default= NULL.}

\item{fq.output.folder}{Directory for trimmed FASTQ files. Default= "db/fq/".}
}
\value{
A `data.table` with:
- `file.type`: Output file labels.
- `path`: Paths to trimmed FASTQ files.
- `cmd`: Shell command to run Trim Galore.
- `job.name`: Default name for the job = "trimGal".
}
\description{
Creates shell commands to trim Illumina adapters from gzipped FASTQ files using Trim Galore.
Supports single-end and paired-end sequencing data.
}
\examples{
# Single-end
cmd <- cmd_trimIlluminaAdaptors(fq1 = "sample_R1.fq.gz")
vl_submit(cmd, execute= FALSE)

# Paired-end
cmd <- cmd_trimIlluminaAdaptors(fq1 = "sample_R1.fq.gz", fq2 = "sample_R2.fq.gz")
vl_submit(cmd, execute= FALSE)

}
