% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_velocyto.R
\name{cmd_velocyto}
\alias{cmd_velocyto}
\title{Generate Cell Ranger Alignment Command}
\usage{
cmd_velocyto(cellranger.output.folder, gtf, cores = 8, mem = 64)
}
\arguments{
\item{mem}{}
}
\description{
Constructs a shell command to run velocyto on a Cell Ranger output, producing a .loom file with spliced/unspliced counts from the sample’s BAM and a provided GTF, with resources and modules recorded.
Velocyto quantifies spliced and unspliced mRNA from single-cell RNA‑seq to infer RNA velocity—i.e., the short‑term future transcriptional state of cells.
}
