% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covBam.R
\name{covBam}
\alias{covBam}
\title{Title}
\usage{
covBam(
  bed,
  bam,
  ignore.strand = TRUE,
  output.prefix = NULL,
  output.folder = "db/counts/",
  bedtools_path = "/software/2020/software/bedtools/2.27.1-foss-2018b/bin/bedtools"
)
}
\arguments{
\item{bed}{Regions in any format compatible with ?importBed.}

\item{bam}{Path to a bam file from which overlapping reads will be counted.}

\item{ignore.strand}{If set to FALSE, only the reads features that are on the same strand
as the bed regions will be counted. If set to TRUE (default), reads are counted irrespective
of their strand.}

\item{output.prefix}{Prefix for the output file name. If specified, a 'cmd' object compatible with
?vl_submit will be returned. If set to NULL (default), the command will be submitted to the system
using fread(cmd= cmd).}

\item{output.folder}{Directory for the output count files. Default: "db/counts/".}

\item{bedtools_path}{Path to the bedtools bin folder.
Default= "/software/2020/software/bedtools/2.27.1-foss-2018b/bin/bedtools".}
}
\value{
A vector of integer counts.
}
\description{
A wrapper around 'bedtools coverage' that computes, for each genomic range in a
bed file, the number of overlapping reads in a bam file.
}
\examples{
covBam(bed= "path/to/bed/file.bed",
bam = "path/to/bam/file.bam")

}
