% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichBed.R
\name{enrichBed}
\alias{enrichBed}
\title{Assess overlap between peaks and regions}
\usage{
enrichBed(
  regions,
  peaks,
  ctl.peaks,
  pseudocount = 0.5,
  maxgap = -1L,
  minoverlap = 0L,
  ignore.strand = TRUE
)
}
\arguments{
\item{regions}{Regions for which overlaps with peaks have to be assessed, in any format compatible with ?importBed.}

\item{peaks}{Peaks regions in any format compatible with ?importBed.}

\item{ctl.peaks}{Control peak regions with no overlaps with peaks, in any format compatible with ?importBed.}

\item{pseudocount}{The pseudocount to use to avoid 0s in fisher test estimates. Default= 0.5.
Note that the p-value will not be affected.}

\item{maxgap}{A single integer specifying the maximum gap allowed between 2 ranges for them to
be considered as overlapping. Default= -1.}

\item{minoverlap}{A single integer specifying the minimum overlap between 2 ranges for them to
be considered as overlapping. Default= 0.}

\item{ignore.strand}{If set to FALSE, only overlaps between regions that are on the same strand will be counted.
If set to TRUE (default), overlaps on both strands are counted.}
}
\value{
A list with elements:
\itemize{
  \item tab: contingency table.
  \item p.value: fisher test p.value.
  \item estimate: the fisher test odd ratio, based on the specified pseudocount.
}
}
\description{
Test whether a set of regions significantly overlaps a set of peaks.
}
\examples{
peaks <- data.table(seqnames= "chr2R",
                    start= seq(1000, 10000, 1000))
peaks[, end:= start+100]
ctl.peaks <- resizeBed(peaks, "end", -100, 200)
enrichBed(peaks = peaks[1:5],
          ctl.peaks = ctl.peaks,
          regions = peaks)

}
