% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSgenomeSize.R
\name{getBSgenomeSize}
\alias{getBSgenomeSize}
\title{Get BS chromosome sizes}
\usage{
getBSgenomeSize(genome)
}
\arguments{
\item{genome}{BSgenome object to use: "dm6", "mm10"...}
}
\value{
data.table containing chromosomes start and end coordinates
}
\description{
Get BS chromosome sizes
}
\examples{
getBSchromSizes("mm10")

}
