% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBSsequence.R
\name{getBSsequence}
\alias{getBSsequence}
\title{Get genomic sequence}
\usage{
getBSsequence(bed, genome)
}
\arguments{
\item{bed}{Input regions in any format compatible with ?importBed.}

\item{genome}{BSgenome name: "dm6", "mm10"...}
}
\value{
Character vector of sequences
}
\description{
Returns the sequences of a bed.
}
\examples{
# Sample random regions
test <- randomRegionsBSgenome("dm6", widths= rep(100, 3))
getBSsequence(test, "dm6")

}
