% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataImage.R
\name{getMetadataImage}
\alias{getMetadataImage}
\title{Title}
\usage{
getMetadataImage(paths)
}
\arguments{
\item{paths}{A vector of file paths for which metadata should be extracted}
}
\description{
Title
}
