% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatkey.R
\name{heatkey}
\alias{heatkey}
\title{Add Color Key (Legend) to Heatmap}
\usage{
heatkey(
  breaks,
  col,
  labels = NULL,
  position = "right",
  adj.x = 0,
  adj.y = 0,
  thickness = 0.75,
  length = 4,
  main = NA,
  cex = 1
)
}
\arguments{
\item{breaks}{A numeric vector specifying the breakpoints (edges) for color mapping.
The number of colors in col must be exactly one less than the number of breaks.}

\item{col}{A vector of colors, where each color corresponds to the interval between consecutive breakpoints in breaks.
Must have length length(breaks) - 1.}

\item{labels}{Labels to be added to the color key. By default, elegant numeric ticks spanning the values in breaks are computed.
If provided as a numeric vector, labels are placed at the corresponding values.
If provided as a character vector of length equal to col, labels are placed at the center of each color interval.}

\item{position}{Position of the color key. Either "right" (vertical legend) or "top" (horizontal legend). Default= "right".}

\item{adj.x}{If specified, x plotting positions are adjusted by adj.x*line.width. Default= 0.}

\item{adj.y}{If specified, y plotting positions are adjusted by adj.y*line.height. Default= 0.}

\item{thickness}{Numeric width of the color key in lines. Default= 0.75.}

\item{length}{Numeric height (if vertical) or width (if horizontal) of the color key in lines. Default= 4.}

\item{main}{Character string for the title of the color key. Default= NA.}

\item{cex}{Numeric scaling factor for the size of the color key text. Default= 1.}
}
\value{
Adds a color key to the current plot, at one line of distance from the plot border. No return value.
}
\description{
Plots a color key (legend) for heatmaps. This function is designed to work alongside heatmap visualizations
and should be called after the main heatmap plot.
}
\details{
The function creates a color key (legend) that maps colors to value intervals in your heatmap.
The number of colors must be exactly one less than the number of breaks.
Labels can be numeric (placed at the corresponding value) or character (placed at the center of each color interval).
}
\examples{
# Create example matrix
set.seed(1234)
test <- matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] <- test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] <- test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] <- test[15:20, seq(2, 10, 2)] + 4
colnames(test) <- paste("Test", 1:10, sep = "")
rownames(test) <- paste("Gene", 1:20, sep = "")

# Basic heatmap with default color key
vl_par()

}
