% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatkey.R
\name{heatkey}
\alias{heatkey}
\title{Add Color Key (Legend) to Heatmap}
\usage{
heatkey(
  breaks,
  col,
  labels = NULL,
  position = "right",
  adj.x = 0,
  adj.y = 0,
  thickness = 0.75,
  length = 4,
  main = NA,
  cex = 1
)
}
\arguments{
\item{breaks}{A numeric vector specifying the breakpoints for color mapping.}

\item{col}{A vector of colors, which can either be the length of breaks (in which case breaks will be centered)
or have one value less than breaks (in which case breaks will be used as edges).}

\item{labels}{Labels to be added to the color key. By default, elegant numeric ticks spanning the values in breaks are computed.
If numeric labels are provided, they are plotted at the corresponding positions.
If a character vector (or factor levels) of length equal to col is provided, labels are centered at breaks' values}

\item{position}{Position of the color key. Either "right" (vertical legend) or "top" (horizontal legend). Default= "right".}

\item{adj.x}{If specified, x plotting positions are adjusted by adj.x*line.width. Default= 0.}

\item{adj.y}{If specified, y plotting positions are adjusted by adj.y*line.height. Default= 0.}

\item{thickness}{Numeric width of the color key in lines. Default= 0.75.}

\item{length}{Numeric height (if vertical) or width (if horizontal) of the color key in lines. Default= 4.}

\item{main}{Character string for the title of the color key. Default= NA.}

\item{cex}{Numeric scaling factor for the size of the color key text. Default= 1.}
}
\value{
Adds a color key to the current plot, at one line of distance from the plot border. No return value.
}
\description{
Plots a color key (legend) for heatmaps. This function is designed to work alongside heatmap visualizations
and should be called after the main heatmap plot.
}
\details{
The function creates a color key (legend) that maps colors to value intervals in your heatmap.
The number of colors must be exactly one less than the number of breaks.
Labels can be numeric (placed at the corresponding value) or character (placed at the center of each color interval).
}
\examples{
# Example with the image function
image(matrix(1:3, ncol= 1), col= c("blue", "white", "red"))

# By default, in this function, breaks are used as edges:
heatkey(0:3, c("blue", "white", "red"), main = "test")

# But in other situations (for example to emphasize 0 as the middle value), one might prefer to use centered breaks:
breaks <- c(-2.5, -1.5, -0.5, 0.5, 1.5, 2.5)
col <- c("blue", "cornflowerblue", "white", "tomato", "red")
image(matrix(-2:2, ncol= 1), breaks= breaks, col= col)
heatkey(breaks= breaks, col= col, pos= "top", main= "Values")

# Character also be used for character labels
heatkey(breaks= 1:3, labels= c("A", "B", "C"), col= c("green", "yellow", "purple"), main= "Test")

}
