% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatkey.R
\name{heatkey}
\alias{heatkey}
\title{Add Color Key (Legend) to Heatmap}
\usage{
heatkey(
  breaks,
  col,
  left = par("usr")[2] + diff(grconvertX(c(0, 1), "line", "user")),
  top = par("usr")[4],
  width = 0.75,
  height = 4,
  cex = 1,
  main = NA
)
}
\arguments{
\item{breaks}{A numeric vector specifying the breakpoints for color mapping.}

\item{col}{A vector of colors corresponding to the values in `breaks`.}

\item{left}{Numeric position for the left edge of the color key. Default is calculated based on the current plot region.}

\item{top}{Numeric position for the top edge of the color key. Default is the top of the current plot region.}

\item{width}{Numeric width of the color key in lines. Default is `0.75`.}

\item{height}{Numeric height of the color key in lines. Default is `4`.}

\item{cex}{Numeric scaling factor for the size of the color key text. Default is `1`.}

\item{main}{Character string for the title of the color key. Default is `NA` (no title).}
}
\value{
Adds a color key to the current plot. No return value.
}
\description{
Plots a color key (legend) for heatmaps. This function is designed to work alongside heatmap visualizations
and should be called after the main heatmap plot.
}
\details{
The function creates a vertical color key that maps colors to values in your heatmap. The `breaks` vector defines
the midpoints of the ranges for each color in `col`. The color key is customizable in terms of position, size, and appearance.
}
\examples{
# Create example matrix
set.seed(1234)
test <- matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] <- test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] <- test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] <- test[15:20, seq(2, 10, 2)] + 4
colnames(test) <- paste("Test", 1:10, sep = "")
rownames(test) <- paste("Gene", 1:20, sep = "")

# Basic heatmap with default color key
vl_par()
vl_image(test)
heatkey(col = heat.colors(100), breaks = seq(-3, 3, length.out = 100))

}
\seealso{
\code{\link{vl_image}} for creating the main heatmap.
}
