% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatkey.R
\name{heatkey}
\alias{heatkey}
\title{Add Color Key (Legend) to Heatmap}
\usage{
heatkey(
  breaks,
  col,
  position = "right",
  adj.x = 0,
  adj.y = 0,
  thickness = 0.75,
  length = 4,
  main = NA,
  cex = 1
)
}
\arguments{
\item{breaks}{A numeric vector (or factors) specifying the breakpoints for color mapping.}

\item{col}{A vector of colors corresponding to the values in `breaks`.}

\item{adj.x}{If specified, x plotting positions are adjusted by adj.x * line.width Default= 0.}

\item{adj.y}{If specified, y plotting positions are adjusted by adj.y * line height. Default= 0.}

\item{thickness}{Numeric width of the color key in lines. Default= 0.75.}

\item{length}{Numeric height of the color key in lines. Default= 4.}

\item{main}{Character string for the title of the color key. Default= NA.}

\item{cex}{Numeric scaling factor for the size of the color key text. Default= 1.}

\item{pos}{Position of the heatkey. Can either be "right" or "top". Default= "right".}
}
\value{
Adds a color key to the current plot, at one line of distance from the plot border. No return value.
}
\description{
Plots a color key (legend) for heatmaps. This function is designed to work alongside heatmap visualizations
and should be called after the main heatmap plot.
}
\details{
The function creates a vertical color key that maps colors to values in your heatmap. The breaks vector defines
the midpoints of the ranges for each color in col. The color key is customizable in terms of position, size, and appearance.
}
\examples{
# Create example matrix
set.seed(1234)
test <- matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] <- test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] <- test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] <- test[15:20, seq(2, 10, 2)] + 4
colnames(test) <- paste("Test", 1:10, sep = "")
rownames(test) <- paste("Gene", 1:20, sep = "")

# Basic heatmap with default color key
vl_par()
vl_image(test)
heatkey(col = heat.colors(100), breaks = seq(-3, 3, length.out = 100))

}
\seealso{
\code{\link{vl_image}} for creating the main heatmap.
}
