% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBamRaw.R
\name{importBamRaw}
\alias{importBamRaw}
\title{Import raw bam using samtools and fread}
\usage{
importBamRaw(
  file,
  extra_arg,
  headN,
  samtools_exe_path = "/software/2020/software/samtools/1.9-foss-2018b/bin/samtools view"
)
}
\arguments{
\item{file}{bam file path.}

\item{extra_arg}{Extra arg to be passed to samtools view.}

\item{headN}{Number of starting lines to import.}

\item{samtools_exe_path}{Path to samtools executable.}
}
\value{
Imported bam file
}
\description{
Uses samtools to import a bam file.
}
\examples{
importBamRaw("path/to/bam/file.bam")
}
