% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBamRaw.R
\name{importBamRaw}
\alias{importBamRaw}
\title{Import Raw BAM File Using Samtools and fread}
\usage{
importBamRaw(
  file,
  extra_arg,
  headN,
  samtools_exe_path = "/software/2020/software/samtools/1.9-foss-2018b/bin/samtools"
)
}
\arguments{
\item{file}{A character string specifying the path to the BAM file.}

\item{extra_arg}{Optional extra arguments that will be passed to `samtools view`.}

\item{headN}{An integer specifying the number of starting lines to import. By default, all reads will be processed.}

\item{samtools_exe_path}{A character string specifying the path to the samtools executable.
Default= "/software/2020/software/samtools/1.9-foss-2018b/bin/samtools".}
}
\value{
A data table containing the imported BAM file.
}
\description{
Uses samtools to import a BAM file and reads the output into R using fread.
}
\examples{
# Import a BAM file
importBamRaw("path/to/bam/file.bam")

# Import with additional arguments
importBamRaw("path/to/bam/file.bam", extra_arg = "-f 4", headN = 100)

}
