% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{importBed}
\alias{importBed}
\title{Import Genomic Ranges in BED Format}
\usage{
importBed(bed, col.names = NULL)
}
\arguments{
\item{bed}{Input genomic ranges in one of the following formats:
\itemize{
  \item Character vector of genomic coordinates ("chr:start-end[:strand]").
  \item File path to .bed, .narrowPeak, or .broadPeak file.
  \item GRanges object.
  \item data.frame or data.table with required columns.
}}

\item{col.names}{Can be used to overwrite default column names.}
}
\description{
Imports genomic coordinates from various input formats into a standardized data.table format.
Supported input types: GRanges objects, character strings ("chr2L:30000-31000:-"), .bed, .narrowPeak
or .broadPeak file paths (see details).
}
\details{
Character coordinates should follow the syntax "chr:start-end[:strand]".
File specifications can be found at https://genome.ucsc.edu/FAQ/FAQformat.html.
GRanges objects are directly converted to data.table and metadata columns are preserved.
Required columns for data.frames/data.tables are 'seqnames', 'start', 'end' and 'strand'.
If missing, end and strand will be set to 'start' and '*', respectively.
}
\examples{
# Import from character coordinates
importBed(c("chr2L:1000-2000:+", "chr3L:1000-2000:-", "chr3L:4000-6000:*"))[]

# From GRanges
importBed(GRanges("chr2L", IRanges(c(1000, 5000), c(2000, 6000))))[]

# From bed file
bed_file <- tempfile(fileext = ".bed")
exportBed(gr, bed_file)
importBed(bed_file)[]

# From narrowpeak file
peaks <- data.table(
  seqnames = "chr2L",
  start = 1000,
  end = 2000,
  name = "peak1",
  score = 100,
  strand = "+",
  signalValue = 5.5,
  pValue = 0.001,
  qValue = 0.05,
  peak = 1500
)
narrowPeak_file <- tempfile(fileext = ".narrowPeak")
importBed(narrowPeak_file)

}
