% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBed.R
\name{importBed}
\alias{importBed}
\title{Import Genomic Ranges in BED Format}
\usage{
importBed(bed)
}
\arguments{
\item{bed}{Input genomic ranges in one of the following formats:
\itemize{
  \item Character vector of 1-based genomic coordinates ("chr:start-end[:strand]").
  \item File path to .bed, .narrowPeak, or .broadPeak file.
  \item GRanges object.
  \item 1-based data.frame or data.table containing genomic coordinates.
}}
}
\description{
Imports genomic coordinates from various input formats into a standardized, 1-based coordinates data.table format.
Supported input types: GRanges objects, character strings ("chr2L:30000-31000:-"), .bed, .narrowPeak
or .broadPeak file paths (see details).
}
\details{
Character coordinates should follow the syntax "chr:start-end[:strand]" and be 1-based.
BED/NARROWPEAK/BROADPEAK files will be converted from BED’s 0-based half-open to 1-based inclusive.
File specifications can be found at https://genome.ucsc.edu/FAQ/FAQformat.html.
GRanges objects are directly converted to data.table and metadata columns are preserved.
data.frames/data.tables should be 1-based and should contain the following columns: 'seqnames', 'start', 'end' and 'strand'.
If missing, end and strand will be set to 'start' and '*', respectively.
}
\examples{
# Import from character coordinates
importBed(c("chr2L:1000-2000:+", "chr3L:1000-2000:-", "chr3L:4000-6000:*"))[]

# From GRanges
gr <- importBed(GRanges("chr2L", IRanges(c(1000, 5000), c(2000, 6000))))[]

# From bed/narrowpeak files
# BED
bed_file <- tempfile(fileext = ".bed")
exportBed(gr, bed_file)
importBed(bed_file)[]
# NARROWPEAK
peaks <- data.table(
  seqnames = "chr2L",
  start = 1000,
  end = 2000,
  name = "peak1",
  score = 100,
  strand = "+",
  signalValue = 5.5,
  pValue = 0.001,
  qValue = 0.05,
  peak = 1500
)
narrowPeak_file <- tempfile(fileext = ".narrowPeak")
exportBed(peaks, narrowPeak_file)
importBed(narrowPeak_file)

}
