% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importJASPAR.R
\name{importJASPAR}
\alias{importJASPAR}
\title{importJASPAR}
\usage{
importJASPAR(
  combinedFile,
  pseudocount = 1e-05,
  bg = c(A = 0.25, C = 0.25, G = 0.25, T = 0.25)
)
}
\arguments{
\item{combinedFile}{Path to a combined file containing JASPAR-formatted taw PFMs.}

\item{pseudocount}{Pseudocount to be added before logging. Default= 1e-5.}

\item{bg}{Background probabilities. Default= c("A"= .25, "C"= .25, "G"= .25, "T"= .25).}
}
\value{
A list containing pwms_perc and pwms_log_odds TFBSTools::PWMatrix sublists.
}
\description{
Imports a JASPAR
}
\examples{
importJASPAR("/groups/stark/vloubiere/motifs_db/hand_curated_Dmel_motifs.txt")
}
