% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAplot.R
\name{MAplot}
\alias{MAplot}
\title{Plots a nice MA plot}
\usage{
MAplot(data, main = NA, padj.cutoff = 0.05, log2FC.cutoff = log2(1.5))
}
\arguments{
\item{data}{Either a valid path to a '.txt' file or a data.table object corresponding to a DESeq2 output sheet  containing both 'log2FoldChange' and 'baseMean' columns.}

\item{main}{Title}

\item{padj.cutoff}{padjust cutoff to be used to call up/down genes. Default= 0.05.}

\item{log2FC.cutoff}{log2FoldChange cutoff to be used to call up/down genes. Default= log2(1.5).}
}
\value{
MA plot
}
\description{
Plots a nice MA plot
}
\examples{
MAplot(data= "/groups/stark/vloubiere/projects/epigenetic_cancer/db/FC_tables/RNA/epiCancer_GFP_PH29_vs_W29.txt",
          main= "PH29")
}
