% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxProjectionImage.R
\name{maxProjectionImage}
\alias{maxProjectionImage}
\title{Title}
\usage{
maxProjectionImage(im, start = 1, end = dim(im)[4], FUN = pmax)
}
\arguments{
\item{im}{}

\item{start}{}

\item{end}{}

\item{FUN}{}
}
\description{
Title
}
