% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifLassoRegression.R
\name{motifLassoRegression}
\alias{motifLassoRegression}
\title{Perform LASSO Regression Using TF Motifs}
\usage{
motifLassoRegression(response, counts, names = NULL, CV = 0.05, seed = 123)
}
\arguments{
\item{response}{A numeric vector representing the variable to predict.}

\item{counts}{A matrix of motif counts, where rows correspond to observations and columns correspond to motifs.}

\item{names}{An optional character or factor vector of motif names. Defaults to the column names of `counts`.}

\item{CV}{A numeric value between 0 and 1 specifying the proportion of data to use as the test set. Default is `0.05`.}

\item{seed}{An integer specifying the random seed for reproducibility. Default is `123`.}
}
\value{
A list with the following components:
\item{test_set}{A `data.table` containing the observed and predicted values for the test set.}
\item{PCC}{The Pearson correlation coefficient between observed and predicted values on the test set.}
\item{R2}{The R-squared value (square of the Pearson correlation coefficient) for the test set.}
\item{best_predictors}{A `data.table` of the most important predictors (motifs) with non-zero coefficients, including their names and coefficients.}
\item{model}{The fitted LASSO model object.}
}
\description{
This function performs LASSO regression to identify transcription factor (TF) motifs associated with a numeric response variable.
It splits the data into training and test sets, scales the data, performs cross-validation to select the best regularization parameter (`lambda`),
and fits a LASSO model. The function returns the model, test set predictions, and the most important predictors (motifs).
}
\examples{
# Download ATAC-Seq FC between PHD11 and control from the Nature paper
tmp <- tempfile(fileext = ".txt.gz")
download.file(url = "https://ftp.ncbi.nlm.nih.gov/geo/series/GSE222nnn/GSE222193/suppl/GSE222193\%5FATAC\%5FTransient\%5Fph\%2DKD\%5FD11\%5Fvs\%5FControl\%5Fno\%5Fph\%2DKD\%5FDESeq2\%5FFC.txt.gz",
              destfile = tmp)

# Retrieve FC
FC <- fread(tmp)
FC <- cbind(importBed(FC$ID), FC[, !"ID"])

# Resize
res <- resizeBed(bed = FC,
                 center = "center",
                 upstream = 250,
                 downstream = 250,
                 genome = "dm6")

# Select JASPAR motifs
load("/groups/stark/vloubiere/motifs_db/vl_Dmel_motifs_DB_full.RData")
sel <- vl_Dmel_motifs_DB_full[collection=="jaspar", pwms_log_odds]
pwms <- vl_Dmel_motifs_DB_full[collection=="jaspar", pwms_perc]
mot.cluster <- vl_Dmel_motifs_DB_full[collection=="jaspar", motif_cluster]

# Compute motif counts
counts <- vl_motifCounts(bed= res,
                         genome= "dm6",
                         pwm_log_odds= sel)

# Lasso regression
lasso <- motifLassoRegression(response= FC$log2FoldChange,
                              counts = counts,
                              names= mot.cluster)

# Check PCC of the test set
vl_par()
plot(lasso$test_set)
addPcc(lasso$PCC)

# Retrieve top predictors
preds <- lasso$best_predictors

# Select top predictors
preds <- preds[, .SD[which.max(abs(s0))], name]
top <- preds[abs(s0)>0.03]
setorderv(top, "s0")

# Plot
barplot(top$s0,
        horiz= TRUE,
        names.arg = top$name,
        xlab= "Lasso coefficient")


}
