% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapBed.R
\name{overlapBed}
\alias{overlapBed}
\title{Find Overlapping Regions Between Two Sets of Genomic Intervals}
\usage{
overlapBed(a, b, minoverlap = 1L, ignore.strand = TRUE)
}
\arguments{
\item{a}{Query regions in any format compatible with `importBed()`.}

\item{b}{Target regions in any format compatible with `importBed()`.}

\item{minoverlap}{A single integer specifying the minimum overlap between 2 ranges for them to
be considered as overlapping. Default= 1L.}

\item{ignore.strand}{If set to FALSE, only reports overlaps between regions that are on the same strand.
If set to TRUE (default), reports overlaps on both strands.}
}
\value{
A data.table with columns:
\itemize{
  \item idx.a: line index of the region in a.
  \item idx.b: line index of the overlapping region in b.
  \item overlap.start: overlap starting position.
  \item overlap.end: overlap ending position.
  \item overlap.width: overlap width.
}
Of note, regions in a with no overlap in b are not returned.
}
\description{
A wrapper around `GenomicRanges::findOverlaps()` that identifies, for each genomic range in a,
the overlapping regions in b.
}
\examples{
# Create example regions
a <- importBed(c("chr1:100-200:+", "chr1:300-400:-", "chr2:100-200:+"))
b <- importBed(c("chr1:150-250:+", "chr1:350-450:+", "chr1:500-600:+"))

# Find all overlaps
overlapBed(a, b)

# Find strand-specific overlaps
overlapBed(a, b, ignore.strand = FALSE)

}
