% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_plot_auc_enrichment.R
\name{plot.vl_auc}
\alias{plot.vl_auc}
\title{Title}
\usage{
\method{plot}{vl_auc}(
  auc,
  cutoff = ifelse(value == "NES", 3, 0.6),
  min.diff.cutoff = 0,
  topN = 10,
  order.var = "score",
  value = "NES",
  ...
)
}
\arguments{
\item{auc}{A vl_auc object created by ?vl_motifAUC.}

\item{cutoff}{Only the variables with a maximum absolute value higher or equal to this cutoff are plotted.}

\item{min.diff.cutoff}{Only the variables with a range higher or equal to this cutoff are plotted.}

\item{topN}{The number of top values to be plotted per ranking variable. Default= 10.}

\item{value}{The value to be plotted. Possible value are "NES" or "auc". Default= "NES".}

\item{...}{Extra arguments to be passed to the heatmap funciton.}
}
\value{
Invisible returns the heatmap object.
}
\description{
Title
}
