% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vl_enr_clusters.R
\name{plot.vl_enr_cl}
\alias{plot.vl_enr_cl}
\title{Plot Clustered Enrichment Results for vl_enr_cl Objects}
\usage{
\method{plot}{vl_enr_cl}(
  obj,
  log2OR.cutoff = 0,
  padj.cutoff = 0.05,
  top.enrich = Inf,
  min.counts = 3L,
  order = "log2OR",
  color.breaks = NULL,
  size.legend.breaks = NULL,
  cex = 1,
  col = c("blue", "red"),
  main = NA,
  plot.empty.clusters = T
)
}
\arguments{
\item{obj}{A data.table of class 'vl_enr_cl' containing columns 'padj', 'log2OR', 'set_hit', 'name' and 'cl'.}

\item{log2OR.cutoff}{Numeric cutoff used to filter positive log2OR values before plotting.}

\item{padj.cutoff}{Numeric cutoff used to filter adjusted p-values before plotting.
Default= 0.05.}

\item{top.enrich}{Integer specifying the maximum number of top enrichments to plot, based on the
'order' parameter. Default= Inf (no selection).}

\item{min.counts}{Integer speifying the minimum number of counts required to include an enrichment
in the plot ('set_hit' column). Default= 3.}

\item{order}{The metric that should be used to order enrichments before selecting the top enrichments.
Possible values are 'padj' or 'log2OR'. Default= 'log2OR'.}

\item{color.breaks}{Color breaks to use for the color scale. Defaults to the range of filtered
-log10(adjusted p-values).}

\item{size.legend.breaks}{Breakpoints for the size legend in the balloon plot. Defaults to the range
of filtered log2OR values.}

\item{cex}{Numeric expansion factor used to adjust balloons' sizes. Default= 1.}

\item{col}{The color scale to use for the balloon plot. Default= c('blue', 'red').}

\item{main}{Character. The main title for the plot. Default= NA.}

\item{plot.empty.clusters}{Should empty clusters be plotted? Default= TRUE.}
}
\value{
Invisibly returns the filtered and plotted data table.
}
\description{
This function generates a balloon plot to visualize clustered enrichment results from an object of class 'vl_enr_cl'.
The balloons represent the log2 odds ratio (log2OR) of enrichment, with their size corresponding to the odds ratio
and their color representing the adjusted p-values (-log10 transformed).
}
\examples{
# Example usage:
For an example using GO enrichment, see ?vl_GOenrich()
For an example using motif enrichment, see ?vl_motifEnrich()

# Assuming `clustered_enrichment_results` is an object of class `vl_enr_cl`:
plot.vl_enr_cl(clustered_enrichment_results, padj.cutoff = 0.01, top.enrich = 5, min.counts = 5)

}
