% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNAletter.R
\name{plotDNAletter}
\alias{plotDNAletter}
\title{plot seqlogo letter}
\usage{
plotDNAletter(letter, xleft, ytop, width, height)
}
\arguments{
\item{letter}{"A", "T", "C" or "G"}

\item{xleft}{xleft position}

\item{ytop}{ytop position}
}
\description{
See function vl_seqlogo
}
