% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R
\name{plotImage}
\alias{plotImage}
\title{Title}
\usage{
plotImage(im, pxSizeX_um = NULL, scale.bar = 100)
}
\arguments{
\item{im}{Two-dimensional image to plot.}

\item{pxSizeX_um}{Pixel width in micrometres.}

\item{scale.bar}{Scale-bar length in micrometres. Use `NA` or `NULL` to omit it.}
}
\description{
Title
}
