% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{profilePlot}
\alias{profilePlot}
\title{Title}
\usage{
profilePlot(x, col, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{}

\item{col}{}

\item{xlim}{}

\item{ylim}{}

\item{...}{}
}
\description{
Title
}
