% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRegionsBed.R
\name{randomRegionsBed}
\alias{randomRegionsBed}
\title{Randomly Sample Control Genomic Regions}
\usage{
randomRegionsBed(
  bed,
  sample.bed,
  no.overlaps = TRUE,
  widths = rep(100, 1000),
  ignore.strand = TRUE
)
}
\arguments{
\item{bed}{Input genomic ranges, in any format compatible with ?importBed.}

\item{sample.bed}{Optional genomic ranges. If provided, the widths arguments will be ignored,
and control regions will be sampled to match the number of regions in bed, as well as their
widths and seqnames distribution.}

\item{no.overlaps}{If set to TRUE (default), ensures that there is no overlap between sampled
regions and sample.bed.}

\item{widths}{Integer vector specifying the width of the regions to sample. The length of the vector
determines the number of sampled regions.}

\item{ignore.strand}{If FALSE and no.overlaps is TRUE, only overlaps on the same strand are avoided.
If TRUE (default), overlaps on both strands are avoided.}
}
\value{
A gr data.table of sampled control regions.
}
\description{
Randomly sample genomic regions from a bed, matching specified widths or the distribution of another bed file.
}
\examples{
# Using widths
bed <- importBed(c("chr2L:1000-2000", "chr3R:1000-2000"))
randomRegionsBed(bed= bed,  widths= rep(10, 10))

# Using a sample.bed file
sample.bed <- importBed(rep(c("chr2L:1800-2000", "chr3R:1800-2000"), each= 3))
randomRegionsBed(bed= bed,  sample.bed= sample.bed)

}
