% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSgenomeTools.R
\name{randomRegionsBSgenome}
\alias{randomRegionsBSgenome}
\title{Randomly Sample Control Genomic Regions}
\usage{
randomRegionsBSgenome(
  genome,
  widths = rep(100, 1000),
  bed,
  restrict.seqnames = NULL,
  no.overlap = TRUE
)
}
\arguments{
\item{genome}{BSgenome name: "dm6", "mm10"...}

\item{widths}{Integer vector specifying the width of the regions to sample. The length of the vector determines the number of regions.}

\item{bed}{Optional genomic ranges. If provided, the widths arguments will be ignored, and control regions will match be sampled to match
the number of regions in bed, as well as their widths and seqnames distribution.}

\item{restrict.seqnames}{Optional character vector. Restrict sampling to these seqnames.}

\item{no.overlap}{If set to TRUE (default), ensures that there is no overlap between sampled and bed regions (if provided).
Stops after 10 iterations.}
}
\value{
A gr data.table of sampled control regions.
}
\description{
Generates random genomic regions from a BSgenome, matching specified widths or the distribution of an input BED file.
}
\examples{
# Sample using widths
test <- randomRegionsBSgenome("dm6", widths= rep(5000, 1e3))

Sampling using bed
randomRegionsBSgenome(genome = "dm6", bed = test)

}
