% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRegionsBSgenome.R
\name{randomRegionsBSgenome}
\alias{randomRegionsBSgenome}
\title{Randomly Sample Control Genomic Regions}
\usage{
randomRegionsBSgenome(
  genome,
  widths,
  restrict.seqnames = NULL,
  no.overlaps = NULL,
  ignore.strand = TRUE
)
}
\arguments{
\item{genome}{BSgenome name: "dm6", "mm10"...}

\item{widths}{Integer vector specifying the width of the regions to sample. The length of the vector
determines the number of sampled regions.}

\item{restrict.seqnames}{If specified, restricts random sampling to the corresponding seqnames.}

\item{no.overlaps}{Genomic ranges for which overlaps should be avoided dhring sampling, in any format compatible
with ?importBed. Default= NULL.}

\item{ignore.strand}{If genomic ranges are provided in no.overlaps, only overlaps on the same strand will be
avoided. If set to TRUE (default), overlaps on both strands are avoided.}
}
\value{
A gr data.table of sampled control regions.
}
\description{
Randomly sample genomic regions from a BSgenome, matching specified widths and seqnames.
}
\examples{
# Sample using widths
test <- randomRegionsBSgenome(
genome= "dm6",
widths= sample(c(1000, 2000, 5000), 1e3, replace= T),
restrict.seqnames= c("chr2L", "chr2R", "chr3L", "chr3R")
)

Sampling non-overlapping set
randomRegionsBSgenome(genome = "dm6", widths = test[, end-start+1], no.overlaps = test)

}
