% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizeBed.R
\name{resizeBed}
\alias{resizeBed}
\title{Resize Genomic Regions in a BED File}
\usage{
resizeBed(
  bed,
  center = "center",
  upstream = 500L,
  downstream = 500L,
  genome = NULL,
  ignore.strand = FALSE
)
}
\arguments{
\item{bed}{Input regions in any format compatible with ?importBed.}

\item{center}{Specifies the origin for resizing. Options are:
\itemize{
  \item `center`: Resize symmetrically around the midpoint of the region (default).
  \item `start`: Resize from the start of the region.
  \item `end`: Resize from the end of the region.
  \item `region`: Extend or contract the entire region's boundaries.
}}

\item{upstream}{Size of the extension upstream of the specified origin. Default= 500L.}

\item{downstream}{Size of the extension downstream of the specified origin. Default= 500L.}

\item{genome}{A BS genome name (e.g. "dm6", "mm10") used to clipped resized regions.
Default= NULL (no clipping).}

\item{ignore.strand}{If set to TRUE, resizing always proceeds from the leftmost coordinates,
irrespective of the strand. If set to FALSE (default), upstream and downstream resizing respect
the strand.}
}
\value{
A data.table containing resized regions.
}
\description{
Resize genomic ranges.
}
\examples{
# Import example BED file
bed <- importBed(c("chr2L:1000-2000:+", "chr2L:1000-2000:-"))

# Resize using different centering options
resizeBed(bed, center = "start", upstream = 100, downstream = 200)[]
resizeBed(bed, center = "center", upstream = 100, downstream = 200)[]
resizeBed(bed, center = "end", upstream = 100, downstream = 200)[]
resizeBed(bed, center = "region", upstream = 100, downstream = 200)[]

# Ignore strand during resizing
resizeBed(bed, center = "center", upstream = 100, downstream = 200, ignore.strand = TRUE)[]

}
