% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revCompDNA.R
\name{revCompDNA}
\alias{revCompDNA}
\title{reverseComplement}
\usage{
revCompDNA(DNA_char, complement = TRUE, reverse = TRUE)
}
\arguments{
\item{DNA_char}{Input DNA character string.}

\item{complement}{Should the DNA sequence be complemented? Default= TRUE}

\item{reverse}{Should the DNA sequence be reversed? Default= TRUE}
}
\value{
Reverse Complemented DNA sequence
}
\description{
Reverse complement DNA sequence character string.
}
\examples{
revCompDNA("ATCG")

}
