% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revCompDNA.R
\name{revCompDNA}
\alias{revCompDNA}
\title{reverseComplement}
\usage{
revCompDNA(DNA_char, reverse = TRUE, complement = TRUE, as.character = TRUE)
}
\arguments{
\item{DNA_char}{Character vector, DNAString or DNAStringSet.}

\item{reverse}{Should the DNA sequence be reversed? Default= TRUE}

\item{complement}{Should the DNA sequence be complemented? Default= TRUE}

\item{as.character}{Should the sequences be returned as a character vector? Otherwise,
it will be returned as a DNAString or a DNAStringSet. Default= TRUE}
}
\value{
Reverse Complemented DNA sequence
}
\description{
Reverse complement DNA sequence character string.
}
\examples{
revCompDNA("ATCG")

}
