% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqProcessing.R
\name{rnaseqProcessing}
\alias{rnaseqProcessing}
\title{Process RNA-Seq Data}
\usage{
rnaseqProcessing(
  fq1,
  fq2 = NULL,
  output.prefix,
  genome,
  genome.idx = NULL,
  gtf = NULL,
  GTF.attrType.extra,
  fq.output.folder = "db/fq/RNASeq/",
  bam.output.folder = "db/bam/RNASeq/",
  alignment.stats.output.folder = "db/alignment_stats/RNASeq/",
  counts.stats.output.folder = "db/stats/RNASeq/",
  counts.output.folder = "db/counts/RNASeq/",
  bw.output.folder = "db/bw/RNASeq/",
  Rpath = "/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript",
  cores = 8
)
}
\arguments{
\item{fq1}{Path(s) to input R1 FASTQ file(s).}

\item{fq2}{Path(s) to input R2 FASTQ file(s) for paired-end data. Default: `NULL`.}

\item{output.prefix}{Prefix for output files.}

\item{genome}{Reference genome identifier (e.g., `"mm10"`, `"hg38"`).}

\item{genome.idx}{Path to the Rsubread genome index. If `NULL`, derived from `genome`. Default: `NULL`.}

\item{gtf}{Path to the GTF annotation file. If `genome` is provided, corresponding protein-coding / mRNA genes will be used. Default: `NULL`.}

\item{GTF.attrType.extra}{Additional GTF attribute to include in the output (e.g., gene symbol).}

\item{fq.output.folder}{Directory for trimmed FASTQ files. Default: `"db/fq/RNASeq/"`.}

\item{bam.output.folder}{Directory for aligned BAM files. Default: `"db/bam/RNASeq/"`.}

\item{alignment.stats.output.folder}{Directory for alignment statistics. Default: `"db/alignment_stats/RNASeq/"`.}

\item{counts.stats.output.folder}{Directory for gene count statistics. Default: `"db/stats/RNASeq/"`.}

\item{counts.output.folder}{Directory for gene counts. Default: `"db/counts/RNASeq/"`.}

\item{bw.output.folder}{Directory for BigWig files. Default: `"db/bw/RNASeq/"`.}

\item{Rpath}{Path to the Rscript binary. Default: `"/software/f2022/software/r/4.3.0-foss-2022b/bin/Rscript"`.}

\item{cores}{Number of CPU cores to use. Default: `8`.}
}
\value{
A `data.table` with:
- `file.types`: Types of output files.
- `path`: Paths to the output files.
- `cmd`: Shell commands for each step in the pipeline.
}
\description{
Implements a pipeline for processing RNA-Seq data, including:
1. Adapter trimming
2. Genome alignment
3. Gene counting
4. BigWig track generation

Supports single-end and paired-end data, with options for custom genome and annotation files.
}
\examples{
# Process paired-end RNA-Seq data
cmd <- rnaseqProcessing(
  fq1 = c("sample1_R1.fq.gz", "sample1_R1_rep2.fq.gz"),
  fq2 = c("sample1_R2.fq.gz", "sample1_R2_rep2.fq.gz"),
  output.prefix = "sample1",
  genome = "hg38",
  gtf = "/data/annotations/hg38.gtf",
  GTF.attrType.extra = "gene_name",
  cores = 8
)
vl_submit(cmd, execute= FALSE)

# Then, differential analysis should be performed using ?cmd_DESeq2()

}
\seealso{
\itemize{
  \item \code{\link{cmd_trimIlluminaAdaptors}} for adapter trimming
  \item \code{\link{cmd_alignRsubread}} for alignment
  \item \code{\link{cmd_countRsubread}} for gene counting
  \item \code{\link{cmd_bamToBigwig}} for BigWig track generation
}
}
