% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_markerHeatmap.R
\name{sc_markerHeatmap}
\alias{sc_markerHeatmap}
\title{Plot and Cluster Top Marker Genes as a Heatmap from Single-Cell Transcriptome Data}
\usage{
sc_markerHeatmap(
  dat,
  topN = 3,
  selection = NULL,
  cluster.selection = FALSE,
  order.selection = FALSE,
  cluster.rows = FALSE,
  cutree.rows = NULL,
  show.col.clusters = TRUE,
  row.annotation.column = NULL,
  row.annotation.colors = c("#A3D8FF", "#B6FF8A", "#FFE066", "#FF6F69"),
  row.annotation.title = NULL,
  col.annotation.column = NULL,
  col.annotation.colors = c("lightgrey", "chartreuse", "red", "gold", "royalblue1",
    "royalblue3", "blue"),
  col.annotation.title = NULL,
  perc.cutoff = 10,
  perc.diff.cutoff = 20,
  order.var = "perc.diff",
  padj.wilcox.cutoff = 0.05,
  logFC.cutoff = 0.25,
  padj.fisher.cutoff = NULL,
  log2OR.cutoff = NULL,
  auc.cutoff = NULL,
  main = NA,
  alternative = "greater",
  add.cells.number = TRUE,
  pdf.file = NULL
)
}
\arguments{
\item{dat}{A data.table containing marker gene statistics (see ?sc_computeMarkerGenes).}

\item{topN}{Number of top genes to select per cluster. Ddefault= 3.
If set to 0 or NULL, only the genes listed in selection (next argument) are used.}

\item{selection}{Character vector of selected gene symbols that should be added to the heatmap.
Default= NULL.}

\item{cluster.selection}{Should select genes be used to cluster rows or not? Default= FALSE.}

\item{order.selection}{Should selected gene columns be ordered?}

\item{cluster.rows}{Logical; whether to cluster heatmap rows. Default= TRUE.}

\item{cutree.rows}{Number of row clusters to cut the dendrogram into. Default= 10.}

\item{show.col.clusters}{Should column cluster names be shown? Default= TRUE.}

\item{perc.cutoff}{Minimum percentage of expressing cells within the cluster. Default= 10.}

\item{perc.diff.cutoff}{Minimum difference in percentage of expressing cells within vs. outside the cluster.
Default= 20.}

\item{order.var}{Variable to order genes by; one of "perc.diff", "auc", "logFC", or "log2OR".
Default= "perc.diff".}

\item{padj.wilcox.cutoff}{Adjusted Wilcoxon p-value cutoff. Default= 0.05.}

\item{logFC.cutoff}{Log fold change cutoff. Default= 0.}

\item{padj.fisher.cutoff}{Adjusted Fisher p-value cutoff. Default= NULL.}

\item{log2OR.cutoff}{Log2 odds ratio cutoff. Default= NULL.}

\item{auc.cutoff}{AUC cutoff. Default= NULL.}

\item{main}{Title.}

\item{alternative}{"greater" for upregulated markers, "smaller" for downregulated. Default= "greater".}

\item{add.cells.number}{If set to TRUE, a barplot specifying the number of cells is plotted. Default= TRUE.}

\item{pdf.file}{If specified, the heatmap will be saved in a pdf with an optimized layout. Default= NULL.}
}
\value{
Invisibly returns a list containing the ordered heatmap matrix.
}
\description{
Selects top marker genes per cluster based on multiple statistical cutoffs, then visualizes their expression as a heatmap.
Optionally clusters rows and/or orders columns, and allows highlighting of selected genes.
}
