% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_topMarkers.R
\name{sc_topMarkers}
\alias{sc_topMarkers}
\title{Plot and cluster top marker genes from single-cell transcriptome data}
\usage{
sc_topMarkers(
  dat,
  perc.cutoff = 10,
  perc.diff.cutoff = 20,
  padj.wilcox.cutoff = 0.05,
  logFC.cutoff = 0,
  padj.fisher.cutoff = 1,
  log2OR.cutoff = 1,
  auc.cutoff = 0.5,
  order.var = "perc.diff",
  topN = Inf,
  alternative = "greater"
)
}
\arguments{
\item{dat}{Path to an .rds file containing marker genes.}

\item{perc.cutoff}{Minimum percentage of expressing within cluster.}

\item{perc.diff.cutoff}{Minimum difference in percentage of expressing cells within cluster vs. outside cluster.}

\item{padj.wilcox.cutoff}{Adjusted Wilcoxon p-value cutoff. Default= 0.05.}

\item{logFC.cutoff}{Log fold change cutoff. Default= 0.1.}

\item{padj.fisher.cutoff}{Adjusted Fisher p-value cutoff. Default= 0.05.}

\item{log2OR.cutoff}{Log2 odds ratio cutoff. Default= 1.}

\item{auc.cutoff}{AUC cutoff. Default= 0.}

\item{order.var}{Variable to order genes by ("perc.diff", "auc", "logFC", "log2OR")}

\item{topN}{Number of top genes per cluster.}
}
\value{
The subset of the rds object corresponding to top marker genes.
}
\description{
Plot and cluster top marker genes from single-cell transcriptome data
}
