% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_UMAP.R
\name{sc_UMAP}
\alias{sc_UMAP}
\title{Plot UMAP from seurat object}
\usage{
sc_UMAP(
  umap,
  value.var = NULL,
  show.labels = TRUE,
  main = NULL,
  col = NULL,
  NA.col = "lightgrey",
  zlim = NULL,
  legend.main = NULL,
  cex.pts = 0.4,
  pch = 16,
  transparency = 0.7,
  cex.lab = 0.7,
  cex.legend = 0.7,
  randomize = FALSE,
  select.cell = NULL,
  layer = "SCT",
  rect.width = NULL,
  rect.height = NULL,
  add = FALSE
)
}
\arguments{
\item{umap}{A seurat object or a data.table/data.frame/matrix containing columns 'cellID', 'umap_1', 'umap_2'.}

\item{value.var}{The value to plot. Should either be a factor vector matching the number of cells in the object,
or a unique gene symbol.}

\item{show.labels}{Should the factor levels be shown on the UMAP itself? Otherwise, they will be shown in a legend.}

\item{cex.lab}{Expansion factor for labels.}

\item{cex.legend}{Expansion factor for the legend.}

\item{layer}{The layer to use. Default= "SCT".}
}
\value{
Invisibly returns a list containing the ordered heatmap matrix.
}
\description{
Function to plot rasterized umaps from a seurat object.
}
