% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somClustering.R
\name{somClustering}
\alias{somClustering}
\title{somClustering}
\usage{
somClustering(
  layers,
  grid = c(2, 3),
  clip.perc = c(0.05, 0.95),
  scale = FALSE,
  user.weights = NULL,
  init.seed = 1,
  maxNA.fraction = 0L
)
}
\arguments{
\item{layers}{Either a matrix or a list of matrices to cluster.}

\item{grid}{THe size of the grid. Default= c(2,3).}

\item{clip.perc}{Percentile used to clip numeric matrices. Set to c(0,1) to skip clipping.}

\item{scale}{Should the data be scaled using z-score? Default= FALSE}

\item{user.weights}{User-defined weights for the different layers. Length should be match the length of layers' list. Default= 1 for all layers.}

\item{init.seed}{Seed used for random initialization/}

\item{maxNA.fraction}{The maxNA fraction tolerated for a given line. Default= 0L.}
}
\description{
A wrapper around the som kohonen function to allow quick tests.
}
