% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringGetDB.R
\name{stringGetDB}
\alias{stringGetDB}
\title{Build STRING db}
\usage{
stringGetDB(species, network_type = "full", version = "11.0")
}
\arguments{
\item{species}{either "Dm" or "Mm"}

\item{network_type}{The type of interactions to be used. Can be one of "full" (full functional, default) or "physical" if only physical interactions are to be considered}

\item{version}{databse version. default= "10"}
}
\description{
Build STRING db
}
