% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_barplot.R
\name{vl_barplot}
\alias{vl_barplot}
\title{Barplot}
\usage{
vl_barplot(
  x,
  names.arg = NULL,
  top.labels = NULL,
  show.sd = TRUE,
  show.points = TRUE,
  beside = FALSE,
  xlim = NULL,
  ylim = NULL,
  tilt.names = TRUE,
  top.labels.cex = 0.7,
  sd.arrow.lwd = 0.5,
  sd.arrow.length = ifelse(length(x) > 1, diff(bar[c(1, 2)])/8, 0.2),
  pch = 16,
  pch.col = adjustcolor("grey20", 0.7),
  pch.cex = 0.5,
  pch.jitter = 0.2,
  xaxt = "s",
  width = 0.6,
  space = NULL,
  xaxs = "i",
  horiz = FALSE,
  ...
)
}
\arguments{
\item{x}{A vector, matrix, data.frame, or a list of numeric variables to plot.}

\item{names.arg}{Names to plot on the x-axis.}

\item{top.labels}{Labels to plot on the top of each bar.}

\item{show.sd}{If x is a list, should the SD be shown? Default = TRUE.}

\item{show.points}{If x is a list, should the points be shown? Default = TRUE.}

\item{beside}{If TRUE and x is a matrix or data.frame, bars are juxtaposed. Default = FALSE (stacked bars).}

\item{xlim}{X-axis limits. Default = NULL.}

\item{ylim}{Y-axis limits. Default = NULL.}

\item{tilt.names}{Should bar names be tilted? Default = TRUE.}

\item{top.labels.cex}{Expansion factor for top.labels.}

\item{sd.arrow.lwd}{Line width of SD arrows. Default = 0.5.}

\item{sd.arrow.length}{Length of SD arrows. Defaults to 1/8 of the distance between bar centers.}

\item{pch}{Plotting character for points. Default = 16.}

\item{pch.col}{Color for points.}

\item{pch.cex}{Expansion factor for points.}

\item{pch.jitter}{Jitter amount for points. Default = 0.2.}

\item{xaxt}{If tilt.names is set to FALSE, style to be used for the x-axis. Default = "s".}

\item{horiz}{Not supported at the moment.}

\item{...}{Additional arguments passed to \code{barplot()}.}
}
\value{
A barplot.
}
\description{
A wrapper around barplot with nicer default settings, allowing input as a list to show both SD and/or individual points.
}
\examples{
vl_barplot(list(1:3, 5:8), top.labels = c("A", "B"), show.sd = FALSE)
vl_barplot(matrix(1:6, ncol = 2), top.labels = c("A", "B"), beside = TRUE)
}
