% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_barplot.R
\name{vl_barplot}
\alias{vl_barplot}
\title{Title}
\usage{
vl_barplot(
  x,
  names.arg = NULL,
  show.sd = TRUE,
  show.points = TRUE,
  beside = FALSE,
  width = 0.6,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  sd.arrow.lwd = 0.5,
  sd.arrow.length = width/5,
  pch = 16,
  pch.col = adjustcolor("grey20", 0.7),
  pch.cex = 0.5,
  pch.jitter = 0.2,
  xaxt = "s",
  xaxs = "i",
  frame = F,
  horiz = FALSE,
  add = F,
  xlab = "",
  ylab = "",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
