% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_barplot.R
\name{vl_barplot}
\alias{vl_barplot}
\title{Barplot allowing to add points and sd}
\usage{
vl_barplot(
  height,
  individual.var = NULL,
  show.sd = !is.null(individual.var),
  bar.top.labels = NULL,
  compute.bar.diff = NULL,
  xlim = NULL,
  ylim = NULL,
  tilt.names = TRUE,
  individual.var.pch = 16,
  individual.var.col = adjustcolor("lightgrey", 0.7),
  individual.var.jitter = 0.2,
  individual.var.cex = 0.5,
  sd.arrow.lwd = 0.5,
  sd.arrow.length = ifelse(length(height) > 1, diff(bar[c(1, 2)])/8, 0.2),
  bar.top.labels.cex = 0.7,
  compute.bar.diff.digits = 2,
  compute.bar.diff.cex = 0.8,
  xpd = NA,
  horiz = F,
  xaxt = "s",
  names.arg = names(height),
  ...
)
}
\arguments{
\item{height}{A vector of values describing the bars which make up the plot.}

\item{individual.var}{List of individual variables to be plotted as points.}

\item{show.sd}{Should sd from individual var be shown as whiskers around bars?}

\item{bar.top.labels}{Labels to be shown on the top or each bar.}

\item{compute.bar.diff}{List of pairwise bar pairs to be compared (V2/V11 ratio)}

\item{individual.var.pch}{pch value for individual variables}

\item{individual.var.col}{color value for individual variables}

\item{individual.var.jitter}{jitter value for individual variables}

\item{individual.var.cex}{cex factor for points}

\item{sd.arrow.lwd}{Line width of sd arrows Default= .5.}

\item{sd.arrow.length}{Length of sd arrow length. defaults to 1/8 of distance between bar centers.}

\item{bar.top.labels.cex}{Expansion factor for bar.top.labels.}

\item{compute.bar.diff.digits}{Number of compute.bar.diff.digits for the computed diff. Default= 2}

\item{compute.bar.diff.cex}{cex value for pairwise comparisons}

\item{xpd}{xpd value for individual.var, sd, pairwaise comparisons...}

\item{horiz}{Not supported atm}

\item{...}{Extra arguments to be passed to barplot}
}
\value{
A simple barplot
}
\description{
Barplot allowing to add points and sd
}
\examples{
vl_barplot(1:3, rep(.2, 3))
}
