% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_iCisTarget.genes.R
\name{vl_iCisTarget.genes}
\alias{vl_iCisTarget.genes}
\title{A wrapper around the rCisTarget package that identify motif enriched in Dmel genes.}
\usage{
vl_iCisTarget.genes(
  geneList,
  highlightTFs = NULL,
  NES.cutoff = 3,
  maxRank = 5000,
  aucMaxRank = 0.05,
  plot = T,
  cluster.rows = F,
  cluster.cols = F,
  drop.empty = FALSE,
  legend.title = "NES",
  show.grid = T,
  show.numbers = T,
  motifRankings.feather =
    "/zssd/scratch/vincent.loubiere/motifs_db/iCisTarget/mc_v10_clust/gene_based/dm6_v10_clust.genes_vs_motifs.rankings.feather",
  motifAnnot.tbl =
    "/zssd/scratch/vincent.loubiere/motifs_db/iCisTarget/motif2tf/motifs-v10nr_clust-nr.flybase-m0.001-o0.0.tbl",
  nCores = 4,
  cleanup.cache = FALSE,
  ...
)
}
\arguments{
\item{geneList}{A named list of Dmel gene symbols.}

\item{highlightTFs}{If a character list of TF names is provided, the column TFinDB in the otuput table will indicate whether
any of those TFs are included within the 'high-confidence' annotation (two asterisks, **) or 'low-confidence' annotation
(one asterisk, *) of the motif. The vector can be named to indicate which TF to highlight for each gene-set. 
Otherwise, all TFs will be used for all geneSets.}

\item{NES.cutoff}{NES threshold to identify significant motifs. The NES is calculated -for each motif- based on the AUC distribution of all the motifs for the gene-set [(x-mean)/sd]. Default= 3.0.}

\item{maxRank}{The maximum rank to take into account for the gene enrichment recovery curve. Default= 5000.}

\item{aucMaxRank}{Threshold to calculate the AUC. In a simplified way, the AUC value represents the fraction of genes
(within the top X genes in the ranking) that are included in the signature. Default= 0.05 (5% of the total number of
genes in the rankings). Common values: 1-10%.}

\item{plot}{Should the result be plotted?}

\item{cluster.rows}{Should the rows of the resulting heatmap be clustered? Default= FALSE}

\item{cluster.cols}{Should the columns of the resulting heatmap be clustered? Default= FALSE}

\item{drop.empty}{Should sub-lists with no enrichment be dropped? Default= FALSE.}

\item{legend.title}{Default= "NES".}

\item{show.grid}{Should the grid be plotted? Default= TRUE.}

\item{show.numbers}{Should the NES values be plotted? Default= TRUE.}

\item{motifRankings.feather}{Path to the gene motifRankings .feather file. Default=
"/zssd/scratch/vincent.loubiere/motifs_db/iCisTarget/mc_v10_clust/gene_based/dm6_v10_clust.genes_vs_motifs.rankings.feather".}

\item{motifAnnot.tbl}{Path to the gene motifAnnot .tbl file. Default=
"/zssd/scratch/vincent.loubiere/motifs_db/iCisTarget/motif2tf/motifs-v10nr_clust-nr.flybase-m0.001-o0.0.tbl"}

\item{nCores}{}

\item{...}{Extra arguments passed to the vl_heatmap function.}

\item{motifAnnot.feather}{}
}
\description{
A wrapper around the rCisTarget package that identify motif enriched in Dmel genes.
}
