% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_MCC.R
\name{vl_MCC}
\alias{vl_MCC}
\title{Compute Matthew's Correlation Coefficient}
\usage{
vl_MCC(predicted, label)
}
\arguments{
\item{predicted}{Predicted values from the model (ranging from 0 to 1).}

\item{label}{A vector of logical labels (or that can be coerced to logical).}
}
\value{
Matthew's correlation coefficient
}
\description{
Matthew’s correlation coefficient (MCC) is a balanced binary-classification metric that measures the correlation between predicted and true labels.
}
