% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_motifPos.R
\name{vl_motifPos}
\alias{vl_motifPos}
\title{Find Motif Positions in Sequences or Genomic Regions}
\usage{
vl_motifPos(
  sequences,
  bed,
  pwm_log_odds,
  genome = NULL,
  bg = "genome",
  p.cutoff = 0.00005,
  pos.strand = FALSE,
  collapse.overlapping = FALSE,
  scratch = "/scratch-cbe/users/vincent.loubiere/motifs/",
  cleanup.cache = FALSE
)
}
\arguments{
\item{sequences}{A named character vector of sequences to analyze. This argument takes precedence over the bed argument.}

\item{bed}{Genomic ranges in a format compatible with ?importBed, from which genomic sequences will be retrieved when sequences is set to NULL.}

\item{pwm_log_odds}{A PWMatrixList (in log2 odds ratio format) containing motifs to map. For example, see "/groups/stark/vloubiere/motifs_db/".}

\item{genome}{The genome to use as background when bg = "genome" and/or to retrieve sequences (when bed is specified).
Default= NULL.}

\item{bg}{The background model for motif detection. Options are "genome", "subject" (inferred from input sequences) or "even" (0.25, 0.25, 0.25, 0.25). Default= "genome".}

\item{p.cutoff}{The p-value cutoff for motif detection. Default= 5e-5.}

\item{pos.strand}{If set to TRUE, only motifs on the positive strand are considered (default= FALSE).}

\item{collapse.overlapping}{Logical. If TRUE, motifs with 70 percent overlap ore more will be collapsed, irrespective of their strand. Default= FALSE.}

\item{scratch}{The path to the scratch directory for temporary files. Default= "/scratch-cbe/users/vincent.loubiere/motifs/".}

\item{cleanup.cache}{If set to TRUE, overwrites cached intermediate results. Default= FALSE.}
}
\value{
A data.table containing motif positions for each sequence, with the following columns:
- `motif`: The name of the motif.
- `seqlvls`: The sequence names.
- `mot.count`: The number of motifs detected.
- `ir`: A nested `data.table` with columns for motif positions (`start`, `end`, `strand`, `width`, `score`).
}
\description{
A wrapper around motifmatchr::matchMotifs that maps motif positions in a set of genomic coordinates or sequences or using a PWMatrixList as input.
}
\examples{
# Load Drosophila motifs
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
pwm <- importJASPAR(pfm.file)$pwms_log_odds

# Test sequence containing two overalapping Jra motifs plus a non overlapping one
test <- c("Jra"= "CATGAGTCAGGTGAGAAATGAGTCAT")
mot <- vl_motifPos(sequences = test, pwm_log_odds = pwm["Jra"], genome = "dm3", p.cutoff = 5e-4)
motifPosToBed(mot)
coll <- vl_motifPos(sequences = test, pwm_log_odds = pwm["Jra"], genome = "dm3", p.cutoff = 5e-4, collapse.overlapping = T)
motifPosToBed(coll)

# Download Dev and Hk enhancers from pe-STARR-Seq paper
tmp <- tempfile(pattern = '.xlsx')
download.file(
url = 'https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41467-024-52921-2/MediaObjects/41467_2024_52921_MOESM4_ESM.xlsx',
destfile = tmp
)
enh <- as.data.table(readxl::read_xlsx(tmp, sheet = 2, skip = 1))
set <- rbind(
  enh[group=='dev' & detail \%in\% c('medium', 'strong')][1:100, .(group, seqnames, start, end, enhancer_sequence)],
  enh[group=='hk' & detail \%in\% c('medium', 'strong'), .(group, seqnames, start, end, enhancer_sequence)]
)
seq <- set$enhancer_sequence
names(seq) <- make.unique(set$group)

# Map motifs
mot <- vl_motifPos(sequences = seq, pwm_log_odds = pwm[c("Dref", "Jra")], bg = "subject", p.cutoff = 5e-4)

# Look at counts: Jra should be more frequent in developemtan and Dref in housekeeping
mot[, class:= tstrsplit(seqlvls, "[.]", keep= 1)]
mot[, rm.NA:= ifelse(is.na(mot.count), 0, mot.count)]
vl_boxplot(rm.NA~motif+class, mot, outline= T)

# Convert to a regular bed file (sequences with no matches are discarded)
motifPosToBed(mot)

# Convert to scores matrix (useful for heatmaps)
mat <- vlite::motifPosToMatrix(mot, seqWidth = nchar(seq)[1])$Dref
vl_heatmap(mat, cluster.rows = , show.row.clusters = "left", main= "Dref max scores")

}
