% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_prAUC.R
\name{vl_prAUC}
\alias{vl_prAUC}
\title{PR AUC}
\usage{
vl_prAUC(
  rank,
  label,
  compute.NES = TRUE,
  N.iter = 100L,
  decreasing = TRUE,
  plot = FALSE,
  main = "PR",
  xlab = "Recall (TPR)",
  ylab = "Precision (PPV)",
  type = "l",
  add = FALSE,
  ...
)
}
\arguments{
\item{rank}{A numeric variable used for ranking (higher = more positive when decreasing=TRUE).}

\item{label}{A vector of logical labels (or that can be coerced to logical).}

\item{compute.NES}{Should a Normalized Enrichment Score be computed by label permutation? Default= TRUE.}

\item{N.iter}{Number of iterations used to compute the NES. Default= 100L.}

\item{decreasing}{Should the sort order for rank be decreasing? If FALSE, ranks are inverted for PRROC. Default= TRUE.}

\item{plot}{Should the PR curve be plotted? Default= FALSE.}

\item{main}{Main title. Default= "PR".}

\item{xlab}{xlab. Default= "Recall (TPR)".}

\item{ylab}{ylab. Default= "Precision (PPV)".}

\item{type}{The type of plot. Default= "l".}

\item{add}{Should the PR curve be added to an existing plot?}

\item{...}{Extra arguments passed to plot (when plot=TRUE and add=FALSE) or lines (when add=TRUE).}

\item{xlim}{Plot xlim. Default= c(0, 1).}

\item{ylim}{Plot ylim. Default= c(0, 1).}
}
\value{
PR AUC, NES and/or plots the PR curve.
}
\description{
Computes Precision-Recall AUC and optionally plots the PR curve.
}
\examples{
# Labels
label <- c(1,0,0,0,0, 1,0,0,0,0)  # 2 positives, 8 negatives (imbalanced)

# Case A: good ROC AUC AND good PR (positives ranked at the very top)
rank_good <- c(0.99, 0.20,0.19,0.18,0.17, 0.98, 0.16,0.15,0.14,0.13)

# Case B: good ROC AUC but bad PR (one negative is ranked above both positives)
# Positives are still above *most* negatives, so ROC AUC stays high, but early precision is poor, so PR suffers.
rank_bad <- c(0.90, 0.99,0.20,0.19,0.18, 0.89, 0.17,0.16,0.15,0.14)

# Compare roc and pr AUCs
vl_par(mfrow= c(2,2))
vl_rocAUC(rank= rank_good, label= label, plot= T)
vl_prAUC(rank= rank_good, label= label, plot= T)
vl_rocAUC(rank= rank_bad, label= label, plot= T)
vl_prAUC(rank= rank_bad, label= label, plot= T)

}
