% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_rocAUC.R
\name{vl_rocAUC}
\alias{vl_rocAUC}
\title{ROC AUC}
\usage{
vl_rocAUC(
  rank,
  label,
  compute.NES = TRUE,
  N.iter = 100L,
  decreasing = TRUE,
  plot = FALSE,
  main = "ROC",
  xlab = "FPR",
  ylab = "TPR",
  type = "l",
  add = FALSE,
  ...
)
}
\arguments{
\item{rank}{A numeric variable used for ranking (decreasing= TRUE by default. See decreasing argument).}

\item{label}{A vector of logical labels (or that can be coerced to logical).}

\item{compute.NES}{Should the Normalized Enrichment Score be computed (Niter= 1000)? Default= TRUE.}

\item{N.iter}{Number of iterations used to compute the NES. Default= 100L}

\item{decreasing}{Should the sort order for rank be decreasing? Default= TRUE.}

\item{plot}{Should the ROC be plotted? Default= FALSE.}

\item{main}{Main title Default= "ROC".}

\item{xlab}{xlab. Default= "FPR".}

\item{ylab}{ylab. Default= "TPR".}

\item{type}{The type of plot. Default= "l".}

\item{add}{Should the ROC line be added to an existing plot?}

\item{...}{Extra arguments to be passed to plot (when plot= TRUE) or lines (when plot= FALSE and add= TRUE).}
}
\value{
ROC AUC
}
\description{
Computes roc AUC and eventually plot the corresponding curve.
}
