% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_rocAUC.R
\name{vl_rocAUC}
\alias{vl_rocAUC}
\title{ROC AUC}
\usage{
vl_rocAUC(
  ranking_score,
  labels,
  compute.NES = TRUE,
  plot = FALSE,
  main = "ROC",
  xlab = "FPR",
  ylab = "TPR",
  type = "l",
  add = FALSE,
  ...
)
}
\arguments{
\item{ranking_score}{ranking_score values from the model.}

\item{labels}{A vector of logical labels (or that can be coerced to logical).}

\item{plot}{Should the ROC be plotted? Default= FALSE.}

\item{main}{Main title Default= "ROC".}

\item{xlab}{xlab. Default= "FPR".}

\item{ylab}{ylab. Default= "TPR".}

\item{type}{The type of plot. Default= "l".}

\item{add}{Should the ROC line be added to an existing plot?}

\item{...}{Extra arguments to be passed to plot (when plot= TRUE) or lines (when plot= FALSE and add= TRUE).}

\item{NES}{Should the Normalized Enrichment Score be computed (Niter= 1000)? Default= TRUE.}
}
\value{
ROC AUC
}
\description{
ROC AUC
}
