% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_rocAUC_grid.R
\name{vl_rocAUC_grid}
\alias{vl_rocAUC_grid}
\title{Title}
\usage{
vl_rocAUC_grid(
  ranks,
  labels,
  compute.NES = TRUE,
  N.iter = 100L,
  decreasing = TRUE,
  ranks.name = "rank",
  labels.name = "label",
  cleanup.cache = FALSE
)
}
\arguments{
\item{ranks}{A data.table of variable used for ranking (decreasing= TRUE by default. See decreasing argument).}

\item{labels}{A data.table of logical labels (or that can be coerced to logical) for which ROC AUC will be computed .}

\item{compute.NES}{Should the Normalized Enrichment Score be computed (Niter= 1000)? Default= TRUE.}

\item{N.iter}{Number of iterations used to compute the NES. Default= 100L}

\item{decreasing}{Should the sort order for ranks be increasing or decreasing? Default= TRUE.}

\item{ranks.name}{The name to give to the ranks variables. Default= "rank".}

\item{labels.name}{The name to give to the labels variables. Default= "label".}

\item{cleanup.cache}{Should the temporary files be overwritten?}
}
\description{
Title
}
