% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_seqLogo.R
\name{vl_seqLogo}
\alias{vl_seqLogo}
\title{Plot motif logo from a PFM, PPM or ICM matrix}
\usage{
vl_seqLogo(
  mat,
  name = NULL,
  xlab = "Position",
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  add = F,
  ...
)
}
\arguments{
\item{mat}{A matrix (or a TFBSTools PFMatrix, PWMatrix or ICMatrix) with rows 'A', 'C', 'G', 'T'.}

\item{name}{The name of the motif which will be added as a title.}

\item{xlab}{xlab. Default= "Position".}

\item{ylab}{ylab. Default= "Frequency" for PFMatrix, "Probability" for PWMatrix and "Bits" for ICMatrix.}

\item{xlim}{xlim for the plot. Default= NULL.}

\item{ylim}{ylim for the plot. Default= NULL.}

\item{...}{Extra parameters for the plot function.}
}
\description{
Generic function that plots a motif logo out of a matrix.
}
\examples{
# Select br PPM
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
mot <- vlite::importJASPAR(pfm.file)

# Plot
vl_par()
vl_seqLogo(mot$PFM[[1]], name= "br")
vl_seqLogo(mot$PPM[[1]], name= "br")
vl_seqLogo(mot$ICM[[1]], name= "br")

}
