% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_seqLogo.R
\name{vl_seqLogo}
\alias{vl_seqLogo}
\title{Plot PWM seqlogo}
\usage{
vl_seqLogo(PWM, name = NA)
}
\arguments{
\item{name}{Name of the motif to plot as a title.}

\item{pwm}{A PWM.}
}
\description{
Transforms a PWM into an ICM before plotting the logo,
}
\examples{
# Retrieve br motif
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
pwm <- importJASPAR(pfm.file)$pwms_log_odds[["br"]]

# Plot
vl_seqLogo(pwm)

}
