% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_seqLogo.R
\name{vl_seqLogo}
\alias{vl_seqLogo}
\title{Plot PWM seqlogo}
\usage{
vl_seqLogo(ICM, sel = 1, name = NULL, ...)
}
\arguments{
\item{ICM}{A TFBSTools ICMatrix or ICMatrixList.}

\item{sel}{If ICM is an ICMatrixList, index or name of the ICM that should be plotted.
Default= 1.}

\item{name}{Name of the motif to plot as a title. By default, the name stored in the ICM will be used.}

\item{...}{Extra plotting parameters for the plot funciton.}
}
\description{
Plots an ICM after removing low information bases (see ICM schneider correction).
}
\examples{
# Retrieve br motif
pfm.file <- system.file("extdata/hand_curated_Dmel_motifs_SCENIC_lite_Dec_2025.pfm", package = "vlite")
ICM <- importJASPAR(pfm.file)$ICM[["br"]]

# Plot
vl_seqLogo(ICM)

}
