% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_submit.R
\name{vl_submit}
\alias{vl_submit}
\title{Submit Commands to a Server with Resource Management}
\usage{
vl_submit(
  cmd,
  cores,
  mem,
  time,
  job.name,
  logs,
  modules,
  overwrite = FALSE,
  execute = TRUE,
  create.output.dirs = TRUE
)
}
\arguments{
\item{cmd}{A data.table containing the commands to execute, which will all be submitted as a single job. Intended columns are:
- `file.type`: Type of output file.
- `path`: Path to the output file.
- `cmd`: Shell command to execute.
- `cores`: (Optional) The number of CPU cores to allocate to the job (only the value from the first row will be used).
- `mem`: (Optional) Memory allocation in GB (only the value from the first row will be used).
- `time`: (Optional) Maximum runtime for the job (HH:MM:SS, only the value from the first row will be used).
- `job.name`: (Optional) The name of the job, only the value from the first row will be used.
- `logs`: (Optional) Directory for log files, only the value from the first row will be used
- `modules`: (Optional) A vector or list of modules to be loaded before executing the command.}

\item{cores}{Number of CPU cores to allocate. If not specified, the first value of cmd$cores column will be used. Default= 8.}

\item{mem}{Memory allocation in GB. If not specified, the first value of cmd$mem column will be used. Default= 32.}

\item{time}{Maximum runtime for the job (HH:MM:SS). If not specified, the first value of cmd$time column will be used. Default= '08:00:00'.}

\item{job.name}{Name of the job. If not specified, the first value of cmd$job.name column will be used. Default= "myJob".}

\item{logs}{Directory for log files. If not specified, the first value of cmd$logs column will be used. Default = 'db/logs'.}

\item{modules}{A character vector of modules to be loaded before executing the command. If not specified, the modules listed in cmd$modules will be used.
By default, the most used modules for genomics are loaded.}

\item{overwrite}{If set to TRUE, overwrites existing output files. Default= FALSE.}

\item{execute}{If set to FALSE, the command is returned and is not submitted to the cluster. Default= TRUE.}

\item{create.output.dirs}{Should missing output directories be created before executing the command? Default= TRUE.}
}
\description{
Submits shell commands to a server using resource management (e.g., LSF). Handles job submission,
resource allocation, and optional directory creation for output files.
}
\examples{
# Submit a command to the server
cmd <- data.table(
  file.type = "bam",
  path = "output.bam",
  cmd = "echo 'Hello world'",
  job.name= "test"
)
vl_submit(cmd)

# Return the constructed command without submitting
vl_submit(cmd, execute = FALSE)

}
